Format a tabulated summary for LaTeX

Format a tabulated summary for LaTeX

# S3 method for demoTable
format(demoTable, formula)

Arguments

demoTable
output from tabSummarize

formula
the same formula as used to create the tabSummarize output

Value

Table with summarized data formatted for inclusion in LaTeX

Note

This function is primarily used for demographics tables

See also

Examples

options(width = 150)
pkpdData = example.pkpdData()
ok = duplicated(pkpdData$id) == FALSE
myFormula =  dose ~ race + wt + bmi + sex
my.formatted.demoTable = format.demoTable(
 tabSummarize(formula = myFormula, data = pkpdData[ok, ], nSignif = 3)
  , formula = myFormula
)
print(my.formatted.demoTable
  , sanitize.text.function = identity
  , booktabs = TRUE
)

#>       ~                        0                    100                  250                  500                  All                 
#>  [1,] "\\hfill "               "(N=32)"             "(N=32)"             "(N=32)"             "(N=32)"             "(N=128)"           
#>  [2,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#>  [3,] "\\textbf{Race}"         ""                   ""                   ""                   ""                   ""                  
#>  [4,] "\\hfill Asian"          "5 (15.6\\%)"        "2 (6.2\\%)"         "7 (21.9\\%)"        "3 (9.4\\%)"         "17 (13.3\\%)"      
#>  [5,] "\\hfill Black"          "8 (25.0\\%)"        "5 (15.6\\%)"        "3 (9.4\\%)"         "5 (15.6\\%)"        "21 (16.4\\%)"      
#>  [6,] "\\hfill Caucasian"      "18 (56.2\\%)"       "25 (78.1\\%)"       "17 (53.1\\%)"       "21 (65.6\\%)"       "81 (63.3\\%)"      
#>  [7,] "\\hfill Other"          "1 (3.1\\%)"         "0 (0.0\\%)"         "5 (15.6\\%)"        "3 (9.4\\%)"         "9 (7.0\\%)"        
#>  [8,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#>  [9,] "\\textbf{Weight (kg)}"  ""                   ""                   ""                   ""                   ""                  
#> [10,] "\\hfill Mean (SD)"      "75.3 (9)"           "76.2 (13)"          "78 (10)"            "75.7 (10)"          "76.3 (10)"         
#> [11,] "\\hfill Median (range)" "76 (56 - 92)"       "74.5 (46 - 105)"    "79.5 (54 - 98)"     "78 (43 - 104)"      "77 (43 - 105)"     
#> [12,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#> [13,] "\\textbf{BMI (kg/m2)}"  ""                   ""                   ""                   ""                   ""                  
#> [14,] "\\hfill Mean (SD)"      "23.7 (4.8)"         "23.2 (5)"           "24.7 (3.7)"         "23.2 (4.1)"         "23.7 (4.4)"        
#> [15,] "\\hfill Median (range)" "23.2 (16.7 - 32.6)" "22.5 (15.2 - 37.1)" "25.6 (16.9 - 31.6)" "22.8 (11.1 - 36.4)" "23.2 (11.1 - 37.1)"
#> [16,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#> [17,] "\\textbf{Gender}"       ""                   ""                   ""                   ""                   ""                  
#> [18,] "\\hfill F"              "18 (56.2\\%)"       "14 (43.8\\%)"       "14 (43.8\\%)"       "18 (56.2\\%)"       "64 (50.0\\%)"      
#> [19,] "\\hfill M"              "14 (43.8\\%)"       "18 (56.2\\%)"       "18 (56.2\\%)"       "14 (43.8\\%)"       "64 (50.0\\%)"