Forced length significant digits

Forced length significant digits

formatted.signif(x, digits = 3, latex = FALSE, align.dot = FALSE)

Arguments

x
a vector of values to be processed into significant digits format

digits
single value specifying the number of significant digits

Value

A formatted character vector containing x in with significant digits format

See also

Examples

formatted.signif(c(5,4.99, 4.99999,5.0000001,5001),4)

#> [1] "5.000" "4.990" "5.000" "5.000" "5001" 

numvec = c(5,4.99, 4.99999,5.0000001,5001, 0.00005,101,39.9)
print(
   xtable(
      data.frame(original=numvec
                 ,latex.TRUE= formatted.signif(numvec,3, TRUE,FALSE)
                 ,align.TRUE= formatted.signif(numvec,3, FALSE,TRUE)
                 ,all.TRUE = formatted.signif(numvec,3, TRUE,TRUE)
      )
   )
   , booktabs=TRUE
   , sanitize.text.function = identity
)

#> % latex table generated in R 3.5.2 by xtable 1.8-3 package
#> % Fri Mar  1 05:51:31 2019
#> \begin{table}[ht]
#> \centering
#> \begin{tabular}{rrlll}
#>   \toprule
#>  & original & latex.TRUE & align.TRUE & all.TRUE \\ 
#>   \midrule
#> 1 & 5.00 & 5.00 & $\phantom{00}$5.00 & $\phantom{00}$5.00 \\ 
#>   2 & 4.99 & 4.99 & $\phantom{00}$4.99 & $\phantom{00}$4.99 \\ 
#>   3 & 5.00 & 5.00 & $\phantom{00}$5.00 & $\phantom{00}$5.00 \\ 
#>   4 & 5.00 & 5.00 & $\phantom{00}$5.00 & $\phantom{00}$5.00 \\ 
#>   5 & 5001.00 & 5.00$\cdot$10$^{3}$ & $\phantom{00}$5.00e+03 & $\phantom{00}$5.00$\cdot$10$^{3}$ \\ 
#>   6 & 0.00 & 5.00$\cdot$10$^{-5}$ & $\phantom{00}$5.00e-05 & $\phantom{00}$5.00$\cdot$10$^{-5}$ \\ 
#>   7 & 101.00 & 101 & $\phantom{}$101 & $\phantom{}$101 \\ 
#>   8 & 39.90 & 39.9 & $\phantom{0}$39.9 & $\phantom{0}$39.9 \\ 
#>    \bottomrule
#> \end{tabular}
#> \end{table}

# now take this to a LaTeX compiler!