Get the covariance matrix

Get the covariance matrix from a completed NONMEM run. If the the run has been zipped, the file will be extracted, then parsed and thereafter the file (not the zipped file) will be discarded.

get.covmat(
  run,
  path = paste(getwd(), "NONMEM", sep = "/"),
  ext = "cov",
  zip.extension = "7z"
)

Arguments

run
run rootname (e.g. run1)

path
directory where run output resides

ext
extension of the covariance

zip.extension
extension of the zip program. Defaults to "7z"

Value

list with covariance matrices for each estimation method

See also

Examples

get.covmat("example2", path = getOption("qpExampleDir"))

#> $`Iterative Two Stage (No Prior)`
#>              THETA1       THETA2       THETA3       THETA4       THETA5       THETA6       THETA7       THETA8       THETA9       THETA10     
#> THETA1        1.44751e-03  3.11467e-06 -4.12336e-04 -4.27881e-07  6.41809e-04 -1.87469e-05 -1.67622e-04  6.13238e-06  1.75125e-05  1.38204e-05
#> THETA2        3.11467e-06  8.31862e-04  1.28374e-06 -2.32055e-04  5.13644e-05  2.12792e-04 -1.53424e-05 -5.25471e-05  5.99954e-05  3.68344e-05
#> THETA3       -4.12336e-04  1.28374e-06  1.23070e-04 -1.23623e-07 -1.67203e-04  6.22605e-06  4.39757e-05 -1.62807e-06 -8.34201e-08 -6.25994e-08
#> THETA4       -4.27881e-07 -2.32055e-04 -1.23623e-07  7.09330e-05 -4.19765e-06 -5.20525e-05  1.59042e-06  1.47629e-05 -9.16374e-06 -4.77929e-06
#> THETA5        6.41809e-04  5.13644e-05 -1.67203e-04 -4.19765e-06  2.34150e-03  2.23585e-05 -6.36124e-04  4.10720e-06  1.15527e-04  1.07481e-04
#> THETA6       -1.87469e-05  2.12792e-04  6.22605e-06 -5.20525e-05  2.23585e-05  1.68452e-03 -9.83157e-06 -4.70937e-04  5.07191e-05  4.82282e-05
#> THETA7       -1.67622e-04 -1.53424e-05  4.39757e-05  1.59042e-06 -6.36124e-04 -9.83157e-06  1.80149e-04 -9.99553e-08 -3.25219e-05 -2.96824e-05
#> THETA8        6.13238e-06 -5.25471e-05 -1.62807e-06  1.47629e-05  4.10720e-06 -4.70937e-04 -9.99553e-08  1.40581e-04 -7.05168e-06 -5.42683e-06
#> THETA9        1.75125e-05  5.99954e-05 -8.34201e-08 -9.16374e-06  1.15527e-04  5.07191e-05 -3.25219e-05 -7.05168e-06  1.10734e-04  6.59534e-05
#> THETA10       1.38204e-05  3.68344e-05 -6.25994e-08 -4.77929e-06  1.07481e-04  4.82282e-05 -2.96824e-05 -5.42683e-06  6.59534e-05  7.98678e-05
#> THETA11      -7.58674e-06 -3.65042e-06  2.68095e-06  1.28805e-06 -8.73652e-06  7.41929e-07  2.49182e-06 -7.99813e-08  1.72676e-06  1.46226e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)    4.13648e-06 -1.54473e-07 -1.32788e-06 -1.72742e-07  1.27511e-06  2.50222e-07 -4.57666e-07 -4.16670e-08 -3.96766e-07  1.23321e-07
#> OMEGA(2,1)    1.54093e-06  1.62940e-07 -6.67053e-07 -5.18700e-08  1.89282e-06  7.42010e-07 -3.75314e-07  1.40560e-08 -1.60800e-07  3.35342e-07
#> OMEGA(2,2)    3.81679e-06  6.74609e-07 -1.07170e-06 -9.10468e-08  5.39505e-06 -7.12323e-06 -7.30935e-07  2.13144e-06 -3.58369e-08 -4.58352e-08
#> OMEGA(3,1)    4.68652e-07  1.83195e-06 -4.21533e-07 -5.80825e-07  1.99883e-06  2.41335e-08 -5.61102e-07  5.50810e-08  7.41226e-07  6.56749e-07
#> OMEGA(3,2)    3.09690e-07  5.58775e-07 -2.35273e-07 -2.85903e-07 -4.52134e-07  1.26022e-06  1.85457e-07 -6.52215e-08 -6.98743e-07 -3.54292e-07
#> OMEGA(3,3)    4.37605e-06  1.65903e-06 -1.33091e-06 -4.45386e-07  1.00909e-05 -5.25857e-06 -2.94259e-06  1.17501e-06 -1.27220e-06 -1.10556e-06
#> OMEGA(4,1)    1.93930e-06 -4.59419e-07 -7.43474e-07  2.95393e-07  6.07709e-06 -1.88627e-07 -1.51649e-06  1.37785e-07  7.56300e-07  9.21583e-07
#> OMEGA(4,2)    2.55401e-06 -4.01886e-07 -6.55296e-07  7.10748e-08  5.01598e-06 -2.17929e-06 -1.25374e-06  9.55437e-07 -4.54703e-08  8.11141e-07
#> OMEGA(4,3)    3.02307e-06  2.74198e-06 -9.62964e-07 -7.27357e-07  7.76784e-06  2.17832e-06 -2.42624e-06 -5.36088e-07 -2.93413e-07  4.27685e-07
#> OMEGA(4,4)   -1.56992e-07  3.53035e-06 -2.40840e-08 -7.30628e-07  8.55255e-06  5.78382e-06 -2.62928e-06 -1.04280e-06  1.17644e-06  3.19055e-06
#>              THETA11      SIGMA(1,1)   OMEGA(1,1)   OMEGA(2,1)   OMEGA(2,2)   OMEGA(3,1)   OMEGA(3,2)   OMEGA(3,3)   OMEGA(4,1)   OMEGA(4,2)  
#> THETA1       -7.58674e-06            0  4.13648e-06  1.54093e-06  3.81679e-06  4.68652e-07  3.09690e-07  4.37605e-06  1.93930e-06  2.55401e-06
#> THETA2       -3.65042e-06            0 -1.54473e-07  1.62940e-07  6.74609e-07  1.83195e-06  5.58775e-07  1.65903e-06 -4.59419e-07 -4.01886e-07
#> THETA3        2.68095e-06            0 -1.32788e-06 -6.67053e-07 -1.07170e-06 -4.21533e-07 -2.35273e-07 -1.33091e-06 -7.43474e-07 -6.55296e-07
#> THETA4        1.28805e-06            0 -1.72742e-07 -5.18700e-08 -9.10468e-08 -5.80825e-07 -2.85903e-07 -4.45386e-07  2.95393e-07  7.10748e-08
#> THETA5       -8.73652e-06            0  1.27511e-06  1.89282e-06  5.39505e-06  1.99883e-06 -4.52134e-07  1.00909e-05  6.07709e-06  5.01598e-06
#> THETA6        7.41929e-07            0  2.50222e-07  7.42010e-07 -7.12323e-06  2.41335e-08  1.26022e-06 -5.25857e-06 -1.88627e-07 -2.17929e-06
#> THETA7        2.49182e-06            0 -4.57666e-07 -3.75314e-07 -7.30935e-07 -5.61102e-07  1.85457e-07 -2.94259e-06 -1.51649e-06 -1.25374e-06
#> THETA8       -7.99813e-08            0 -4.16670e-08  1.40560e-08  2.13144e-06  5.50810e-08 -6.52215e-08  1.17501e-06  1.37785e-07  9.55437e-07
#> THETA9        1.72676e-06            0 -3.96766e-07 -1.60800e-07 -3.58369e-08  7.41226e-07 -6.98743e-07 -1.27220e-06  7.56300e-07 -4.54703e-08
#> THETA10       1.46226e-06            0  1.23321e-07  3.35342e-07 -4.58352e-08  6.56749e-07 -3.54292e-07 -1.10556e-06  9.21583e-07  8.11141e-07
#> THETA11       8.90711e-06            0 -4.26489e-07 -5.62027e-07 -1.07487e-06 -5.16314e-07 -6.87116e-07 -3.16978e-06 -3.41297e-07 -3.64467e-07
#> SIGMA(1,1)    0.00000e+00            0  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -4.26489e-07            0  1.01349e-06  3.68128e-07  9.09517e-08  5.33240e-07  2.98979e-07  4.44190e-07  4.32181e-07  1.96997e-07
#> OMEGA(2,1)   -5.62027e-07            0  3.68128e-07  7.48452e-07  4.66091e-07  3.32471e-07  4.26822e-07  4.59863e-07  2.95866e-07  3.57108e-07
#> OMEGA(2,2)   -1.07487e-06            0  9.09517e-08  4.66091e-07  2.24035e-06  6.41438e-08  5.29799e-07  2.92092e-07  1.49214e-07  5.21322e-07
#> OMEGA(3,1)   -5.16314e-07            0  5.33240e-07  3.32471e-07  6.41438e-08  1.66946e-06  7.84275e-07  1.28553e-06  1.03286e-06  4.64176e-07
#> OMEGA(3,2)   -6.87116e-07            0  2.98979e-07  4.26822e-07  5.29799e-07  7.84275e-07  2.06848e-06  1.53066e-06  5.04977e-07  1.12555e-06
#> OMEGA(3,3)   -3.16978e-06            0  4.44190e-07  4.59863e-07  2.92092e-07  1.28553e-06  1.53066e-06  7.99314e-06  6.62489e-07  7.05170e-07
#> OMEGA(4,1)   -3.41297e-07            0  4.32181e-07  2.95866e-07  1.49214e-07  1.03286e-06  5.04977e-07  6.62489e-07  1.24943e-06  5.26018e-07
#> OMEGA(4,2)   -3.64467e-07            0  1.96997e-07  3.57108e-07  5.21322e-07  4.64176e-07  1.12555e-06  7.05170e-07  5.26018e-07  1.33618e-06
#> OMEGA(4,3)   -2.29212e-06            0  4.17746e-07  3.94410e-07  2.27696e-07  9.03355e-07  1.20571e-06  4.54079e-06  8.46590e-07  7.90842e-07
#> OMEGA(4,4)   -1.83304e-06            0  4.12753e-07  3.77326e-07  9.81674e-08  7.35442e-07  8.20353e-07  2.40959e-06  9.64396e-07  9.46160e-07
#>              OMEGA(4,3)   OMEGA(4,4)  
#> THETA1        3.02307e-06 -1.56992e-07
#> THETA2        2.74198e-06  3.53035e-06
#> THETA3       -9.62964e-07 -2.40840e-08
#> THETA4       -7.27357e-07 -7.30628e-07
#> THETA5        7.76784e-06  8.55255e-06
#> THETA6        2.17832e-06  5.78382e-06
#> THETA7       -2.42624e-06 -2.62928e-06
#> THETA8       -5.36088e-07 -1.04280e-06
#> THETA9       -2.93413e-07  1.17644e-06
#> THETA10       4.27685e-07  3.19055e-06
#> THETA11      -2.29212e-06 -1.83304e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00
#> OMEGA(1,1)    4.17746e-07  4.12753e-07
#> OMEGA(2,1)    3.94410e-07  3.77326e-07
#> OMEGA(2,2)    2.27696e-07  9.81674e-08
#> OMEGA(3,1)    9.03355e-07  7.35442e-07
#> OMEGA(3,2)    1.20571e-06  8.20353e-07
#> OMEGA(3,3)    4.54079e-06  2.40959e-06
#> OMEGA(4,1)    8.46590e-07  9.64396e-07
#> OMEGA(4,2)    7.90842e-07  9.46160e-07
#> OMEGA(4,3)    3.93925e-06  3.04672e-06
#> OMEGA(4,4)    3.04672e-06  3.99993e-06
#> 
#> $`Importance Sampling (No Prior)`
#>              THETA1       THETA2       THETA3       THETA4       THETA5       THETA6       THETA7       THETA8       THETA9       THETA10     
#> THETA1        1.06846e-03  1.47960e-05 -3.01735e-04 -9.15615e-07  2.86116e-04  2.75862e-05 -7.98033e-05 -4.57762e-06  4.29417e-05  2.62814e-05
#> THETA2        1.47960e-05  8.18787e-04 -2.71053e-06 -2.28161e-04  2.42613e-05  2.83666e-04 -7.65207e-06 -7.63512e-05  4.31239e-05  2.97573e-05
#> THETA3       -3.01735e-04 -2.71053e-06  9.08904e-05  7.07287e-08 -7.73786e-05 -5.54496e-06  2.24978e-05  9.39329e-07 -9.13147e-06 -4.89935e-06
#> THETA4       -9.15615e-07 -2.28161e-04  7.07287e-08  6.90860e-05 -2.46187e-06 -7.46733e-05  7.87778e-07  2.20591e-05 -4.27059e-06 -3.26417e-06
#> THETA5        2.86116e-04  2.42613e-05 -7.73786e-05 -2.46187e-06  1.53349e-03  4.00240e-05 -4.29636e-04 -5.86681e-06  5.66247e-05  5.00986e-05
#> THETA6        2.75862e-05  2.83666e-04 -5.54496e-06 -7.46733e-05  4.00240e-05  1.26600e-03 -1.21037e-05 -3.52919e-04  6.44417e-05  5.40617e-05
#> THETA7       -7.98033e-05 -7.65207e-06  2.24978e-05  7.87778e-07 -4.29636e-04 -1.21037e-05  1.28165e-04  1.78675e-06 -1.79605e-05 -1.47254e-05
#> THETA8       -4.57762e-06 -7.63512e-05  9.39329e-07  2.20591e-05 -5.86681e-06 -3.52919e-04  1.78675e-06  1.06971e-04 -1.04998e-05 -8.08454e-06
#> THETA9        4.29417e-05  4.31239e-05 -9.13147e-06 -4.27059e-06  5.66247e-05  6.44417e-05 -1.79605e-05 -1.04998e-05  1.08104e-04  6.16742e-05
#> THETA10       2.62814e-05  2.97573e-05 -4.89935e-06 -3.26417e-06  5.00986e-05  5.40617e-05 -1.47254e-05 -8.08454e-06  6.16742e-05  7.26410e-05
#> THETA11       7.85988e-07  2.20027e-07 -1.76723e-07  1.00190e-07  2.09739e-06 -2.11751e-06 -5.93722e-07  8.93779e-07  2.18187e-06  1.64548e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)    4.69155e-07 -1.62544e-07 -8.40233e-08  6.16110e-08  7.08494e-07  8.63241e-07 -1.19694e-07 -1.89554e-07  2.29135e-07  2.08287e-07
#> OMEGA(2,1)    3.30392e-07 -1.60111e-07 -5.04489e-08  8.01413e-08  1.66105e-07  1.30526e-07  8.34457e-08  2.49383e-08  2.16364e-07  1.89493e-07
#> OMEGA(2,2)   -5.59839e-07 -2.50834e-07  1.71633e-07  6.62637e-08 -1.68869e-06 -1.15813e-06  6.11475e-07  3.21756e-07 -5.06850e-07 -4.03926e-07
#> OMEGA(3,1)    7.11854e-07  3.11291e-09 -7.39177e-08 -1.33439e-08  1.74917e-06  1.83917e-06 -3.87527e-07 -4.49514e-07  5.77428e-08  2.16085e-07
#> OMEGA(3,2)    5.99380e-07  1.12500e-07 -1.24134e-07 -3.25812e-08  1.98288e-06 -7.08503e-08 -3.29773e-07  3.89090e-08  4.34860e-07  2.24558e-07
#> OMEGA(3,3)    6.90435e-07  5.37322e-07 -9.11805e-08 -2.57637e-07  2.63725e-06  4.65792e-06 -6.88982e-07 -1.44452e-06 -1.87412e-06 -1.16262e-06
#> OMEGA(4,1)    1.19815e-06 -1.41126e-06 -3.02861e-07  4.75714e-07  1.23414e-06  1.42943e-06 -3.00442e-07 -3.41111e-07  3.28532e-07  1.76795e-07
#> OMEGA(4,2)    2.83319e-07 -1.64995e-07 -5.04415e-08  6.49908e-08  1.74778e-06 -2.19745e-06 -4.66710e-07  7.03710e-07  4.42669e-07  2.86316e-07
#> OMEGA(4,3)    5.27376e-07  1.19284e-06 -7.84530e-08 -3.97165e-07  1.70204e-06  5.02879e-06 -4.69174e-07 -1.51226e-06 -9.45931e-07 -7.83999e-07
#> OMEGA(4,4)    2.34959e-07  1.40796e-06 -1.75743e-08 -4.24504e-07  8.08915e-07  5.04705e-06 -2.29239e-07 -1.48073e-06 -5.25750e-07 -3.80794e-07
#>              THETA11      SIGMA(1,1)   OMEGA(1,1)   OMEGA(2,1)   OMEGA(2,2)   OMEGA(3,1)   OMEGA(3,2)   OMEGA(3,3)   OMEGA(4,1)   OMEGA(4,2)  
#> THETA1        7.85988e-07            0  4.69155e-07  3.30392e-07 -5.59839e-07  7.11854e-07  5.99380e-07  6.90435e-07  1.19815e-06  2.83319e-07
#> THETA2        2.20027e-07            0 -1.62544e-07 -1.60111e-07 -2.50834e-07  3.11291e-09  1.12500e-07  5.37322e-07 -1.41126e-06 -1.64995e-07
#> THETA3       -1.76723e-07            0 -8.40233e-08 -5.04489e-08  1.71633e-07 -7.39177e-08 -1.24134e-07 -9.11805e-08 -3.02861e-07 -5.04415e-08
#> THETA4        1.00190e-07            0  6.16110e-08  8.01413e-08  6.62637e-08 -1.33439e-08 -3.25812e-08 -2.57637e-07  4.75714e-07  6.49908e-08
#> THETA5        2.09739e-06            0  7.08494e-07  1.66105e-07 -1.68869e-06  1.74917e-06  1.98288e-06  2.63725e-06  1.23414e-06  1.74778e-06
#> THETA6       -2.11751e-06            0  8.63241e-07  1.30526e-07 -1.15813e-06  1.83917e-06 -7.08503e-08  4.65792e-06  1.42943e-06 -2.19745e-06
#> THETA7       -5.93722e-07            0 -1.19694e-07  8.34457e-08  6.11475e-07 -3.87527e-07 -3.29773e-07 -6.88982e-07 -3.00442e-07 -4.66710e-07
#> THETA8        8.93779e-07            0 -1.89554e-07  2.49383e-08  3.21756e-07 -4.49514e-07  3.89090e-08 -1.44452e-06 -3.41111e-07  7.03710e-07
#> THETA9        2.18187e-06            0  2.29135e-07  2.16364e-07 -5.06850e-07  5.77428e-08  4.34860e-07 -1.87412e-06  3.28532e-07  4.42669e-07
#> THETA10       1.64548e-06            0  2.08287e-07  1.89493e-07 -4.03926e-07  2.16085e-07  2.24558e-07 -1.16262e-06  1.76795e-07  2.86316e-07
#> THETA11       8.21184e-06            0 -4.97656e-07 -4.23220e-07 -1.05694e-06 -1.01193e-06 -5.80233e-07 -4.22538e-06 -7.91710e-07 -3.24708e-07
#> SIGMA(1,1)    0.00000e+00            0  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -4.97656e-07            0  9.61379e-07  2.76354e-07  6.06523e-08  6.17651e-07  1.94535e-07  6.94163e-07  4.14391e-07  9.50709e-08
#> OMEGA(2,1)   -4.23220e-07            0  2.76354e-07  6.79099e-07  3.16480e-07  2.31364e-07  3.27460e-07  3.70329e-07  2.17943e-07  2.05611e-07
#> OMEGA(2,2)   -1.05694e-06            0  6.06523e-08  3.16480e-07  1.75268e-06  2.09021e-08  2.36357e-07  1.94262e-07  4.18997e-08  3.03450e-07
#> OMEGA(3,1)   -1.01193e-06            0  6.17651e-07  2.31364e-07  2.09021e-08  1.86459e-06  7.28011e-07  2.60114e-06  1.05359e-06  2.82627e-07
#> OMEGA(3,2)   -5.80233e-07            0  1.94535e-07  3.27460e-07  2.36357e-07  7.28011e-07  1.97828e-06  1.68649e-06  3.93455e-07  9.33019e-07
#> OMEGA(3,3)   -4.22538e-06            0  6.94163e-07  3.70329e-07  1.94262e-07  2.60114e-06  1.68649e-06  1.06181e-05  1.49895e-06  5.67038e-07
#> OMEGA(4,1)   -7.91710e-07            0  4.14391e-07  2.17943e-07  4.18997e-08  1.05359e-06  3.93455e-07  1.49895e-06  1.11692e-06  3.20210e-07
#> OMEGA(4,2)   -3.24708e-07            0  9.50709e-08  2.05611e-07  3.03450e-07  2.82627e-07  9.33019e-07  5.67038e-07  3.20210e-07  1.18041e-06
#> OMEGA(4,3)   -2.98551e-06            0  4.70871e-07  2.63571e-07  9.83338e-08  1.67061e-06  1.19905e-06  6.53632e-06  1.22574e-06  5.75383e-07
#> OMEGA(4,4)   -2.21613e-06            0  3.13252e-07  1.82152e-07  6.97531e-08  1.01127e-06  6.96745e-07  3.78753e-06  9.49447e-07  6.18241e-07
#>              OMEGA(4,3)   OMEGA(4,4)  
#> THETA1        5.27376e-07  2.34959e-07
#> THETA2        1.19284e-06  1.40796e-06
#> THETA3       -7.84530e-08 -1.75743e-08
#> THETA4       -3.97165e-07 -4.24504e-07
#> THETA5        1.70204e-06  8.08915e-07
#> THETA6        5.02879e-06  5.04705e-06
#> THETA7       -4.69174e-07 -2.29239e-07
#> THETA8       -1.51226e-06 -1.48073e-06
#> THETA9       -9.45931e-07 -5.25750e-07
#> THETA10      -7.83999e-07 -3.80794e-07
#> THETA11      -2.98551e-06 -2.21613e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00
#> OMEGA(1,1)    4.70871e-07  3.13252e-07
#> OMEGA(2,1)    2.63571e-07  1.82152e-07
#> OMEGA(2,2)    9.83338e-08  6.97531e-08
#> OMEGA(3,1)    1.67061e-06  1.01127e-06
#> OMEGA(3,2)    1.19905e-06  6.96745e-07
#> OMEGA(3,3)    6.53632e-06  3.78753e-06
#> OMEGA(4,1)    1.22574e-06  9.49447e-07
#> OMEGA(4,2)    5.75383e-07  6.18241e-07
#> OMEGA(4,3)    5.27618e-06  3.76726e-06
#> OMEGA(4,4)    3.76726e-06  4.02294e-06
#> 
#> $`Stochastic Approximation Expectation-Maximization (No Prior)`
#>              THETA1       THETA2       THETA3       THETA4       THETA5       THETA6       THETA7       THETA8       THETA9       THETA10     
#> THETA1        1.42743e-03 -1.24999e-05 -4.07166e-04  2.52553e-06  6.37776e-04 -1.81995e-05 -1.64563e-04  4.52207e-06  9.39146e-07  4.51263e-06
#> THETA2       -1.24999e-05  8.64282e-04  5.05893e-06 -2.41601e-04  2.51316e-05  2.87214e-04 -9.25911e-06 -7.42537e-05  5.54853e-05  3.40128e-05
#> THETA3       -4.07166e-04  5.05893e-06  1.21545e-04 -9.79205e-07 -1.66756e-04  4.84259e-06  4.29769e-05 -1.04548e-06  2.96440e-06  1.07677e-06
#> THETA4        2.52553e-06 -2.41601e-04 -9.79205e-07  7.34932e-05 -1.21665e-06 -7.34739e-05  9.74263e-07  2.09424e-05 -1.12747e-05 -6.21909e-06
#> THETA5        6.37776e-04  2.51316e-05 -1.66756e-04 -1.21665e-06  2.33118e-03 -5.32620e-06 -6.29867e-04  7.10906e-06  7.64863e-05  8.11970e-05
#> THETA6       -1.81995e-05  2.87214e-04  4.84259e-06 -7.34739e-05 -5.32620e-06  1.79300e-03 -1.90614e-06 -5.00419e-04  3.08537e-05  3.66683e-05
#> THETA7       -1.64563e-04 -9.25911e-06  4.29769e-05  9.74263e-07 -6.29867e-04 -1.90614e-06  1.76672e-04 -1.15713e-06 -2.34815e-05 -2.41780e-05
#> THETA8        4.52207e-06 -7.42537e-05 -1.04548e-06  2.09424e-05  7.10906e-06 -5.00419e-04 -1.15713e-06  1.48858e-04 -5.35027e-06 -4.46464e-06
#> THETA9        9.39146e-07  5.54853e-05  2.96440e-06 -1.12747e-05  7.64863e-05  3.08537e-05 -2.34815e-05 -5.35027e-06  8.94379e-05  4.89397e-05
#> THETA10       4.51263e-06  3.40128e-05  1.07677e-06 -6.21909e-06  8.11970e-05  3.66683e-05 -2.41780e-05 -4.46464e-06  4.89397e-05  6.76058e-05
#> THETA11      -5.69400e-06  5.24227e-07  2.04907e-06  3.45010e-07 -3.79193e-06 -1.28349e-06  1.09617e-06  8.42116e-07  3.83419e-06  3.56705e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)    4.33425e-06 -4.67742e-07 -1.29913e-06 -8.49832e-08  8.26069e-07  6.70848e-07 -3.28299e-07 -2.03474e-07 -2.81530e-07  1.34757e-07
#> OMEGA(2,1)    1.40594e-06 -4.41340e-07 -5.99040e-07  6.97954e-08  5.60742e-07  1.13739e-06 -8.33745e-08 -2.04361e-07 -3.89060e-07  1.13086e-07
#> OMEGA(2,2)    3.05027e-06 -7.51359e-07 -8.89591e-07  1.78838e-07  5.28325e-06 -6.59167e-06 -9.77549e-07  1.69559e-06 -1.02574e-06 -1.07521e-06
#> OMEGA(3,1)   -3.59578e-07  1.75056e-06 -3.32032e-08 -5.55865e-07 -1.43754e-07  1.72723e-06  5.40344e-08 -4.60135e-07  5.02981e-07  4.41716e-07
#> OMEGA(3,2)   -6.02804e-07 -4.35580e-07  9.07527e-08 -1.55955e-08 -4.05025e-06  3.32389e-06  1.09969e-06 -6.66404e-07 -1.15116e-07 -2.46564e-07
#> OMEGA(3,3)    2.68621e-06 -4.45985e-06 -6.71931e-07  1.16022e-06  7.96513e-06 -3.92732e-06 -2.04844e-06  1.11569e-06 -1.76429e-06 -1.70673e-06
#> OMEGA(4,1)    1.68182e-06 -5.63247e-07 -5.80455e-07  3.12784e-07  4.64239e-06  8.81274e-07 -1.12062e-06 -2.09235e-07  5.38116e-07  6.52716e-07
#> OMEGA(4,2)    2.33102e-06 -1.14001e-06 -5.76634e-07  2.68450e-07  2.88501e-06 -5.14221e-07 -7.46404e-07  3.29924e-07  2.06143e-07  7.25625e-07
#> OMEGA(4,3)    1.92257e-06 -9.22863e-07 -5.46191e-07  2.20935e-07  4.44429e-06  2.93067e-06 -1.26994e-06 -5.99708e-07 -6.35518e-07 -2.26549e-07
#> OMEGA(4,4)   -2.03014e-07  1.54863e-06  4.57311e-09 -2.39462e-07  5.11847e-06  6.00639e-06 -1.52516e-06 -1.10399e-06  7.09438e-07  2.43957e-06
#>              THETA11      SIGMA(1,1)   OMEGA(1,1)   OMEGA(2,1)   OMEGA(2,2)   OMEGA(3,1)   OMEGA(3,2)   OMEGA(3,3)   OMEGA(4,1)   OMEGA(4,2)  
#> THETA1       -5.69400e-06            0  4.33425e-06  1.40594e-06  3.05027e-06 -3.59578e-07 -6.02804e-07  2.68621e-06  1.68182e-06  2.33102e-06
#> THETA2        5.24227e-07            0 -4.67742e-07 -4.41340e-07 -7.51359e-07  1.75056e-06 -4.35580e-07 -4.45985e-06 -5.63247e-07 -1.14001e-06
#> THETA3        2.04907e-06            0 -1.29913e-06 -5.99040e-07 -8.89591e-07 -3.32032e-08  9.07527e-08 -6.71931e-07 -5.80455e-07 -5.76634e-07
#> THETA4        3.45010e-07            0 -8.49832e-08  6.97954e-08  1.78838e-07 -5.55865e-07 -1.55955e-08  1.16022e-06  3.12784e-07  2.68450e-07
#> THETA5       -3.79193e-06            0  8.26069e-07  5.60742e-07  5.28325e-06 -1.43754e-07 -4.05025e-06  7.96513e-06  4.64239e-06  2.88501e-06
#> THETA6       -1.28349e-06            0  6.70848e-07  1.13739e-06 -6.59167e-06  1.72723e-06  3.32389e-06 -3.92732e-06  8.81274e-07 -5.14221e-07
#> THETA7        1.09617e-06            0 -3.28299e-07 -8.33745e-08 -9.77549e-07  5.40344e-08  1.09969e-06 -2.04844e-06 -1.12062e-06 -7.46404e-07
#> THETA8        8.42116e-07            0 -2.03474e-07 -2.04361e-07  1.69559e-06 -4.60135e-07 -6.66404e-07  1.11569e-06 -2.09235e-07  3.29924e-07
#> THETA9        3.83419e-06            0 -2.81530e-07 -3.89060e-07 -1.02574e-06  5.02981e-07 -1.15116e-07 -1.76429e-06  5.38116e-07  2.06143e-07
#> THETA10       3.56705e-06            0  1.34757e-07  1.13086e-07 -1.07521e-06  4.41716e-07 -2.46564e-07 -1.70673e-06  6.52716e-07  7.25625e-07
#> THETA11       8.67371e-06            0 -2.29778e-07 -3.47250e-07 -9.51777e-07 -1.89258e-07 -1.78263e-07 -1.94570e-06 -8.52873e-08 -1.04863e-08
#> SIGMA(1,1)    0.00000e+00            0  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -2.29778e-07            0  8.24356e-07  2.35427e-07  2.43382e-09  1.61147e-07  1.01649e-07  1.45584e-08  1.44519e-07  5.91879e-08
#> OMEGA(2,1)   -3.47250e-07            0  2.35427e-07  6.02739e-07  2.45046e-07  5.88544e-08  1.25672e-07  7.77186e-08  8.72130e-08  1.23116e-07
#> OMEGA(2,2)   -9.51777e-07            0  2.43382e-09  2.45046e-07  1.87649e-06 -1.01666e-07 -3.95173e-08  9.72785e-08 -8.16781e-09  5.90099e-08
#> OMEGA(3,1)   -1.89258e-07            0  1.61147e-07  5.88544e-08 -1.01666e-07  1.05805e-06  3.38050e-07  1.06701e-07  5.61729e-07  1.70369e-07
#> OMEGA(3,2)   -1.78263e-07            0  1.01649e-07  1.25672e-07 -3.95173e-08  3.38050e-07  1.44938e-06  1.49900e-07  1.65606e-07  5.97216e-07
#> OMEGA(3,3)   -1.94570e-06            0  1.45584e-08  7.77186e-08  9.72785e-08  1.06701e-07  1.49900e-07  4.20690e-06 -1.01740e-07 -7.72694e-08
#> OMEGA(4,1)   -8.52873e-08            0  1.44519e-07  8.72130e-08 -8.16781e-09  5.61729e-07  1.65606e-07 -1.01740e-07  8.51136e-07  2.57517e-07
#> OMEGA(4,2)   -1.04863e-08            0  5.91879e-08  1.23116e-07  5.90099e-08  1.70369e-07  5.97216e-07 -7.72694e-08  2.57517e-07  8.85291e-07
#> OMEGA(4,3)   -1.26232e-06            0  8.38639e-08  1.01155e-07 -6.70874e-09  8.60773e-08  2.54312e-07  1.80789e-06  1.79687e-07  4.34446e-08
#> OMEGA(4,4)   -9.98459e-07            0  1.38313e-07  1.38639e-07 -1.55823e-07  1.53171e-07  7.78006e-08  4.90694e-07  3.70305e-07  2.29703e-07
#>              OMEGA(4,3)   OMEGA(4,4)  
#> THETA1        1.92257e-06 -2.03014e-07
#> THETA2       -9.22863e-07  1.54863e-06
#> THETA3       -5.46191e-07  4.57311e-09
#> THETA4        2.20935e-07 -2.39462e-07
#> THETA5        4.44429e-06  5.11847e-06
#> THETA6        2.93067e-06  6.00639e-06
#> THETA7       -1.26994e-06 -1.52516e-06
#> THETA8       -5.99708e-07 -1.10399e-06
#> THETA9       -6.35518e-07  7.09438e-07
#> THETA10      -2.26549e-07  2.43957e-06
#> THETA11      -1.26232e-06 -9.98459e-07
#> SIGMA(1,1)    0.00000e+00  0.00000e+00
#> OMEGA(1,1)    8.38639e-08  1.38313e-07
#> OMEGA(2,1)    1.01155e-07  1.38639e-07
#> OMEGA(2,2)   -6.70874e-09 -1.55823e-07
#> OMEGA(3,1)    8.60773e-08  1.53171e-07
#> OMEGA(3,2)    2.54312e-07  7.78006e-08
#> OMEGA(3,3)    1.80789e-06  4.90694e-07
#> OMEGA(4,1)    1.79687e-07  3.70305e-07
#> OMEGA(4,2)    4.34446e-08  2.29703e-07
#> OMEGA(4,3)    1.73734e-06  1.21444e-06
#> OMEGA(4,4)    1.21444e-06  2.26373e-06
#> 
#> $`Objective Function Evaluation by Importance Sampling (No Prior)`
#>              THETA1       THETA2       THETA3       THETA4       THETA5       THETA6       THETA7       THETA8       THETA9       THETA10     
#> THETA1        1.07206e-03  1.20706e-05 -3.02363e-04 -6.72316e-07  2.99920e-04  2.38278e-05 -8.27217e-05 -4.07897e-06  3.69707e-05  2.32186e-05
#> THETA2        1.20706e-05  8.38601e-04 -2.38161e-06 -2.33973e-04  2.00641e-05  3.25819e-04 -6.96693e-06 -8.81028e-05  4.08881e-05  2.88801e-05
#> THETA3       -3.02363e-04 -2.38161e-06  9.08726e-05  8.80171e-08 -8.08094e-05 -4.96062e-06  2.30400e-05  8.79885e-07 -8.09462e-06 -4.63377e-06
#> THETA4       -6.72316e-07 -2.33973e-04  8.80171e-08  7.07805e-05 -2.06130e-06 -8.62724e-05  7.89928e-07  2.53538e-05 -4.54825e-06 -3.52133e-06
#> THETA5        2.99920e-04  2.00641e-05 -8.08094e-05 -2.06130e-06  1.53354e-03  3.39081e-05 -4.28640e-04 -4.76698e-06  4.71462e-05  4.57473e-05
#> THETA6        2.38278e-05  3.25819e-04 -4.96062e-06 -8.62724e-05  3.39081e-05  1.31605e-03 -1.16349e-05 -3.65994e-04  6.47050e-05  5.59936e-05
#> THETA7       -8.27217e-05 -6.96693e-06  2.30400e-05  7.89928e-07 -4.28640e-04 -1.16349e-05  1.27320e-04  1.85502e-06 -1.61927e-05 -1.44085e-05
#> THETA8       -4.07897e-06 -8.81028e-05  8.79885e-07  2.53538e-05 -4.76698e-06 -3.65994e-04  1.85502e-06  1.10465e-04 -1.18558e-05 -9.20239e-06
#> THETA9        3.69707e-05  4.08881e-05 -8.09462e-06 -4.54825e-06  4.71462e-05  6.47050e-05 -1.61927e-05 -1.18558e-05  1.03554e-04  5.80115e-05
#> THETA10       2.32186e-05  2.88801e-05 -4.63377e-06 -3.52133e-06  4.57473e-05  5.59936e-05 -1.44085e-05 -9.20239e-06  5.80115e-05  7.00696e-05
#> THETA11       9.51798e-07 -3.70278e-07 -3.53542e-07  1.38644e-07  4.62924e-06 -3.12880e-06 -1.60510e-06  1.04493e-06  5.91876e-07  6.95708e-07
#> SIGMA(1,1)    0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)    7.86722e-07 -1.06177e-07 -1.34324e-07  1.15256e-07  3.81644e-07  5.16649e-07 -1.36849e-08 -5.02208e-08  4.98366e-07  3.79703e-07
#> OMEGA(2,1)    6.31057e-07 -3.12922e-07 -9.32907e-08  1.66427e-07  2.59410e-07 -7.96044e-08  1.01188e-07  1.51539e-07  3.61657e-07  2.79333e-07
#> OMEGA(2,2)   -3.59152e-07 -2.04262e-07  1.58871e-07  7.42810e-08 -9.73646e-07 -1.17516e-06  6.22419e-07  4.13000e-07 -4.07283e-07 -3.53125e-07
#> OMEGA(3,1)    1.08618e-06  6.70286e-07 -6.26326e-08 -1.89375e-08  1.13973e-06  1.62863e-06 -1.53697e-07 -2.57740e-07  1.31912e-06  1.01864e-06
#> OMEGA(3,2)    8.96785e-07  6.71302e-07 -1.29352e-07 -5.55487e-08  2.05459e-06  1.01502e-06 -1.41818e-07  3.61066e-09  1.65150e-06  1.02531e-06
#> OMEGA(3,3)    1.43574e-06  1.48506e-06 -1.10067e-07 -1.34503e-07  4.73242e-07  5.25142e-06  1.22060e-07 -1.20633e-06  2.78938e-06  1.76918e-06
#> OMEGA(4,1)    1.20324e-06 -1.05004e-06 -2.12911e-07  4.82703e-07  5.28106e-07  1.09808e-06 -5.19159e-08 -1.73244e-07  9.98994e-07  6.99190e-07
#> OMEGA(4,2)    2.90477e-07  1.63894e-08 -1.41004e-08  6.90465e-08  1.35248e-06 -1.51591e-06 -1.67856e-07  6.45788e-07  8.03868e-07  6.34660e-07
#> OMEGA(4,3)    7.46914e-07  2.00459e-06 -8.37705e-09 -3.64257e-07 -4.37965e-07  5.64638e-06  2.82728e-07 -1.40842e-06  2.18830e-06  1.50045e-06
#> OMEGA(4,4)    1.39006e-07  2.02962e-06  8.96078e-08 -4.40110e-07 -1.15549e-06  5.58038e-06  4.06641e-07 -1.45983e-06  1.36073e-06  1.29450e-06
#>              THETA11      SIGMA(1,1)   OMEGA(1,1)   OMEGA(2,1)   OMEGA(2,2)   OMEGA(3,1)   OMEGA(3,2)   OMEGA(3,3)   OMEGA(4,1)   OMEGA(4,2)  
#> THETA1        9.51798e-07            0  7.86722e-07  6.31057e-07 -3.59152e-07  1.08618e-06  8.96785e-07  1.43574e-06  1.20324e-06  2.90477e-07
#> THETA2       -3.70278e-07            0 -1.06177e-07 -3.12922e-07 -2.04262e-07  6.70286e-07  6.71302e-07  1.48506e-06 -1.05004e-06  1.63894e-08
#> THETA3       -3.53542e-07            0 -1.34324e-07 -9.32907e-08  1.58871e-07 -6.26326e-08 -1.29352e-07 -1.10067e-07 -2.12911e-07 -1.41004e-08
#> THETA4        1.38644e-07            0  1.15256e-07  1.66427e-07  7.42810e-08 -1.89375e-08 -5.55487e-08 -1.34503e-07  4.82703e-07  6.90465e-08
#> THETA5        4.62924e-06            0  3.81644e-07  2.59410e-07 -9.73646e-07  1.13973e-06  2.05459e-06  4.73242e-07  5.28106e-07  1.35248e-06
#> THETA6       -3.12880e-06            0  5.16649e-07 -7.96044e-08 -1.17516e-06  1.62863e-06  1.01502e-06  5.25142e-06  1.09808e-06 -1.51591e-06
#> THETA7       -1.60510e-06            0 -1.36849e-08  1.01188e-07  6.22419e-07 -1.53697e-07 -1.41818e-07  1.22060e-07 -5.19159e-08 -1.67856e-07
#> THETA8        1.04493e-06            0 -5.02208e-08  1.51539e-07  4.13000e-07 -2.57740e-07  3.61066e-09 -1.20633e-06 -1.73244e-07  6.45788e-07
#> THETA9        5.91876e-07            0  4.98366e-07  3.61657e-07 -4.07283e-07  1.31912e-06  1.65150e-06  2.78938e-06  9.98994e-07  8.03868e-07
#> THETA10       6.95708e-07            0  3.79703e-07  2.79333e-07 -3.53125e-07  1.01864e-06  1.02531e-06  1.76918e-06  6.99190e-07  6.34660e-07
#> THETA11       9.60544e-06            0 -4.86859e-07 -4.80744e-07 -1.37838e-06 -1.05268e-06 -9.69136e-07 -4.65426e-06 -7.75758e-07 -4.98689e-07
#> SIGMA(1,1)    0.00000e+00            0  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -4.86859e-07            0  8.24216e-07  2.01748e-07  6.38042e-08  3.37932e-07  1.28999e-07  3.56873e-07  2.05741e-07  5.22561e-08
#> OMEGA(2,1)   -4.80744e-07            0  2.01748e-07  5.98506e-07  2.67558e-07  5.63490e-08  1.95576e-07  2.53291e-07  8.75263e-08  1.02584e-07
#> OMEGA(2,2)   -1.37838e-06            0  6.38042e-08  2.67558e-07  1.68557e-06  3.55632e-08  3.66425e-08  4.26573e-07  3.83352e-08  1.46211e-07
#> OMEGA(3,1)   -1.05268e-06            0  3.37932e-07  5.63490e-08  3.55632e-08  1.38757e-06  5.85017e-07  1.54751e-06  6.99230e-07  1.98526e-07
#> OMEGA(3,2)   -9.69136e-07            0  1.28999e-07  1.95576e-07  3.66425e-08  5.85017e-07  1.87343e-06  1.13000e-06  2.74882e-07  7.91634e-07
#> OMEGA(3,3)   -4.65426e-06            0  3.56873e-07  2.53291e-07  4.26573e-07  1.54751e-06  1.13000e-06  8.23122e-06  8.29025e-07  3.15425e-07
#> OMEGA(4,1)   -7.75758e-07            0  2.05741e-07  8.75263e-08  3.83352e-08  6.99230e-07  2.74882e-07  8.29025e-07  8.21913e-07  2.15260e-07
#> OMEGA(4,2)   -4.98689e-07            0  5.22561e-08  1.02584e-07  1.46211e-07  1.98526e-07  7.91634e-07  3.15425e-07  2.15260e-07  9.91311e-07
#> OMEGA(4,3)   -3.22911e-06            0  2.25265e-07  1.66930e-07  2.17232e-07  9.49320e-07  8.56822e-07  4.81407e-06  6.50206e-07  2.53157e-07
#> OMEGA(4,4)   -2.30797e-06            0  1.37691e-07  1.01581e-07  1.16092e-07  5.20050e-07  4.50484e-07  2.60570e-06  4.65426e-07  2.71767e-07
#>              OMEGA(4,3)   OMEGA(4,4)  
#> THETA1        7.46914e-07  1.39006e-07
#> THETA2        2.00459e-06  2.02962e-06
#> THETA3       -8.37705e-09  8.96078e-08
#> THETA4       -3.64257e-07 -4.40110e-07
#> THETA5       -4.37965e-07 -1.15549e-06
#> THETA6        5.64638e-06  5.58038e-06
#> THETA7        2.82728e-07  4.06641e-07
#> THETA8       -1.40842e-06 -1.45983e-06
#> THETA9        2.18830e-06  1.36073e-06
#> THETA10       1.50045e-06  1.29450e-06
#> THETA11      -3.22911e-06 -2.30797e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00
#> OMEGA(1,1)    2.25265e-07  1.37691e-07
#> OMEGA(2,1)    1.66930e-07  1.01581e-07
#> OMEGA(2,2)    2.17232e-07  1.16092e-07
#> OMEGA(3,1)    9.49320e-07  5.20050e-07
#> OMEGA(3,2)    8.56822e-07  4.50484e-07
#> OMEGA(3,3)    4.81407e-06  2.60570e-06
#> OMEGA(4,1)    6.50206e-07  4.65426e-07
#> OMEGA(4,2)    2.53157e-07  2.71767e-07
#> OMEGA(4,3)    3.90368e-06  2.66385e-06
#> OMEGA(4,4)    2.66385e-06  2.99811e-06
#> 
#> $`MCMC Bayesian Analysis`
#>              THETA1       THETA2       THETA3       THETA4       THETA5       THETA6       THETA7       THETA8       THETA9       THETA10     
#> THETA1        1.06389e-03  3.38427e-05 -3.05001e-04 -1.12291e-05  2.65206e-04  5.06165e-05 -7.33286e-05 -2.44878e-05  4.17149e-06  9.59716e-06
#> THETA2        3.38427e-05  7.26788e-04 -1.47485e-05 -2.06757e-04 -1.69986e-05  2.26446e-04  1.66842e-06 -6.57796e-05 -1.53691e-06 -4.64313e-06
#> THETA3       -3.05001e-04 -1.47485e-05  9.30219e-05  4.81073e-06 -7.22684e-05 -1.41174e-05  2.04283e-05  6.61450e-06  8.61092e-07 -1.10629e-06
#> THETA4       -1.12291e-05 -2.06757e-04  4.81073e-06  6.37575e-05  2.96806e-06 -6.10056e-05 -2.63782e-07  1.86067e-05  3.22749e-06  2.94780e-06
#> THETA5        2.65206e-04 -1.69986e-05 -7.22684e-05  2.96806e-06  1.61905e-03 -1.12048e-04 -4.51098e-04  2.70576e-05  7.81994e-06  2.53842e-05
#> THETA6        5.06165e-05  2.26446e-04 -1.41174e-05 -6.10056e-05 -1.12048e-04  1.22245e-03  2.61601e-05 -3.50566e-04 -1.70017e-05 -1.38367e-05
#> THETA7       -7.33286e-05  1.66842e-06  2.04283e-05 -2.63782e-07 -4.51098e-04  2.61601e-05  1.34803e-04 -6.00686e-06 -4.49691e-06 -8.18818e-06
#> THETA8       -2.44878e-05 -6.57796e-05  6.61450e-06  1.86067e-05  2.70576e-05 -3.50566e-04 -6.00686e-06  1.07245e-04  5.44142e-06  5.62503e-06
#> THETA9        4.17149e-06 -1.53691e-06  8.61092e-07  3.22749e-06  7.81994e-06 -1.70017e-05 -4.49691e-06  5.44142e-06  6.12583e-05  2.66950e-05
#> THETA10       9.59716e-06 -4.64313e-06 -1.10629e-06  2.94780e-06  2.53842e-05 -1.38367e-05 -8.18818e-06  5.62503e-06  2.66950e-05  4.69026e-05
#> THETA11      -3.26975e-06 -3.84893e-06  1.02589e-06  1.24013e-06  1.25594e-05 -3.17350e-06 -3.12322e-06  6.94138e-07  1.28087e-06  1.02714e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -7.85737e-07  6.73381e-07  3.13142e-07 -1.31338e-07  2.26083e-06  2.06591e-06 -8.86028e-07 -4.78376e-07  8.50850e-07  7.67238e-07
#> OMEGA(2,1)    7.71554e-07 -8.18664e-07 -1.61120e-07  3.01677e-07  3.33181e-06 -1.26151e-06 -8.43329e-07  4.97053e-07  3.97237e-07  2.75795e-07
#> OMEGA(2,2)   -3.11781e-07 -3.85619e-06  1.16162e-08  1.00529e-06 -3.79791e-06 -8.56869e-06  1.32714e-06  2.87970e-06 -3.87906e-07 -5.60893e-07
#> OMEGA(3,1)    2.22839e-07  2.11136e-07  7.05596e-08  1.14755e-07  3.09660e-06  2.95771e-06 -1.39994e-06 -9.57670e-07  8.41076e-07  9.45211e-07
#> OMEGA(3,2)    1.30198e-06 -2.85411e-06 -4.28658e-07  5.83112e-07  4.08387e-07 -8.15602e-06 -1.22981e-07  2.28455e-06 -2.23906e-07 -2.40352e-07
#> OMEGA(3,3)   -5.62274e-07 -1.93869e-06  2.04324e-07  7.58048e-07 -2.24533e-06  6.87887e-07  7.40479e-08 -2.44772e-09 -9.35682e-07  3.16742e-08
#> OMEGA(4,1)    7.58793e-07  7.78120e-07 -1.40413e-07 -1.74396e-07  2.11190e-06  6.28926e-06 -9.47663e-07 -1.94451e-06  8.66985e-08  2.73956e-07
#> OMEGA(4,2)    7.95487e-07 -1.27431e-06 -3.03343e-07  6.66717e-08  2.42713e-07 -9.16191e-07  1.30287e-07  3.31424e-07 -1.29378e-06 -8.15097e-07
#> OMEGA(4,3)    3.41817e-07  3.24213e-07 -6.32124e-08 -1.40435e-07 -3.85174e-07  7.26978e-06 -1.90620e-07 -2.19908e-06 -1.73619e-06 -6.26631e-07
#> OMEGA(4,4)    2.28786e-06  3.16314e-06 -6.55464e-07 -1.01432e-06 -4.31024e-07  1.36132e-05  6.02411e-08 -4.09083e-06 -2.13715e-06 -8.29300e-07
#>              THETA11      SIGMA(1,1)   OMEGA(1,1)   OMEGA(2,1)   OMEGA(2,2)   OMEGA(3,1)   OMEGA(3,2)   OMEGA(3,3)   OMEGA(4,1)   OMEGA(4,2)  
#> THETA1       -3.26975e-06            0 -7.85737e-07  7.71554e-07 -3.11781e-07  2.22839e-07  1.30198e-06 -5.62274e-07  7.58793e-07  7.95487e-07
#> THETA2       -3.84893e-06            0  6.73381e-07 -8.18664e-07 -3.85619e-06  2.11136e-07 -2.85411e-06 -1.93869e-06  7.78120e-07 -1.27431e-06
#> THETA3        1.02589e-06            0  3.13142e-07 -1.61120e-07  1.16162e-08  7.05596e-08 -4.28658e-07  2.04324e-07 -1.40413e-07 -3.03343e-07
#> THETA4        1.24013e-06            0 -1.31338e-07  3.01677e-07  1.00529e-06  1.14755e-07  5.83112e-07  7.58048e-07 -1.74396e-07  6.66717e-08
#> THETA5        1.25594e-05            0  2.26083e-06  3.33181e-06 -3.79791e-06  3.09660e-06  4.08387e-07 -2.24533e-06  2.11190e-06  2.42713e-07
#> THETA6       -3.17350e-06            0  2.06591e-06 -1.26151e-06 -8.56869e-06  2.95771e-06 -8.15602e-06  6.87887e-07  6.28926e-06 -9.16191e-07
#> THETA7       -3.12322e-06            0 -8.86028e-07 -8.43329e-07  1.32714e-06 -1.39994e-06 -1.22981e-07  7.40479e-08 -9.47663e-07  1.30287e-07
#> THETA8        6.94138e-07            0 -4.78376e-07  4.97053e-07  2.87970e-06 -9.57670e-07  2.28455e-06 -2.44772e-09 -1.94451e-06  3.31424e-07
#> THETA9        1.28087e-06            0  8.50850e-07  3.97237e-07 -3.87906e-07  8.41076e-07 -2.23906e-07 -9.35682e-07  8.66985e-08 -1.29378e-06
#> THETA10       1.02714e-06            0  7.67238e-07  2.75795e-07 -5.60893e-07  9.45211e-07 -2.40352e-07  3.16742e-08  2.73956e-07 -8.15097e-07
#> THETA11       6.14626e-06            0 -2.15336e-07 -5.76491e-08 -3.89255e-07 -1.04511e-08  6.32822e-07 -6.55410e-07  1.95811e-08  3.93878e-07
#> SIGMA(1,1)    0.00000e+00            0  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -2.15336e-07            0  9.08349e-07  1.98681e-07 -1.13335e-07  4.14482e-07 -2.65691e-07  4.27549e-07  1.93364e-07 -2.48755e-07
#> OMEGA(2,1)   -5.76491e-08            0  1.98681e-07  5.97650e-07  1.80550e-07  2.55716e-08  2.35202e-08  3.14547e-08  7.36423e-08  3.44455e-08
#> OMEGA(2,2)   -3.89255e-07            0 -1.13335e-07  1.80550e-07  1.46942e-06 -4.39198e-07  1.75827e-07 -3.26451e-07 -2.19248e-07  4.32946e-07
#> OMEGA(3,1)   -1.04511e-08            0  4.14482e-07  2.55716e-08 -4.39198e-07  1.23642e-06 -3.90690e-07  7.77103e-07  5.53390e-07 -6.10225e-07
#> OMEGA(3,2)    6.32822e-07            0 -2.65691e-07  2.35202e-08  1.75827e-07 -3.90690e-07  1.26245e-06 -7.13089e-07 -2.02292e-07  8.86053e-07
#> OMEGA(3,3)   -6.55410e-07            0  4.27549e-07  3.14547e-08 -3.26451e-07  7.77103e-07 -7.13089e-07  2.59021e-06  3.13664e-07 -6.97188e-07
#> OMEGA(4,1)    1.95811e-08            0  1.93364e-07  7.36423e-08 -2.19248e-07  5.53390e-07 -2.02292e-07  3.13664e-07  7.34975e-07 -1.80590e-07
#> OMEGA(4,2)    3.93878e-07            0 -2.48755e-07  3.44455e-08  4.32946e-07 -6.10225e-07  8.86053e-07 -6.97188e-07 -1.80590e-07  1.47283e-06
#> OMEGA(4,3)   -5.50195e-07            0  5.64932e-08 -1.09799e-07 -1.16116e-07  7.48324e-08 -7.41818e-08  6.92580e-07  1.58276e-07  8.99276e-09
#> OMEGA(4,4)   -5.16391e-07            0 -2.11621e-07 -1.64470e-07  7.52267e-08 -3.99293e-07  3.00117e-07 -3.64921e-07  1.37627e-09  6.38448e-07
#>              OMEGA(4,3)   OMEGA(4,4)  
#> THETA1        3.41817e-07  2.28786e-06
#> THETA2        3.24213e-07  3.16314e-06
#> THETA3       -6.32124e-08 -6.55464e-07
#> THETA4       -1.40435e-07 -1.01432e-06
#> THETA5       -3.85174e-07 -4.31024e-07
#> THETA6        7.26978e-06  1.36132e-05
#> THETA7       -1.90620e-07  6.02411e-08
#> THETA8       -2.19908e-06 -4.09083e-06
#> THETA9       -1.73619e-06 -2.13715e-06
#> THETA10      -6.26631e-07 -8.29300e-07
#> THETA11      -5.50195e-07 -5.16391e-07
#> SIGMA(1,1)    0.00000e+00  0.00000e+00
#> OMEGA(1,1)    5.64932e-08 -2.11621e-07
#> OMEGA(2,1)   -1.09799e-07 -1.64470e-07
#> OMEGA(2,2)   -1.16116e-07  7.52267e-08
#> OMEGA(3,1)    7.48324e-08 -3.99293e-07
#> OMEGA(3,2)   -7.41818e-08  3.00117e-07
#> OMEGA(3,3)    6.92580e-07 -3.64921e-07
#> OMEGA(4,1)    1.58276e-07  1.37627e-09
#> OMEGA(4,2)    8.99276e-09  6.38448e-07
#> OMEGA(4,3)    9.74189e-07  7.91542e-07
#> OMEGA(4,4)    7.91542e-07  2.02555e-06
#> 
#> $`First Order Conditional Estimation with Interaction (No Prior)`
#>              THETA1       THETA2       THETA3       THETA4       THETA5       THETA6       THETA7       THETA8       THETA9       THETA10     
#> THETA1        1.06649e-03  1.55722e-05 -3.01019e-04 -9.87836e-07  2.90814e-04  2.82244e-05 -8.10159e-05 -4.55030e-06  4.38036e-05  2.72250e-05
#> THETA2        1.55722e-05  8.20629e-04 -2.84835e-06 -2.28295e-04  2.48686e-05  2.92956e-04 -7.84363e-06 -7.82961e-05  4.50918e-05  3.15896e-05
#> THETA3       -3.01019e-04 -2.84835e-06  9.06487e-05  1.01661e-07 -7.84963e-05 -5.48638e-06  2.28177e-05  9.14400e-07 -9.15400e-06 -4.97214e-06
#> THETA4       -9.87836e-07 -2.28295e-04  1.01661e-07  6.90568e-05 -2.43912e-06 -7.63316e-05  8.20337e-07  2.24196e-05 -4.54958e-06 -3.50281e-06
#> THETA5        2.90814e-04  2.48686e-05 -7.84963e-05 -2.43912e-06  1.53442e-03  3.98071e-05 -4.29911e-04 -5.45271e-06  5.74296e-05  5.17727e-05
#> THETA6        2.82244e-05  2.92956e-04 -5.48638e-06 -7.63316e-05  3.98071e-05  1.27935e-03 -1.24269e-05 -3.55135e-04  7.04798e-05  5.98171e-05
#> THETA7       -8.10159e-05 -7.84363e-06  2.28177e-05  8.20337e-07 -4.29911e-04 -1.24269e-05  1.28243e-04  1.85078e-06 -1.80195e-05 -1.49990e-05
#> THETA8       -4.55030e-06 -7.82961e-05  9.14400e-07  2.24196e-05 -5.45271e-06 -3.55135e-04  1.85078e-06  1.07260e-04 -1.17738e-05 -9.23764e-06
#> THETA9        4.38036e-05  4.50918e-05 -9.15400e-06 -4.54958e-06  5.74296e-05  7.04798e-05 -1.80195e-05 -1.17738e-05  1.09587e-04  6.30676e-05
#> THETA10       2.72250e-05  3.15896e-05 -4.97214e-06 -3.50281e-06  5.17727e-05  5.98171e-05 -1.49990e-05 -9.23764e-06  6.30676e-05  7.35362e-05
#> THETA11       1.36198e-06 -1.91665e-07 -3.21404e-07  8.36271e-08  3.18560e-06 -2.81737e-06 -1.03493e-06  7.04715e-07  5.03115e-07  8.38692e-08
#> SIGMA(1,1)    0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)    2.67499e-07 -1.04752e-07 -7.35903e-09  7.22186e-08  3.54220e-07  4.69081e-07  3.89225e-10 -3.05576e-08  4.27859e-07  3.99633e-07
#> OMEGA(2,1)    2.19855e-07 -2.02160e-07  3.19349e-09  1.24780e-07  1.44328e-07 -1.88050e-07  1.44112e-07  2.01962e-07  3.42233e-07  3.21297e-07
#> OMEGA(2,2)   -2.54098e-07 -1.11022e-07  1.49304e-07  1.16017e-07 -1.26757e-06 -1.20065e-06  7.10307e-07  5.48861e-07  1.29698e-07  1.68789e-07
#> OMEGA(3,1)    4.82606e-07  1.26123e-07  4.25565e-08  3.98484e-08  1.13094e-06  1.17016e-06 -1.63944e-07 -1.21219e-07  1.00682e-06  1.00150e-06
#> OMEGA(3,2)    6.63884e-07  8.91285e-08 -7.13691e-08  8.43542e-08  2.06538e-06 -5.94858e-07 -2.44085e-07  3.90135e-07  1.19925e-06  9.22755e-07
#> OMEGA(3,3)    7.68004e-07  1.68960e-06  5.28654e-08 -1.00742e-07  1.39937e-06  5.40109e-06 -8.53007e-08 -9.42243e-07  1.52974e-06  1.92956e-06
#> OMEGA(4,1)    9.60351e-07 -1.32544e-06 -1.93201e-07  5.05145e-07  6.33105e-07  9.55186e-07 -9.17110e-08 -1.18202e-07  8.92238e-07  6.86379e-07
#> OMEGA(4,2)    2.89725e-07 -2.92601e-07 -1.42161e-08  1.47668e-07  1.68168e-06 -2.49700e-06 -3.21265e-07  8.98272e-07  6.27586e-07  4.76220e-07
#> OMEGA(4,3)    4.58645e-07  2.08878e-06  5.87177e-08 -3.03522e-07  4.13053e-07  5.76010e-06  3.90765e-08 -1.22950e-06  1.68031e-06  1.52384e-06
#> OMEGA(4,4)    5.50412e-08  2.03007e-06  1.28568e-07 -3.67569e-07 -3.72218e-07  5.69728e-06  1.66746e-07 -1.34711e-06  1.31969e-06  1.19550e-06
#>              THETA11      SIGMA(1,1)   OMEGA(1,1)   OMEGA(2,1)   OMEGA(2,2)   OMEGA(3,1)   OMEGA(3,2)   OMEGA(3,3)   OMEGA(4,1)   OMEGA(4,2)  
#> THETA1        1.36198e-06            0  2.67499e-07  2.19855e-07 -2.54098e-07  4.82606e-07  6.63884e-07  7.68004e-07  9.60351e-07  2.89725e-07
#> THETA2       -1.91665e-07            0 -1.04752e-07 -2.02160e-07 -1.11022e-07  1.26123e-07  8.91285e-08  1.68960e-06 -1.32544e-06 -2.92601e-07
#> THETA3       -3.21404e-07            0 -7.35903e-09  3.19349e-09  1.49304e-07  4.25565e-08 -7.13691e-08  5.28654e-08 -1.93201e-07 -1.42161e-08
#> THETA4        8.36271e-08            0  7.22186e-08  1.24780e-07  1.16017e-07  3.98484e-08  8.43542e-08 -1.00742e-07  5.05145e-07  1.47668e-07
#> THETA5        3.18560e-06            0  3.54220e-07  1.44328e-07 -1.26757e-06  1.13094e-06  2.06538e-06  1.39937e-06  6.33105e-07  1.68168e-06
#> THETA6       -2.81737e-06            0  4.69081e-07 -1.88050e-07 -1.20065e-06  1.17016e-06 -5.94858e-07  5.40109e-06  9.55186e-07 -2.49700e-06
#> THETA7       -1.03493e-06            0  3.89225e-10  1.44112e-07  7.10307e-07 -1.63944e-07 -2.44085e-07 -8.53007e-08 -9.17110e-08 -3.21265e-07
#> THETA8        7.04715e-07            0 -3.05576e-08  2.01962e-07  5.48861e-07 -1.21219e-07  3.90135e-07 -9.42243e-07 -1.18202e-07  8.98272e-07
#> THETA9        5.03115e-07            0  4.27859e-07  3.42233e-07  1.29698e-07  1.00682e-06  1.19925e-06  1.52974e-06  8.92238e-07  6.27586e-07
#> THETA10       8.38692e-08            0  3.99633e-07  3.21297e-07  1.68789e-07  1.00150e-06  9.22755e-07  1.92956e-06  6.86379e-07  4.76220e-07
#> THETA11       7.99352e-06            0 -3.99795e-07 -4.31942e-07 -1.24239e-06 -7.30882e-07 -5.81091e-07 -3.49656e-06 -6.10366e-07 -3.91482e-07
#> SIGMA(1,1)    0.00000e+00            0  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00  0.00000e+00
#> OMEGA(1,1)   -3.99795e-07            0  9.37401e-07  2.73915e-07  8.54841e-08  5.41565e-07  1.76162e-07  4.55486e-07  3.63086e-07  9.38582e-08
#> OMEGA(2,1)   -4.31942e-07            0  2.73915e-07  6.84761e-07  3.68647e-07  2.20081e-07  3.55338e-07  3.35564e-07  2.11829e-07  2.28911e-07
#> OMEGA(2,2)   -1.24239e-06            0  8.54841e-08  3.68647e-07  1.88955e-06  7.88405e-08  3.82860e-07  3.85840e-07  7.88729e-08  4.14615e-07
#> OMEGA(3,1)   -7.30882e-07            0  5.41565e-07  2.20081e-07  7.88405e-08  1.63128e-06  6.76399e-07  1.96211e-06  8.98668e-07  2.86533e-07
#> OMEGA(3,2)   -5.81091e-07            0  1.76162e-07  3.55338e-07  3.82860e-07  6.76399e-07  2.06287e-06  1.62362e-06  3.59206e-07  9.97267e-07
#> OMEGA(3,3)   -3.49656e-06            0  4.55486e-07  3.35564e-07  3.85840e-07  1.96211e-06  1.62362e-06  9.27813e-06  1.07701e-06  6.27629e-07
#> OMEGA(4,1)   -6.10366e-07            0  3.63086e-07  2.11829e-07  7.88729e-08  8.98668e-07  3.59206e-07  1.07701e-06  1.00635e-06  3.19705e-07
#> OMEGA(4,2)   -3.91482e-07            0  9.38582e-08  2.28911e-07  4.14615e-07  2.86533e-07  9.97267e-07  6.27629e-07  3.19705e-07  1.20192e-06
#> OMEGA(4,3)   -2.49408e-06            0  3.04314e-07  2.41175e-07  2.25568e-07  1.22836e-06  1.16140e-06  5.56572e-06  9.32249e-07  6.15462e-07
#> OMEGA(4,4)   -1.90109e-06            0  2.01771e-07  1.66530e-07  1.46116e-07  7.19430e-07  6.62756e-07  3.13490e-06  7.57332e-07  6.40542e-07
#>              OMEGA(4,3)   OMEGA(4,4)  
#> THETA1        4.58645e-07  5.50412e-08
#> THETA2        2.08878e-06  2.03007e-06
#> THETA3        5.87177e-08  1.28568e-07
#> THETA4       -3.03522e-07 -3.67569e-07
#> THETA5        4.13053e-07 -3.72218e-07
#> THETA6        5.76010e-06  5.69728e-06
#> THETA7        3.90765e-08  1.66746e-07
#> THETA8       -1.22950e-06 -1.34711e-06
#> THETA9        1.68031e-06  1.31969e-06
#> THETA10       1.52384e-06  1.19550e-06
#> THETA11      -2.49408e-06 -1.90109e-06
#> SIGMA(1,1)    0.00000e+00  0.00000e+00
#> OMEGA(1,1)    3.04314e-07  2.01771e-07
#> OMEGA(2,1)    2.41175e-07  1.66530e-07
#> OMEGA(2,2)    2.25568e-07  1.46116e-07
#> OMEGA(3,1)    1.22836e-06  7.19430e-07
#> OMEGA(3,2)    1.16140e-06  6.62756e-07
#> OMEGA(3,3)    5.56572e-06  3.13490e-06
#> OMEGA(4,1)    9.32249e-07  7.57332e-07
#> OMEGA(4,2)    6.15462e-07  6.40542e-07
#> OMEGA(4,3)    4.58930e-06  3.31375e-06
#> OMEGA(4,4)    3.31375e-06  3.69975e-06
#>