VPC plot using ggplot2 (Xpose-alike version)

Plot basic vpc with 95% CI of 2.5th, 50th & 97.5th predicted percentile and observed data. This call is to be extended with for stratification, and axes definition using the layering custom to ggplot2 objects. See the examples below to learn more.

ggvpc_xpose(vpc, PI = c(0.025, 0.975),
  area.col.central = PI.ci.med.arcol, area.col.outer = gray(0.2),
  linecol.pred = "blue", linesize.pred = 1, linetype.obs = "dashed",
  linecol.obs = "black", linesize.obs = 0.5, alpha = 0.33,
  point.shape = 1, point.size = 1.25, point.col = "darkslategrey",
  yrange.stretch = c(0.9, 1.1), quiet = TRUE)

Arguments

vpc
output from nm.read.vpc

PI
prediction interval (c(0.025,0.975) for 95% CI)

area.col.central
color of prediction polygon for the central tendency

area.col.outer
color of prediction polygon for the outer percentiles

linecol.pred
color of predicted lines

linesize.pred
line width of of predicted lines

linetype.obs
line type of predicted lines

linecol.obs
line color of observed data

linesize.obs
line width of observed data

alpha
transparancy scalar (between 0 and 1)

point.shape
numeric value for dot shape

point.size
scalar of the observed data dot size

point.col
color of observed data dots

yrange.stretch
vector of c(min,max) which will proportionally rescale the lower and upper limits of the Y axis

Value

A ggplot object to be extended optionally

Note

Editing and stratification to be done by adding ggplot layer

See also

Examples

## example of vpc NOT involving stratification
 nm.read.vpc(path =  file.path(getOption("qpExampleDir"),"vpc_final_strt")) -> vpc.all

#> /home/travis/build/qPharmetra/qpToolkit/NONMEM/vpc_final_strt/vpc_results.csv was not found for reading

#> Warning: cannot open file '/home/travis/build/qPharmetra/qpToolkit/NONMEM/vpc_final_strt/NA': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

ggvpc_xpose(vpc.all) +
  labs(x="Time (h)", y="Concentration (ng/ml)")

#> Error in names(vpc$obs)[names(vpc$obs) == unique(vpc$obs$idv.var)] = "IVAR": object 'vpc.all' not found

ggvpc_xpose(vpc.all) +
  labs(x="Time (h)", y="Concentration (ng/ml)") +
  scale_y_log10() + scale_x_log10()

#> Error in ggvpc_xpose(vpc.all): object 'vpc.all' not found

## demonstration of changing colors & stratification
ggvpc_xpose(vpc.all
            , area.col.outer = qp.green
            , area.col.central = qp.blue
            , linecol.pred = steel
            , alpha = 1
            , linesize.pred=3
            , PI = c(0.05,0.95)) +
  labs(x="Time (h)", y="Concentration (ng/ml)") +
  scale_y_log10() + scale_x_log10() + facet_grid(~strata)

#> Error in ggvpc_xpose(vpc.all, area.col.outer = qp.green, area.col.central = qp.blue,     linecol.pred = steel, alpha = 1, linesize.pred = 3, PI = c(0.05,         0.95)): object 'vpc.all' not found

## PI smoothed
nm.read.vpc(path =  file.path(getOption("qpExampleDir"),"vpc_final_strt"),PI.ci.area.smooth=TRUE) -> vpc.all

#> /home/travis/build/qPharmetra/qpToolkit/NONMEM/vpc_final_strt/vpc_results.csv was not found for reading

#> Warning: cannot open file '/home/travis/build/qPharmetra/qpToolkit/NONMEM/vpc_final_strt/NA': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

ggvpc_xpose(vpc.all
            , PI = c(0.05,0.95)) +
  labs(x="Time (h)", y="Concentration (ng/ml)") +
  scale_y_log10() + scale_x_log10() + facet_grid(~strata)

#> Error in ggvpc_xpose(vpc.all, PI = c(0.05, 0.95)): object 'vpc.all' not found