General PNG grpah creator

A general wrapper function for a call to png(), which opens a plot device (i.e. enables you to "print" a graphic to a file). Note that the plot must be generated and then dev.off() must be called to close the plot device (i.e. close the output graphic file)

myPNG(
  filename = "C:/Test/test.png",
  w = 8,
  h = 6,
  units = "in",
  psize = 10,
  bg = "transparent",
  res = 600,
  ...
)

Arguments

filename
filename including path. To follow qP workflow standards should read the internal relative reference directory "../WorkAre/output/graphs"

w
width in units

h
height in units

units
units ('in', 'cm', or 'pt')

psize
pointsize in units

bg
background color

res
resolution of PNG graph. High values create top quality graph at the expense of large file size.

...
any additional parameters to be passed on the png()

Value

A graph written to file in .png format.