Extract information from nlme object

Given an nlme object, extract all available information for further processing. In addition parse the information into a ready-to-use parameter estimate table. Furthermore, when method is set to samples, it will create samples to be used for a VPC (see nlme.vpc).

# S3 method for extract
nlme(
  obj,
  method = "components",
  newdata = NULL,
  sample = FALSE,
  digits = 3,
  digits.CV = 0
)

Arguments

obj
an nlme object

method
can be either components (extracting function) or samples (for VPC)

newdata
a new data set with at the minimum the data items (independent, dependent variables, covariates, and grouping variables) required to build the nlme object.

sample
logical (default is FALSE) indicating whether to sample or not

digits
number of significant digits for parameter estimate table

digits.CV
number of significant digits for coefficient of variation in the parameter estimate table

Value

list with extracted nlme object information, model summary, model parameter table, or, when method = "samples", a matrix with predictions for data.frame getData(obj) or newdata.

See also

Examples

fm1 <- nlme::nlme(height ~ SSasymp(age, Asym, R0, lrc),
data = Loblolly,
fixed = Asym + R0 + lrc ~ 1,
random = Asym ~ 1,
start = c(Asym = 103, R0 = -8.5, lrc = -3.3))
fm1.out = nlme.extract(fm1)
fm1.out

#> $form
#> SSasymp(age, Asym, R0, lrc)
#> 
#> $response
#> height
#> 
#> $groups
#> Seed
#> 
#> $dataVars
#> [1] "age"
#> 
#> $fixef
#> $fixef$coef
#>       Asym         R0        lrc 
#> 101.449600  -8.627331  -3.233751 
#> 
#> $fixef$var
#>             Asym           R0          lrc
#> Asym  5.84351635  0.531301643 -0.073847624
#> R0    0.53130164  0.097482045 -0.008690417
#> lrc  -0.07384762 -0.008690417  0.001132500
#> 
#> 
#> $ranef
#> $ranef$Seed
#>     Asym 
#> 3.650642 
#> 
#> $ranef$var
#>               reStruct.Seed        lSigma
#> reStruct.Seed  0.0389578347 -0.0003654294
#> lSigma        -0.0003654294  0.0071578871
#> 
#> 
#> $varef
#> $varef$sigma
#> [1] 0.7188625
#> 
#> $varef$varCoef
#> NULL
#> 
#> $varef$varClass
#> [1] "NULL"
#> 
#> $varef$varFunc
#> NULL
#> 
#> 
#> $table
#>       Parameter         Est.se  CV            CI95
#> 1          Asym     101 (2.46)  2%    (96.6 - 106)
#> 2            R0  -8.63 (0.318)  4%    (-9.25 - -8)
#> 3           lrc -3.23 (0.0343)  1%  (-3.3 - -3.17)
#> 4 reStruct.Seed    3.65 (0.72) 20%   (2.48 - 5.37)
#> 5         sigma 0.719 (0.0595)  8% (0.609 - 0.849)
#>