Predict partial residuals from an nls object

Predict partial residuals from an nls object to be plotted with plot.fit

nls.predict(func, object, newdata = NULL, xrange = 150)

Arguments

func
formula of y ~ x | z detailing what to predict. Requires y to be the response variable of the nls obhect

object
and nls object

newdata
optional dataset to simulate across. Needs the elements required to create the prediction

xrange
number of equal-spaced x covariate values created from min(x) to max(x)

Value

a graph or a multi-level list with observed and predicted output (class c('nlsfit','fit','list'))

See also

Examples

DNase1 <- subset(DNase, Run == 1)

## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
summary(fm1DNase1)

#> 
#> Formula: density ~ SSlogis(log(conc), Asym, xmid, scal)
#> 
#> Parameters:
#>      Estimate Std. Error t value Pr(>|t|)    
#> Asym  2.34518    0.07815   30.01 2.17e-13 ***
#> xmid  1.48309    0.08135   18.23 1.22e-10 ***
#> scal  1.04146    0.03227   32.27 8.51e-14 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.01919 on 13 degrees of freedom
#> 
#> Number of iterations to convergence: 0 
#> Achieved convergence tolerance: 8.283e-06
#> 

fm1DNase1.predict = nls.predict(density ~ conc,object = fm1DNase1)

#> Error in eval(if ("data" %in% names(object)) object$data else mCall$data): object 'DNase1' not found

plot(fm1DNase1.predict)

#> Error in plot(fm1DNase1.predict): object 'fm1DNase1.predict' not found

fm1DNase1.predict = nls.predict(density ~ conc,object = fm1DNase1)

#> Error in eval(if ("data" %in% names(object)) object$data else mCall$data): object 'DNase1' not found

plot(fm1DNase1.predict, logX = TRUE)

#> Error in plot(fm1DNase1.predict, logX = TRUE): object 'fm1DNase1.predict' not found