NONMEM output extraction from XML output

Extracts all output embedded in the XML file of a NONMEM run

nm.extract.xml(
  run = "run1",
  path = getOption("nmDir"),
  filename = paste(path, run, paste(run, xml.extension, sep = ""), sep = "/"),
  digits = 3,
  digits.se = digits,
  digits.cv = digits,
  digits.est = digits,
  xml.extension = ".xml",
  zip.extension = ".7z",
  zip.call = getOption("unzip.call"),
  remove.obsolete = FALSE,
  na.value = "n.d.",
  control_stream = FALSE,
  get.xpose.tables = TRUE,
  quiet = TRUE
)

Arguments

run
run rootname (e.g. run1)

path
directory where run resides

filename
with full path of the XML file to be parsed. This is automatically created using qP workflow standards but can be specified as well

digits
number of significant digits

digits.se
number of significant digits in SE (when point estimates = T)

digits.cv
number of significant digits in CV% (when point.estimates = T)

digits.est
number of significant digits in parameter estimates (when point.estimates = T)

xml.extension
extension of the xml file. Defaults to ".xml"

zip.extension
extension of the zip program. Defaults to ".7z"

zip.call
full OS call to zip utility for unzipping. Defaults to "c:/progra~1/7-zip/7z e %s.7z" on windows

remove.obsolete
logical (F) defining if non-estimated parameters should be dropped for the parameter table.

na.value
what to do with an NA value in the parameter table. Defaults to "n.d."

control_stream
logical (F) defining if the control stream should be included as well.

get.xpose.tables
logical (T) defining if the xpose tables should be integrated in the result as a single data frame using the function get.xpose.tables

quiet
if TRUE (default) silences any message returned by the function and subfunctions.

Value

An extensive list with all elements extracted from the XML output

Note

Works for NM7.2 (up to 9 OMEGAs) and NM7.3 (bug free)

Examples

tmp = nm.extract.xml("example2", path = getOption("qpExampleDir"))

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> extraction of example2.xml complete.

names(tmp)

#>  [1] "fixed"                         "fixedse"                       "omega"                         "omegase"                      
#>  [5] "sigma"                         "sigmase"                       "covariance"                    "correlation"                  
#>  [9] "table"                         "eigenvalues"                   "etabar"                        "etabarse"                     
#> [13] "etabarpval"                    "etashrink"                     "epsshrink"                     "estimation.method"            
#> [17] "parallel_est"                  "monitor"                       "termination_status"            "termination_information"      
#> [21] "estimation_elapsed_time"       "covariance_elapsed_time"       "final_objective_function"      "final_objective_function_text"
#> [25] "final_objective_function_std" 

tmp$final_objective_function

#> $`Iterative Two Stage (No Prior)`
#> [1] -10771.86
#> 
#> $`Importance Sampling (No Prior)`
#> [1] -10783.92
#> 
#> $`Stochastic Approximation Expectation-Maximization (No Prior)`
#> [1] -19993.88
#> 
#> $`Objective Function Evaluation by Importance Sampling (No Prior)`
#> [1] -10777.68
#> 
#> $`MCMC Bayesian Analysis`
#> [1] -19185.25
#> 
#> $`First Order Conditional Estimation with Interaction (No Prior)`
#> [1] -10772.14
#> 

tmp$table

#> $`Iterative Two Stage (No Prior)`
#>    Parameter     Estimate           CV.perc          SE estimated
#> 1     THETA1  3.305220000 0.988049207011939   0.0326572 estimated
#> 2     THETA2  3.257970000 0.879277586963662   0.0286466 estimated
#> 3     THETA3 -0.611824000 -1.55615994142106  0.00952096 estimated
#> 4     THETA4 -0.208226000 -3.99087529895402  0.00831004 estimated
#> 5     THETA5  0.735323000   5.3271283503984   0.0391716 estimated
#> 6     THETA6  1.139920000   3.1377640536178    0.035768 estimated
#> 7     THETA7  0.336296000  3.36742036777125   0.0113245 estimated
#> 8     THETA8  0.191801000  5.39971115896163   0.0103567 estimated
#> 9     THETA9  0.694545000   1.5072313529001   0.0104684 estimated
#> 10   THETA10  2.301570000 0.372585669781932  0.00857532 estimated
#> 11   THETA11  0.100030000  2.82643207037889  0.00282728 estimated
#> 12  OMEGA1.1  0.010282100  9.41631573316735 0.000968195 estimated
#> 13  OMEGA2.1  0.000191938  431.130365013702 0.000827503 estimated
#> 14  OMEGA2.2  0.007948230  17.2945423069035  0.00137461 estimated
#> 15  OMEGA3.1  0.001281790  99.6434673386436  0.00127722 estimated
#> 16  OMEGA3.2 -0.000170245 -843.648858997327  0.00143627 estimated
#> 17  OMEGA3.3  0.010006800  30.4393012751329    0.003046 estimated
#> 18  OMEGA4.1 -0.000595539 -168.447406467083  0.00100317 estimated
#> 19  OMEGA4.2  0.000542183  202.204790633421  0.00109632 estimated
#> 20  OMEGA4.3  0.002000550  107.083552023194  0.00214226 estimated
#> 21  OMEGA4.4  0.009670120  19.8908596790939  0.00192347 estimated
#> 22  SIGMA1.1  1.000000000             .....       .....     fixed
#> 
#> $`Importance Sampling (No Prior)`
#>    Parameter     Estimate           CV.perc          SE estimated
#> 1     THETA1  3.305220000 0.988049207011939   0.0326572 estimated
#> 2     THETA2  3.257970000 0.879277586963662   0.0286466 estimated
#> 3     THETA3 -0.611824000 -1.55615994142106  0.00952096 estimated
#> 4     THETA4 -0.208226000 -3.99087529895402  0.00831004 estimated
#> 5     THETA5  0.735323000   5.3271283503984   0.0391716 estimated
#> 6     THETA6  1.139920000   3.1377640536178    0.035768 estimated
#> 7     THETA7  0.336296000  3.36742036777125   0.0113245 estimated
#> 8     THETA8  0.191801000  5.39971115896163   0.0103567 estimated
#> 9     THETA9  0.694545000   1.5072313529001   0.0104684 estimated
#> 10   THETA10  2.301570000 0.372585669781932  0.00857532 estimated
#> 11   THETA11  0.100030000  2.82643207037889  0.00282728 estimated
#> 12  OMEGA1.1  0.010282100  9.41631573316735 0.000968195 estimated
#> 13  OMEGA2.1  0.000191938  431.130365013702 0.000827503 estimated
#> 14  OMEGA2.2  0.007948230  17.2945423069035  0.00137461 estimated
#> 15  OMEGA3.1  0.001281790  99.6434673386436  0.00127722 estimated
#> 16  OMEGA3.2 -0.000170245 -843.648858997327  0.00143627 estimated
#> 17  OMEGA3.3  0.010006800  30.4393012751329    0.003046 estimated
#> 18  OMEGA4.1 -0.000595539 -168.447406467083  0.00100317 estimated
#> 19  OMEGA4.2  0.000542183  202.204790633421  0.00109632 estimated
#> 20  OMEGA4.3  0.002000550  107.083552023194  0.00214226 estimated
#> 21  OMEGA4.4  0.009670120  19.8908596790939  0.00192347 estimated
#> 22  SIGMA1.1  1.000000000             .....       .....     fixed
#> 
#> $`Stochastic Approximation Expectation-Maximization (No Prior)`
#>    Parameter     Estimate           CV.perc          SE estimated
#> 1     THETA1  3.305220000 0.988049207011939   0.0326572 estimated
#> 2     THETA2  3.257970000 0.879277586963662   0.0286466 estimated
#> 3     THETA3 -0.611824000 -1.55615994142106  0.00952096 estimated
#> 4     THETA4 -0.208226000 -3.99087529895402  0.00831004 estimated
#> 5     THETA5  0.735323000   5.3271283503984   0.0391716 estimated
#> 6     THETA6  1.139920000   3.1377640536178    0.035768 estimated
#> 7     THETA7  0.336296000  3.36742036777125   0.0113245 estimated
#> 8     THETA8  0.191801000  5.39971115896163   0.0103567 estimated
#> 9     THETA9  0.694545000   1.5072313529001   0.0104684 estimated
#> 10   THETA10  2.301570000 0.372585669781932  0.00857532 estimated
#> 11   THETA11  0.100030000  2.82643207037889  0.00282728 estimated
#> 12  OMEGA1.1  0.010282100  9.41631573316735 0.000968195 estimated
#> 13  OMEGA2.1  0.000191938  431.130365013702 0.000827503 estimated
#> 14  OMEGA2.2  0.007948230  17.2945423069035  0.00137461 estimated
#> 15  OMEGA3.1  0.001281790  99.6434673386436  0.00127722 estimated
#> 16  OMEGA3.2 -0.000170245 -843.648858997327  0.00143627 estimated
#> 17  OMEGA3.3  0.010006800  30.4393012751329    0.003046 estimated
#> 18  OMEGA4.1 -0.000595539 -168.447406467083  0.00100317 estimated
#> 19  OMEGA4.2  0.000542183  202.204790633421  0.00109632 estimated
#> 20  OMEGA4.3  0.002000550  107.083552023194  0.00214226 estimated
#> 21  OMEGA4.4  0.009670120  19.8908596790939  0.00192347 estimated
#> 22  SIGMA1.1  1.000000000             .....       .....     fixed
#> 
#> $`Objective Function Evaluation by Importance Sampling (No Prior)`
#>    Parameter     Estimate           CV.perc          SE estimated
#> 1     THETA1  3.305220000 0.988049207011939   0.0326572 estimated
#> 2     THETA2  3.257970000 0.879277586963662   0.0286466 estimated
#> 3     THETA3 -0.611824000 -1.55615994142106  0.00952096 estimated
#> 4     THETA4 -0.208226000 -3.99087529895402  0.00831004 estimated
#> 5     THETA5  0.735323000   5.3271283503984   0.0391716 estimated
#> 6     THETA6  1.139920000   3.1377640536178    0.035768 estimated
#> 7     THETA7  0.336296000  3.36742036777125   0.0113245 estimated
#> 8     THETA8  0.191801000  5.39971115896163   0.0103567 estimated
#> 9     THETA9  0.694545000   1.5072313529001   0.0104684 estimated
#> 10   THETA10  2.301570000 0.372585669781932  0.00857532 estimated
#> 11   THETA11  0.100030000  2.82643207037889  0.00282728 estimated
#> 12  OMEGA1.1  0.010282100  9.41631573316735 0.000968195 estimated
#> 13  OMEGA2.1  0.000191938  431.130365013702 0.000827503 estimated
#> 14  OMEGA2.2  0.007948230  17.2945423069035  0.00137461 estimated
#> 15  OMEGA3.1  0.001281790  99.6434673386436  0.00127722 estimated
#> 16  OMEGA3.2 -0.000170245 -843.648858997327  0.00143627 estimated
#> 17  OMEGA3.3  0.010006800  30.4393012751329    0.003046 estimated
#> 18  OMEGA4.1 -0.000595539 -168.447406467083  0.00100317 estimated
#> 19  OMEGA4.2  0.000542183  202.204790633421  0.00109632 estimated
#> 20  OMEGA4.3  0.002000550  107.083552023194  0.00214226 estimated
#> 21  OMEGA4.4  0.009670120  19.8908596790939  0.00192347 estimated
#> 22  SIGMA1.1  1.000000000             .....       .....     fixed
#> 
#> $`MCMC Bayesian Analysis`
#>    Parameter     Estimate           CV.perc          SE estimated
#> 1     THETA1  3.305220000 0.988049207011939   0.0326572 estimated
#> 2     THETA2  3.257970000 0.879277586963662   0.0286466 estimated
#> 3     THETA3 -0.611824000 -1.55615994142106  0.00952096 estimated
#> 4     THETA4 -0.208226000 -3.99087529895402  0.00831004 estimated
#> 5     THETA5  0.735323000   5.3271283503984   0.0391716 estimated
#> 6     THETA6  1.139920000   3.1377640536178    0.035768 estimated
#> 7     THETA7  0.336296000  3.36742036777125   0.0113245 estimated
#> 8     THETA8  0.191801000  5.39971115896163   0.0103567 estimated
#> 9     THETA9  0.694545000   1.5072313529001   0.0104684 estimated
#> 10   THETA10  2.301570000 0.372585669781932  0.00857532 estimated
#> 11   THETA11  0.100030000  2.82643207037889  0.00282728 estimated
#> 12  OMEGA1.1  0.010282100  9.41631573316735 0.000968195 estimated
#> 13  OMEGA2.1  0.000191938  431.130365013702 0.000827503 estimated
#> 14  OMEGA2.2  0.007948230  17.2945423069035  0.00137461 estimated
#> 15  OMEGA3.1  0.001281790  99.6434673386436  0.00127722 estimated
#> 16  OMEGA3.2 -0.000170245 -843.648858997327  0.00143627 estimated
#> 17  OMEGA3.3  0.010006800  30.4393012751329    0.003046 estimated
#> 18  OMEGA4.1 -0.000595539 -168.447406467083  0.00100317 estimated
#> 19  OMEGA4.2  0.000542183  202.204790633421  0.00109632 estimated
#> 20  OMEGA4.3  0.002000550  107.083552023194  0.00214226 estimated
#> 21  OMEGA4.4  0.009670120  19.8908596790939  0.00192347 estimated
#> 22  SIGMA1.1  1.000000000             .....       .....     fixed
#> 
#> $`First Order Conditional Estimation with Interaction (No Prior)`
#>    Parameter     Estimate           CV.perc          SE estimated
#> 1     THETA1  3.305220000 0.988049207011939   0.0326572 estimated
#> 2     THETA2  3.257970000 0.879277586963662   0.0286466 estimated
#> 3     THETA3 -0.611824000 -1.55615994142106  0.00952096 estimated
#> 4     THETA4 -0.208226000 -3.99087529895402  0.00831004 estimated
#> 5     THETA5  0.735323000   5.3271283503984   0.0391716 estimated
#> 6     THETA6  1.139920000   3.1377640536178    0.035768 estimated
#> 7     THETA7  0.336296000  3.36742036777125   0.0113245 estimated
#> 8     THETA8  0.191801000  5.39971115896163   0.0103567 estimated
#> 9     THETA9  0.694545000   1.5072313529001   0.0104684 estimated
#> 10   THETA10  2.301570000 0.372585669781932  0.00857532 estimated
#> 11   THETA11  0.100030000  2.82643207037889  0.00282728 estimated
#> 12  OMEGA1.1  0.010282100  9.41631573316735 0.000968195 estimated
#> 13  OMEGA2.1  0.000191938  431.130365013702 0.000827503 estimated
#> 14  OMEGA2.2  0.007948230  17.2945423069035  0.00137461 estimated
#> 15  OMEGA3.1  0.001281790  99.6434673386436  0.00127722 estimated
#> 16  OMEGA3.2 -0.000170245 -843.648858997327  0.00143627 estimated
#> 17  OMEGA3.3  0.010006800  30.4393012751329    0.003046 estimated
#> 18  OMEGA4.1 -0.000595539 -168.447406467083  0.00100317 estimated
#> 19  OMEGA4.2  0.000542183  202.204790633421  0.00109632 estimated
#> 20  OMEGA4.3  0.002000550  107.083552023194  0.00214226 estimated
#> 21  OMEGA4.4  0.009670120  19.8908596790939  0.00192347 estimated
#> 22  SIGMA1.1  1.000000000             .....       .....     fixed
#> 

tmp$monitor

#> $`Iterative Two Stage (No Prior)`
#>         0         5        10        15        20        25        30        35        40        45        50        55        60        65        70 
#>  43391.70 -10716.47 -10763.01 -10768.73 -10770.40 -10771.08 -10771.41 -10771.59 -10771.69 -10771.76 -10771.80 -10771.82 -10771.84 -10771.84 -10771.85 
#>        75        80        85        90        95       100       105       110       115       120       125       130       135       140       145 
#> -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 
#>       150       155       160       165       170       175       180       185       190       195       200 
#> -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 -10771.86 
#> 
#> $`Importance Sampling (No Prior)`
#>         0         1         2         3         4         5         6         7         8         9        10        11        12        13        14 
#> -10781.56 -10781.91 -10781.78 -10782.46 -10781.39 -10782.74 -10784.55 -10782.07 -10781.89 -10782.42 -10785.72 -10781.99 -10783.01 -10782.82 -10784.40 
#>        15        16        17        18        19        20        21        22        23        24        25        26        27        28        29 
#> -10781.43 -10780.02 -10781.79 -10781.33 -10780.36 -10782.81 -10783.50 -10782.73 -10778.68 -10780.78 -10779.90 -10779.94 -10785.65 -10778.24 -10783.85 
#>        30        31        32        33        34        35        36        37        37 
#> -10781.36 -10779.76 -10779.90 -10782.58 -10782.48 -10781.61 -10782.31 -10779.77 -10783.92 
#> 
#> $`Stochastic Approximation Expectation-Maximization (No Prior)`
#>     -3000     -2990     -2980     -2970     -2960     -2950     -2940     -2930     -2920     -2910     -2900     -2890     -2880     -2870         0 
#> -19904.69 -19867.43 -19694.05 -19714.93 -19682.76 -19645.50 -19590.94 -19599.59 -19398.13 -19568.94 -19536.53 -19503.30 -19589.66 -19505.91 -19548.29 
#>        10        20        30        40        50        60        70        80        90       100       110       120       130       140       150 
#> -19871.07 -19915.51 -19947.26 -19947.70 -19951.32 -19960.73 -19965.45 -19967.37 -19972.69 -19975.31 -19977.98 -19982.08 -19982.21 -19981.39 -19981.63 
#>       160       170       180       190       200       210       220       230       240       250       260       270       280       290       300 
#> -19981.89 -19982.25 -19982.35 -19982.16 -19981.89 -19981.80 -19981.45 -19980.36 -19979.15 -19978.47 -19978.39 -19979.11 -19979.15 -19979.37 -19980.63 
#>       310       320       330       340       350       360       370       380       390       400       410       420       430       440       450 
#> -19981.47 -19982.69 -19983.88 -19984.41 -19985.05 -19985.26 -19985.48 -19985.78 -19986.25 -19985.67 -19985.79 -19986.07 -19985.93 -19986.48 -19986.27 
#>       460       470       480       490       500       510       520       530       540       550       560       570       580       590       600 
#> -19986.83 -19986.55 -19986.74 -19986.78 -19986.86 -19986.44 -19986.41 -19986.59 -19987.46 -19987.79 -19988.01 -19988.88 -19989.71 -19989.90 -19990.02 
#>       610       620       630       640       650       660       670       680       690       700       710       720       730       740       750 
#> -19989.93 -19989.83 -19989.64 -19989.80 -19989.72 -19989.74 -19990.24 -19990.52 -19990.45 -19989.98 -19989.85 -19989.49 -19989.19 -19989.05 -19989.05 
#>       760       770       780       790       800       810       820       830       840       850       860       870       880       890       900 
#> -19989.03 -19989.48 -19989.91 -19989.85 -19989.42 -19989.24 -19989.11 -19989.39 -19989.50 -19989.71 -19989.77 -19989.93 -19990.14 -19989.90 -19990.20 
#>       910       920       930       940       950       960       970       980       990      1000      1010      1020      1030      1040      1050 
#> -19990.77 -19990.70 -19990.66 -19990.79 -19991.05 -19991.18 -19991.47 -19991.71 -19992.08 -19992.53 -19992.61 -19992.58 -19992.35 -19991.97 -19991.94 
#>      1060      1070      1080      1090      1100      1110      1120      1130      1140      1150      1160      1170      1180      1190      1200 
#> -19991.96 -19992.26 -19992.50 -19992.84 -19993.12 -19993.33 -19993.32 -19993.48 -19993.48 -19993.28 -19993.15 -19993.48 -19993.57 -19993.76 -19993.97 
#>      1210      1220      1230      1240      1250      1260      1270      1280      1290      1300      1310      1320      1330      1340      1350 
#> -19993.76 -19993.85 -19993.83 -19993.75 -19993.71 -19994.13 -19994.24 -19994.26 -19994.33 -19994.30 -19994.04 -19994.02 -19994.24 -19994.45 -19994.52 
#>      1360      1370      1380      1390      1400      1410      1420      1430      1440      1450      1460      1470      1480      1490      1500 
#> -19994.62 -19994.53 -19994.27 -19994.04 -19993.88 -19993.76 -19993.60 -19993.56 -19993.04 -19992.96 -19992.85 -19992.68 -19992.55 -19992.63 -19992.44 
#>      1510      1520      1530      1540      1550      1560      1570      1580      1590      1600      1610      1620      1630      1640      1650 
#> -19992.52 -19992.86 -19992.89 -19992.88 -19992.89 -19992.80 -19992.86 -19992.95 -19992.77 -19992.64 -19992.93 -19992.83 -19992.86 -19992.76 -19992.91 
#>      1660      1670      1680      1690      1700      1710      1720      1730      1740      1750      1760      1770      1780      1790      1800 
#> -19992.96 -19992.98 -19993.02 -19993.21 -19993.25 -19993.38 -19993.44 -19993.39 -19993.29 -19993.23 -19993.05 -19992.95 -19993.14 -19993.11 -19993.01 
#>      1810      1820      1830      1840      1850      1860      1870      1880      1890      1900      1910      1920      1930      1940      1950 
#> -19992.89 -19992.84 -19992.84 -19993.04 -19993.05 -19993.06 -19993.01 -19992.94 -19993.13 -19993.09 -19993.08 -19993.06 -19993.20 -19993.28 -19993.36 
#>      1960      1970      1980      1990      2000 
#> -19993.33 -19993.42 -19993.55 -19993.65 -19993.88 
#> 
#> $`Objective Function Evaluation by Importance Sampling (No Prior)`
#>         0         1         2         3         4         5 
#> -10776.76 -10776.87 -10777.54 -10776.67 -10776.07 -10777.68 
#> 
#> $`MCMC Bayesian Analysis`
#>    -10000     -9900     -9800     -9700     -9600     -9500     -9400     -9300     -9200     -9100     -9000         0       100       200       300 
#>         0  -1968720  -3907800  -5834437  -7770001  -9708001 -11644983 -13577312 -15513872 -17450136 -19379332         0  -1915303  -3832383  -5758395 
#>       400       500       600       700       800       900      1000 
#>  -7677565  -9593900 -11510444 -13426121 -15343141 -17261970 -19185051 
#> 
#> $`First Order Conditional Estimation with Interaction (No Prior)`
#>         0         5        10        15        20        25 
#> -10769.55 -10770.79 -10770.99 -10771.87 -10772.00 -10772.14 
#> 

tmp = nm.extract.xml("example2", path = getOption("qpExampleDir"))

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> Joining, by = "Parameter"

#> extraction of example2.xml complete.