Insert doses in other compartments

To model drug entry into the system from multiple absorption sites this function allows to turn a NONMEM datset into one with additional doses in differetn compartments.

nm.insert.dose.multiple.compartments(
  data,
  dose.in.cmt = c(1, 2),
  ...,
  quiet = TRUE
)

Arguments

data
NONMEM data.frame, should contain variable 'CMT'

dose.in.cmt
numeric double or vector for new CMT doses

...
(unquoted) columns to sort the result by

quiet
when TRUE (default) silences the messages returned

Value

A data.frame with input dataset supplied with additional doses

See also

Examples

library(dplyr)
## original data:
nmData = example.NONMEM.dataset(TIME=seq(0,24,4))
nmData$CMT = with(nmData, swap(EVID, 0:1, 2:1))

## adding doses to compartments 3, 4 and 5
nmData3 = nm.insert.dose.multiple.compartments(data = nmData
                                               , dose.in.cmt=c(3,4,5)
                                               , ID, TIME, -EVID

)
tbl_df(nmData3[, c('ID','TIME','EVID','AMT','CMT','DV')])

#> # A tibble: 99 x 6
#>       ID  TIME  EVID   AMT   CMT    DV
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1     0     1     1     1     0
#>  2     1     0     1     1     3     0
#>  3     1     0     1     1     4     0
#>  4     1     0     1     1     5     0
#>  5     1     0     0     0     2     0
#>  6     1     4     0     0     2     0
#>  7     1     8     0     0     2     0
#>  8     1    12     0     0     2     0
#>  9     1    16     0     0     2     0
#> 10     1    20     0     0     2     0
#> # … with 89 more rows