Create NONMEM Parameter Estimate Table

Create NONMEM Parameter Estimate Table

nm.params.table(
  run,
  path = getOption("nmDir"),
  file.ext = ".ext",
  runIndex,
  fixed.text = ".....",
  return.all = FALSE
)

Arguments

run
run rootname (e.g. run1)

path
directory where rootname.ext resides

file.ext
file extentions of the .ext file. Defaults to .ext

runIndex
which (numeric) estimation method needs to be tabulated? Defaults to the last estimation method outputted to the .ext file

fixed.text
character string to replace the entries for standard error and coefficient of variation (CV) for parameters that were fixed in the control stream

return.all
logical indicator to return the standard data set for inclusion in reports (default) or if TRUE the data set with everything.

Value

data frame with parameter name, estimate, coefficient of variation, standard error, and estimated/fixed information.

See also

Examples

nm.params.table(run = "example1", path = getOption("qpExampleDir"))

#> Joining, by = "Parameter"

#>    Parameter     Estimate           CV.perc         SE estimated
#> 1     THETA1  1.686930000  2.55967941764033    0.04318 estimated
#> 2     THETA2  1.611290000  2.95483122218843  0.0476109 estimated
#> 3     THETA3  0.819605000  8.53864971541169  0.0699832 estimated
#> 4     THETA4  2.391610000  2.15700720435188  0.0515872 estimated
#> 5   OMEGA1.1  0.165061000  16.4708198787115  0.0271869 estimated
#> 6   OMEGA2.1 -0.000741979  -2803.0712459517  0.0207982 estimated
#> 7   OMEGA2.2  0.131429000  23.8227484040813    0.03131 estimated
#> 8   OMEGA3.1  0.012411200  240.912240556916  0.0299001 estimated
#> 9   OMEGA3.2  0.015956500  223.893711026854  0.0357256 estimated
#> 10  OMEGA3.3  0.187548000  38.3405848102886   0.071907 estimated
#> 11  OMEGA4.1 -0.012735500 -185.120332927643   0.023576 estimated
#> 12  OMEGA4.2  0.013906100  173.090945700088  0.0240702 estimated
#> 13  OMEGA4.3  0.033270200  127.986005494406  0.0425812 estimated
#> 14  OMEGA4.4  0.149906000  25.4973116486331   0.038222 estimated
#> 15  SIGMA1.1  0.057163200  13.2438876759873 0.00757063 estimated

nm.params.table("example2",  path = getOption("qpExampleDir"),
fixed.text = "(fixed to 0)", return.all = TRUE)

#> Joining, by = "Parameter"

#>        Estimate           SE     Eigen     CondNum     sd.var   sd.sd.var Parameter index level   ord estimated           CV.perc      Run
#> 1   3.305220000    0.0326572 0.0258399 158.3920000  0.0000000 0.00000e+00    THETA1   1.0 THETA 1 1.0 estimated 0.988049207011939 example2
#> 2   3.257970000    0.0286466 0.0276883   0.0258399  0.0000000 0.00000e+00    THETA2   2.0 THETA 1 2.0 estimated 0.879277586963662 example2
#> 3  -0.611824000   0.00952096 0.0352403   4.0928400  0.0000000 0.00000e+00    THETA3   3.0 THETA 1 3.0 estimated -1.55615994142106 example2
#> 4  -0.208226000   0.00831004 0.0444284   0.0000000  0.0000000 0.00000e+00    THETA4   4.0 THETA 1 4.0 estimated -3.99087529895402 example2
#> 5   0.735323000    0.0391716 0.0596110   0.0000000  0.0000000 0.00000e+00    THETA5   5.0 THETA 1 5.0 estimated   5.3271283503984 example2
#> 6   1.139920000     0.035768 0.2052890   0.0000000  0.0000000 0.00000e+00    THETA6   6.0 THETA 1 6.0 estimated   3.1377640536178 example2
#> 7   0.336296000    0.0113245 0.2703130   0.0000000  0.0000000 0.00000e+00    THETA7   7.0 THETA 1 7.0 estimated  3.36742036777125 example2
#> 8   0.191801000    0.0103567 0.2981540   0.0000000  0.0000000 0.00000e+00    THETA8   8.0 THETA 1 8.0 estimated  5.39971115896163 example2
#> 9   0.694545000    0.0104684 0.4992870   0.0000000  0.0000000 0.00000e+00    THETA9   9.0 THETA 1 9.0 estimated   1.5072313529001 example2
#> 10  2.301570000   0.00857532 0.5624250   0.0000000  0.0000000 0.00000e+00   THETA10  10.0 THETA 110.0 estimated 0.372585669781932 example2
#> 11  0.100030000   0.00282728 0.6202310   0.0000000  0.0000000 0.00000e+00   THETA11  11.0 THETA 111.0 estimated  2.82643207037889 example2
#> 12  0.010282100  0.000968195 1.1033500   0.0000000  0.1014010 4.77410e-03  OMEGA1.1   1.1 OMEGA 2 1.1 estimated  9.41631573316735 example2
#> 13  0.000191938  0.000827503 1.3254500   0.0000000  0.0212316 9.06201e-02  OMEGA2.1   2.1 OMEGA 2 2.1 estimated  431.130365013702 example2
#> 14  0.007948230   0.00137461 1.3571500   0.0000000  0.0891528 7.70929e-03  OMEGA2.2   2.2 OMEGA 2 2.2 estimated  17.2945423069035 example2
#> 15  0.001281790   0.00127722 1.4866100   0.0000000  0.1263650 1.14875e-01  OMEGA3.1   3.1 OMEGA 2 3.1 estimated  99.6434673386436 example2
#> 16 -0.000170245   0.00143627 1.5644400   0.0000000 -0.0190894 1.62476e-01  OMEGA3.2   3.2 OMEGA 2 3.2 estimated -843.648858997327 example2
#> 17  0.010006800     0.003046 1.5993100   0.0000000  0.1000340 1.52248e-02  OMEGA3.3   3.3 OMEGA 2 3.3 estimated  30.4393012751329 example2
#> 18 -0.000595539   0.00100317 2.4044700   0.0000000 -0.0597246 1.04157e-01  OMEGA4.1   4.1 OMEGA 2 4.1 estimated -168.447406467083 example2
#> 19  0.000542183   0.00109632 2.7388500   0.0000000  0.0618437 1.21953e-01  OMEGA4.2   4.2 OMEGA 2 4.2 estimated  202.204790633421 example2
#> 20  0.002000550   0.00214226 4.0928400   0.0000000  0.2033690 1.75727e-01  OMEGA4.3   4.3 OMEGA 2 4.3 estimated  107.083552023194 example2
#> 21  0.009670120   0.00192347 0.0000000   0.0000000  0.0983368 9.78003e-03  OMEGA4.4   4.4 OMEGA 2 4.4 estimated  19.8908596790939 example2
#> 22  1.000000000 (fixed to 0) 0.6790310   0.0000000  1.0000000 1.00000e+10  SIGMA1.1   1.1 SIGMA 3 1.1     fixed      (fixed to 0) example2