Parse NONMEM control stream

Function to read NONMEM control stream and parse it out as named list

nm.parse.control.stream(
  run,
  path = getOption("nmDir"),
  file.ext = ".mod",
  quiet = TRUE
)

Arguments

run
run rootname (e.g. run1)

path
directory where run resides

file.ext
file extension of NONMEM control stream (.mod)

quiet
if TRUE no message will be returned.

Value

A named list with all $PK, $ERROR, $THETA etc... elements of the control stream

Examples

nm.parse.control.stream("example1", path = getOption("qpExampleDir"), file.ext = ".ctl")

#> $PROB
#>  [1] "$PROB RUN# Example 1 (from samp5l)"                                                             
#>  [2] ";; 1. Based on:0"                                                                               
#>  [3] ";; 2. Description:"                                                                             
#>  [4] ";;    2 compartment model"                                                                      
#>  [5] ";; 3. Label:"                                                                                   
#>  [6] ";;    2CMT"                                                                                     
#>  [7] ";; 4. Structural model:"                                                                        
#>  [8] ";;    ADVAN3 TRANS4"                                                                            
#>  [9] ";; 5. Covariate model:"                                                                         
#> [10] ";;    None"                                                                                     
#> [11] ";; 6. Inter-individual variability:"                                                            
#> [12] ";;    CL, V1, Q, V2"                                                                            
#> [13] ";; 7. Inter-occasion variability:"                                                              
#> [14] ";;    None"                                                                                     
#> [15] ";; 8. Residual variability:"                                                                    
#> [16] ";;    proportional"                                                                             
#> [17] ";; 9. Estimation:"                                                                              
#> [18] ";;    FOCE INTER, SAEM, BAYES, IMP"                                                             
#> [19] "; carries over to the next $EST statement, within a $PROB).  The SAEM is a Monte Carlo process,"
#> [20] "; so setting the SEED assures repeatability of results.  Each iteration obtains only 2 Monte"   
#> [21] "; Carlo samples ISAMPLE), so they are very fast.  But many iterations are needed, so PRINT only"
#> [22] "; every 100th iteration.  After the stochastic phase, 500 accumulation iterations will be"      
#> [23] "; Performed (NITER), to obtain good parameters estimates with little stochastic noise."         
#> [24] "; As a new FILE has not been given, the SAEM results will append to example1.ext."              
#> 
#> $INPUT
#> [1] "$INPUT C SET ID JID TIME  DV=CONC AMT=DOSE RATE EVID MDV CMT CLX V1X QX V2X SDIX SDSX"
#> 
#> $DATA
#> [1] "$DATA example1.csv IGNORE=C"
#> 
#> $SUBROUTINES
#> [1] "$SUBROUTINES ADVAN3 TRANS4"                                                                
#> [2] ";NTHETA=number of Thetas to be estimated"                                                  
#> [3] ";NETA=number of Etas to be estimated (and to be described by NETAxBETA OMEGA matrix)"      
#> [4] ";NTHP=number of thetas which have a prior"                                                 
#> [5] ";NETP=number of Omegas with prior"                                                         
#> [6] ";Prior information is important for MCMC Bayesian analysis, not necessary for maximization"
#> [7] "; methods"                                                                                 
#> 
#> $PRIOR
#> [1] "$PRIOR NWPRI NTHETA=4, NETA=4, NTHP=4, NETP=4"
#> 
#> $PK
#>  [1] "$PK"                                                                                          
#>  [2] "; The thetas are MU modeled.  Best that there is a linear relationship between THETAs and Mus"
#>  [3] ";  The linear MU modeling of THETAS allows them to be efficiently Gibbs sampled."             
#>  [4] "MU_1=THETA(1)"                                                                                
#>  [5] "MU_2=THETA(2)"                                                                                
#>  [6] "MU_3=THETA(3)"                                                                                
#>  [7] "MU_4=THETA(4)"                                                                                
#>  [8] "CL=DEXP(MU_1+ETA(1))"                                                                         
#>  [9] "V1=DEXP(MU_2+ETA(2))"                                                                         
#> [10] "Q=DEXP(MU_3+ETA(3))"                                                                          
#> [11] "V2=DEXP(MU_4+ETA(4))"                                                                         
#> [12] "S1=V1"                                                                                        
#> [13] "STRT = 0"                                                                                     
#> [14] "IF(ID.GT.50) STRT = 1"                                                                        
#> 
#> $ERROR
#> [1] "$ERROR"                    "IPRED = F"                 "Y = F + F*EPS(1)"          "; Initial values of THETA"
#> 
#> $THETA
#>  [1] "$THETA "                                                                                          
#>  [2] "(0.001, 2.0) ;[LN(CL)]"                                                                           
#>  [3] "(0.001, 2.0) ;[LN(V1)]"                                                                           
#>  [4] "(0.001, 2.0) ;[LN(Q)]"                                                                            
#>  [5] "(0.001, 2.0) ;[LN(V2)]"                                                                           
#>  [6] ";INITIAL values of OMEGA"                                                                         
#>  [7] "$THETA (2.0 FIX) (2.0 FIX) (2.0 FIX) (2.0 FIX)"                                                   
#>  [8] "; Variance to prior information of THETAS.  Because variances are very large, this"               
#>  [9] "; means that the prior information to the THETAS is highly uninformative."                        
#> [10] "$THETA (4 FIX)"                                                                                   
#> [11] "; The first analysis is iterative two-stage, maximum of 500 iterations (NITER), iteration results"
#> [12] "; are printed every 5 iterations, gradient precision (SIGL) is 4. Termination is tested on all of"
#> [13] "; the population parameters (CTYPE=3), and for less then 2 significant digits change (NSIG)."     
#> [14] "; Prior information is not necessary for ITS, so NOPRIOR=1.  The intermediate and final results"  
#> [15] "; of the ITS method will be recoded in row/column format in example1.ext"                         
#> 
#> $OMEGA
#>  [1] "$OMEGA BLOCK(4)"                                                                                 
#>  [2] "0.15   ;[P]"                                                                                     
#>  [3] "0.01  ;[F]"                                                                                      
#>  [4] "0.15   ;[P]"                                                                                     
#>  [5] "0.01  ;[F]"                                                                                      
#>  [6] "0.01  ;[F]"                                                                                      
#>  [7] "0.15   ;[P]"                                                                                     
#>  [8] "0.01  ;[F]"                                                                                      
#>  [9] "0.01  ;[F]"                                                                                      
#> [10] "0.01  ;[F]"                                                                                      
#> [11] "0.15   ;[P]"                                                                                     
#> [12] ";Initial value of SIGMA"                                                                         
#> [13] "$OMEGA BLOCK(4)"                                                                                 
#> [14] "10000 FIX "                                                                                      
#> [15] "0.00 10000"                                                                                      
#> [16] "0.00  0.00 10000"                                                                                
#> [17] "0.00  0.00 0.0 10000"                                                                            
#> [18] "; Prior information to the OMEGAS."                                                              
#> [19] "$OMEGA BLOCK(4)"                                                                                 
#> [20] "0.2 FIX "                                                                                        
#> [21] "0.0  0.2 "                                                                                       
#> [22] "0.0  0.0 0.2"                                                                                    
#> [23] "0.0  0.0 0.0 0.2"                                                                                
#> [24] ";Degrees of freedom to prior OMEGA matrix.  Because degrees of freedom is very low, equal to the"
#> [25] "; the dimension of the prior OMEGA, this means that the prior information to the OMEGAS is"      
#> [26] "; highly uninformative"                                                                          
#> 
#> $SIGMA
#> [1] "$SIGMA "                       "(0.6 )   ;[P]"                 "; Prior information of THETAS"
#> 
#> $EST
#>  [1] "$EST METHOD=ITS INTERACTION FILE=example1.ext NITER=500 PRINT=5 NOABORT SIGL=4 CTYPE=3 CITER=10   "
#>  [2] "     CALPHA=0.05 NOPRIOR=1 NSIG=2"                                                                 
#>  [3] "; The results of ITS are used as the initial values for the SAEM method.  A maximum of 3000"       
#>  [4] "; stochastic iterations (NBURN) is requested, but may end early if statistical test determines"    
#>  [5] "; that variations in all parameters is stationary (note that any settings from the previous $EST"  
#>  [6] "; carries over to the next $EST statement, within a $PROB).  The SAEM is a Monte Carlo process,"   
#>  [7] "; so setting the SEED assures repeatability of results.  Each iteration obtains only 2 Monte"      
#>  [8] "; Carlo samples ISAMPLE), so they are very fast.  But many iterations are needed, so PRINT only"   
#>  [9] "; every 100th iteration.  After the stochastic phase, 500 accumulation iterations will be"         
#> [10] "; Performed (NITER), to obtain good parameters estimates with little stochastic noise."            
#> [11] "; As a new FILE has not been given, the SAEM results will append to example1.ext."                 
#> [12] "$EST METHOD=SAEM INTERACTION NBURN=3000 NITER=500 PRINT=100 SEED=1556678 ISAMPLE=2"                
#> [13] "; After the SAEM method, obtain good estimates of the marginal density (objective function),"      
#> [14] "; along with good estimates of the standard errors.  This is best done with importance sampling"   
#> [15] "; (IMP), performing the expectation step only (EONLY=1), so that final population parameters"      
#> [16] "; remain at the final SAEM result.  Five iterations (NITER) should allow the importance sampling"  
#> [17] "; proposal density to become stationary.  This is observed by the objective function settling "    
#> [18] "; to a particular value (with some stochastic noise).  By using 3000 Monte Carlo samples"          
#> [19] "; (ISAMPLE), this assures a precise assessment of standard errors."                                
#> [20] "$EST METHOD=IMP MAPITER=0 INTERACTION EONLY=1 NITER=5 ISAMPLE=3000 PRINT=1 SIGL=8 NOPRIOR=1"       
#> [21] "; The Bayesian analysis is performed.  While 10000 burn-in"                                        
#> [22] "; iterations are requested as a maximum, because the termination test is on (CTYPE<>0, set at the" 
#> [23] "; first $EST statement), and because the initial parameters are at the SAEM result, which is the"  
#> [24] "; maximum likelihood position, the analysis should settle down to a stationary distribution in"    
#> [25] "; several hundred iterations.  Prior information is also used to facilitate Bayesian analysis."    
#> [26] "; The individual Bayesian iteration results are important, and may be need for post-processing"    
#> [27] "; analysis. So specify a separate FILE for the Bayesian analysis. "                                
#> [28] "$EST METHOD=BAYES INTERACTION FILE=example1.txt NBURN=10000 NITER=10000 PRINT=100 NOPRIOR=0"       
#> [29] "; Just for old-times sake, let?s see what the traditional FOCE method will give us.  "             
#> [30] "; And, remember to introduce a new FILE, so its results won?t append to our Bayesian FILE. "       
#> [31] "; Appending to example1.ext with the EM methods is fine."                                          
#> [32] "$EST METHOD=COND INTERACTION MAXEVAL=9999 NSIG=3 SIGL=10 PRINT=5 NOABORT NOPRIOR=1"                
#> [33] "     FILE=example1.ext"                                                                            
#> [34] "; Time for the standard error results.  You may request a more precise gradient precision (SIGL)"  
#> [35] "; that differed from that used during estimation."                                                 
#> 
#> $COV
#> [1] "$COV MATRIX=R PRINT=E UNCONDITIONAL SIGL=12"                                   
#> [2] "; Print out results in tables. Include some of the new weighted residual types"
#> 
#> $TABLE
#> [1] "$TABLE ID TIME DV IPRED PRED EVID RES WRES CPRED CWRES EPRED ERES EWRES NOAPPEND ONEHEADER "
#> [2] " FILE=example1.sdtab NOPRINT"                                                               
#> [3] "$TABLE ID CL V1 Q V2 ETA1 ETA2 ETA3 ETA4 NOAPPEND ONEHEADER NOPRINT FILE=example1.patab"    
#>