Parsing SCM output

Read and process SCM output file such that the output can be directly copied into Microsoft products and LaTeX typesetting.

nm.process.scm(path, digits = 3)

Arguments

path
full directory path where scmlog files reside

digits
number of significant digits

Value

A character vector with processed SCM output, ready for inclusion into a (LaTeX) report

Examples

library(xtable)
## the code below works if a number of scmlog.txt files are in folder NONMEM/scm files
myPath = file.path(getOption("qpExampleDir"),"scm_example2")
scm = nm.process.scm(myPath)
names(scm)

#> [1] "full.scm"  "summary"   "model"     "steps"     "scm.latex"

# Fail if scm file not found
foo <- try(nm.process.scm(file.path(getOption("qpExampleDir"),"scm_example0")))

#> Error in nm.process.scm(file.path(getOption("qpExampleDir"), "scm_example0")) : 
#>   cannot find /tmp/Rtmp59s2Ad/temp_libpath378afcc4f59/qpToolkit/NONMEM/scm_example0/scmlog.txt

scm$summary

#>      MODEL TEST     BASE.OFV      NEW.OFV       TEST OFV.DROP    GOAL dDF SIGNIFICANT      PVAL
#> 1 CLGNDR-2 PVAL  -5494.54612  -9557.26778 4062.72165        > 6.63490   1        YES! 0.00e+000
#> 2  CLAGE-5 PVAL  -9557.26778  -9746.45167  189.18389        > 6.63490   1        YES! 4.79e-043
#> 3  V1AGE-5 PVAL  -9746.45167 -10176.09933  429.64766        > 6.63490   1        YES! 0.00e+000
#> 4 V1GNDR-2 PVAL -10176.09933 -10195.27729   19.17797        > 6.63490   1        YES!  0.000012

# populate an appendix with scm output
print(xtable(scm$scm.latex
   , caption = "Full Stepwise Covariate Building output"
   , label =  'Tab:scmfull')
      , sanitize.text.function = function(x)x
      , include.rownames = FALSE
      , size = "small"
      , caption.placement = "top"
      , booktabs = TRUE
      , tabular.environment = "longtable"
      , floating = FALSE
      #  , only.contents = TRUE
)

#> % latex table generated in R 3.6.2 by xtable 1.8-4 package
#> % Fri Jan 17 23:28:55 2020
#> \begingroup\small
#> \begin{longtable}{llllllll}
#> \caption{Full Stepwise Covariate Building output} \\ 
#>   \toprule
#> MODEL & OFV$_{base}$ & OFV$_{test}$ & $\Delta$OFV & GOAL & $\Delta$DF & SIGNIFICANT & P value \\ 
#>   \midrule
#> $\textit{START FORWARD STEP 1}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   CLAGE-5 & $\phantom{0}$-5494.5 & $\phantom{0}$-8791.9 & $\phantom{}$3297.3 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   CLGNDR-2 & $\phantom{0}$-5494.5 & $\phantom{0}$-9557.3 & $\phantom{}$4062.7 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   QAGE-5 & $\phantom{0}$-5494.5 & $\phantom{0}$-6123 & $\phantom{0}$628.5 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   QGNDR-2 & $\phantom{0}$-5494.5 & $\phantom{0}$-8789.9 & $\phantom{}$3295.4 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V1AGE-5 & $\phantom{0}$-5494.5 & $\phantom{0}$-9145.8 & $\phantom{}$3651.3 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V1GNDR-2 & $\phantom{0}$-5494.5 & $\phantom{0}$-9049.8 & $\phantom{}$3555.3 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V2AGE-5 & $\phantom{0}$-5494.5 & $\phantom{0}$-8810.1 & $\phantom{}$3315.6 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V2GNDR-2 & $\phantom{0}$-5494.5 & $\phantom{0}$-8790.2 & $\phantom{}$3295.7 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   $\textit{ADDED CL-GNDR-2}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   \cmidrule{2-8}% & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   $\textit{START FORWARD STEP 2}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   CLAGE-5 & $\phantom{0}$-9557.3 & $\phantom{0}$-9746.5 & $\phantom{0}$189.2 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   QAGE-5 & $\phantom{0}$-9557.3 & $\phantom{0}$-9560 & $\phantom{000}$2.8 & $>$ 6.63 & 1 &  & 0.0960 \\ 
#>   QGNDR-2 & $\phantom{0}$-9557.3 & $\phantom{0}$-9559.5 & $\phantom{000}$2.2 & $>$ 6.63 & 1 &  & 0.1370 \\ 
#>   V1AGE-5 & $\phantom{0}$-9557.3 & $\phantom{0}$-9658.2 & $\phantom{0}$100.9 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V1GNDR-2 & $\phantom{0}$-9557.3 & $\phantom{0}$-9565.9 & $\phantom{000}$8.6 & $>$ 6.63 & 1 & YES! & 0.0033 \\ 
#>   V2AGE-5 & $\phantom{0}$-9557.3 & $\phantom{0}$-9570.1 & $\phantom{00}$12.8 & $>$ 6.63 & 1 & YES! & 0.0003 \\ 
#>   V2GNDR-2 & $\phantom{0}$-9557.3 & $\phantom{0}$-9557.3 & $\phantom{000}$0 & $>$ 6.63 & 1 &  & 0.9510 \\ 
#>   $\textit{ADDED CL-AGE-5}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   \cmidrule{2-8}% & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   $\textit{START FORWARD STEP 3}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   QAGE-5 & $\phantom{0}$-9746.5 & $\phantom{0}$-9746.7 & $\phantom{000}$0.3 & $>$ 6.63 & 1 &  & 0.6170 \\ 
#>   QGNDR-2 & $\phantom{0}$-9746.5 & $\phantom{0}$-9748.5 & $\phantom{000}$2.1 & $>$ 6.63 & 1 &  & 0.1510 \\ 
#>   V1AGE-5 & $\phantom{0}$-9746.5 & $\phantom{}$-10176.1 & $\phantom{0}$429.6 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V1GNDR-2 & $\phantom{0}$-9746.5 & $\phantom{0}$-9752.9 & $\phantom{000}$6.4 & $>$ 6.63 & 1 &  & 0.0113 \\ 
#>   V2AGE-5 & $\phantom{0}$-9746.5 & $\phantom{0}$-9771.7 & $\phantom{00}$25.2 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V2GNDR-2 & $\phantom{0}$-9746.5 & $\phantom{0}$-9746.5 & $\phantom{000}$0 & $>$ 6.63 & 1 &  & 0.8840 \\ 
#>   $\textit{ADDED V1-AGE-5}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   \cmidrule{2-8}% & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   $\textit{START FORWARD STEP 4}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   QAGE-5 & $\phantom{}$-10176.1 & $\phantom{}$-10176.1 & $\phantom{000}$0 & $>$ 6.63 & 1 &  & 0.9760 \\ 
#>   QGNDR-2 & $\phantom{}$-10176.1 & $\phantom{}$-10177.8 & $\phantom{000}$1.7 & $>$ 6.63 & 1 &  & 0.1960 \\ 
#>   V1GNDR-2 & $\phantom{}$-10176.1 & $\phantom{}$-10195.3 & $\phantom{00}$19.2 & $>$ 6.63 & 1 & YES! & $<$0.0001 \\ 
#>   V2AGE-5 & $\phantom{}$-10176.1 & $\phantom{}$-10176.9 & $\phantom{000}$0.8 & $>$ 6.63 & 1 &  & 0.3790 \\ 
#>   V2GNDR-2 & $\phantom{}$-10176.1 & $\phantom{}$-10176.2 & $\phantom{000}$0.1 & $>$ 6.63 & 1 &  & 0.7180 \\ 
#>   $\textit{ADDED V1-GNDR-2}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   \cmidrule{2-8}% & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   $\textit{START FORWARD STEP 5}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   QAGE-5 & $\phantom{}$-10195.3 & $\phantom{}$-10195.3 & $\phantom{000}$0 & $>$ 6.63 & 1 &  & 0.9580 \\ 
#>   QGNDR-2 & $\phantom{}$-10195.3 & $\phantom{}$-10197.8 & $\phantom{000}$2.5 & $>$ 6.63 & 1 &  & 0.1130 \\ 
#>   V2AGE-5 & $\phantom{}$-10195.3 & $\phantom{}$-10195.9 & $\phantom{000}$0.6 & $>$ 6.63 & 1 &  & 0.4380 \\ 
#>   V2GNDR-2 & $\phantom{}$-10195.3 & $\phantom{}$-10195.8 & $\phantom{000}$0.5 & $>$ 6.63 & 1 &  & 0.4600 \\ 
#>   $\textit{NOTHING ADDED}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   \cmidrule{2-8}% & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   $\textit{START BACKWARD STEP 1}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   CLAGE-1 & $\phantom{}$-10195.3 & $\phantom{0}$-9671.9 & $\phantom{}$-523.4 & $>$ -10.8 & -1 &  & $<$0.0001 \\ 
#>   CLGNDR-1 & $\phantom{}$-10195.3 & $\phantom{0}$-9250.3 & $\phantom{}$-945 & $>$ -10.8 & -1 &  & $<$0.0001 \\ 
#>   V1AGE-1 & $\phantom{}$-10195.3 & $\phantom{0}$-9752.9 & $\phantom{}$-442.4 & $>$ -10.8 & -1 &  & $<$0.0001 \\ 
#>   V1GNDR-1 & $\phantom{}$-10195.3 & $\phantom{}$-10169 & $\phantom{0}$-26.3 & $>$ -10.8 & -1 &  & $<$0.0001 \\ 
#>   $\textit{NOTHING REMOVED}$ & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>   \cmidrule{2-8}% & ~ & ~ & ~ & $~$ ~ & ~ & ~ & ~ \\ 
#>    \bottomrule
#> \label{Tab:scmfull}
#> \end{longtable}
#> \endgroup