Extract control stream sections

Function to read NONMEM control stream and extract specific sections

nm.remove.section(secNames, ctl)

Arguments

secNames
character vector with sections to remove

ctl
character vector of scanned control stream

Value

A list with as many (named) element as there are $.... elements in the control stream

Note

This function is not intended to be called alone, but is called in the wrapper function nm.parse.control.stream

See also

Examples

nm.remove.section("EST"
   , read.mod("example1", path = getOption("qpExampleDir"), file.ext=".ctl")
   )[[2]]

#> $EST
#> $EST[[1]]
#> [1] "$EST METHOD=ITS INTERACTION FILE=example1.ext NITER=500 PRINT=5 NOABORT SIGL=4 CTYPE=3 CITER=10   "
#> [2] "     CALPHA=0.05 NOPRIOR=1 NSIG=2"                                                                 
#> [3] "; The results of ITS are used as the initial values for the SAEM method.  A maximum of 3000"       
#> [4] "; stochastic iterations (NBURN) is requested, but may end early if statistical test determines"    
#> 
#> $EST[[2]]
#> [1] "; that variations in all parameters is stationary (note that any settings from the previous $EST"
#> 
#> $EST[[3]]
#> [1] "; carries over to the next $EST statement, within a $PROB).  The SAEM is a Monte Carlo process,"
#> [2] "; so setting the SEED assures repeatability of results.  Each iteration obtains only 2 Monte"   
#> [3] "; Carlo samples ISAMPLE), so they are very fast.  But many iterations are needed, so PRINT only"
#> [4] "; every 100th iteration.  After the stochastic phase, 500 accumulation iterations will be"      
#> [5] "; Performed (NITER), to obtain good parameters estimates with little stochastic noise."         
#> [6] "; As a new FILE has not been given, the SAEM results will append to example1.ext."              
#> 
#> $EST[[4]]
#> [1] "$EST METHOD=SAEM INTERACTION NBURN=3000 NITER=500 PRINT=100 SEED=1556678 ISAMPLE=2"              
#> [2] "; After the SAEM method, obtain good estimates of the marginal density (objective function),"    
#> [3] "; along with good estimates of the standard errors.  This is best done with importance sampling" 
#> [4] "; (IMP), performing the expectation step only (EONLY=1), so that final population parameters"    
#> [5] "; remain at the final SAEM result.  Five iterations (NITER) should allow the importance sampling"
#> [6] "; proposal density to become stationary.  This is observed by the objective function settling "  
#> [7] "; to a particular value (with some stochastic noise).  By using 3000 Monte Carlo samples"        
#> [8] "; (ISAMPLE), this assures a precise assessment of standard errors."                              
#> 
#> $EST[[5]]
#> [1] "$EST METHOD=IMP MAPITER=0 INTERACTION EONLY=1 NITER=5 ISAMPLE=3000 PRINT=1 SIGL=8 NOPRIOR=1"      
#> [2] "; The Bayesian analysis is performed.  While 10000 burn-in"                                       
#> [3] "; iterations are requested as a maximum, because the termination test is on (CTYPE<>0, set at the"
#> 
#> $EST[[6]]
#> [1] "; first $EST statement), and because the initial parameters are at the SAEM result, which is the"
#> [2] "; maximum likelihood position, the analysis should settle down to a stationary distribution in"  
#> [3] "; several hundred iterations.  Prior information is also used to facilitate Bayesian analysis."  
#> [4] "; The individual Bayesian iteration results are important, and may be need for post-processing"  
#> [5] "; analysis. So specify a separate FILE for the Bayesian analysis. "                              
#> 
#> $EST[[7]]
#> [1] "$EST METHOD=BAYES INTERACTION FILE=example1.txt NBURN=10000 NITER=10000 PRINT=100 NOPRIOR=0"
#> [2] "; Just for old-times sake, let?s see what the traditional FOCE method will give us.  "      
#> [3] "; And, remember to introduce a new FILE, so its results won?t append to our Bayesian FILE. "
#> [4] "; Appending to example1.ext with the EM methods is fine."                                   
#> 
#> $EST[[8]]
#> [1] "$EST METHOD=COND INTERACTION MAXEVAL=9999 NSIG=3 SIGL=10 PRINT=5 NOABORT NOPRIOR=1"              
#> [2] "     FILE=example1.ext"                                                                          
#> [3] "; Time for the standard error results.  You may request a more precise gradient precision (SIGL)"
#> [4] "; that differed from that used during estimation."                                               
#> 
#>