like as.ordered with extended formatting

like as.ordered with extended formatting

orderedNumeric(x, prefix = "", suffix = "", paste.function = paste,
  special.x = NULL)

Arguments

x
a vector of numeric values

prefix
text string to preceed the number (like drug name)

suffix
text string to ome after the number (like mg)

paste.function
function to use when glueing together prefix, x, and suffix

special.x
list of length 2, with first (numeric) element being the number that will be exchanged for the 2nd element which should be a text string.

Value

A formatted character vector containing count and percentage of unique values in x.

Examples

out = expand.grid(ID=1:5, DOSE = c(0,500,100,4000))
factor(out$DOSE)

#>  [1] 0    0    0    0    0    500  500  500  500  500  100  100  100  100  100  4000 4000 4000 4000 4000
#> Levels: 0 100 500 4000

#Levels: 0 100 500 4000  ## W H A T ???
orderedNumeric(out$DOSE)

#>  [1]  0      0      0      0      0      500    500    500    500    500    100    100    100    100    100    4000   4000   4000   4000   4000 
#> Levels:  0   100   500   4000 

#Levels:  0   100   500   4000    #that's better, similar to ordered(out$DOSE)
orderedNumeric(out$DOSE, special.x = list(0, "Placebo"))

#>  [1] Placebo Placebo Placebo Placebo Placebo  500     500     500     500     500     100     100     100     100     100     4000    4000    4000  
#> [19]  4000    4000  
#> Levels: Placebo  100   500   4000 

#Levels: Placebo  100   500   4000 
orderedNumeric(out$DOSE, prefix = "DrugName ")

#>  [1] DrugName  0     DrugName  0     DrugName  0     DrugName  0     DrugName  0     DrugName  500   DrugName  500   DrugName  500   DrugName  500  
#> [10] DrugName  500   DrugName  100   DrugName  100   DrugName  100   DrugName  100   DrugName  100   DrugName  4000  DrugName  4000  DrugName  4000 
#> [19] DrugName  4000  DrugName  4000 
#> Levels: DrugName  0  DrugName  100  DrugName  500  DrugName  4000 

orderedNumeric(out$DOSE, prefix = "DrugName", suffix = "mcg")

#>  [1] DrugName 0 mcg    DrugName 0 mcg    DrugName 0 mcg    DrugName 0 mcg    DrugName 0 mcg    DrugName 500 mcg  DrugName 500 mcg  DrugName 500 mcg 
#>  [9] DrugName 500 mcg  DrugName 500 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 4000 mcg
#> [17] DrugName 4000 mcg DrugName 4000 mcg DrugName 4000 mcg DrugName 4000 mcg
#> Levels: DrugName 0 mcg DrugName 100 mcg DrugName 500 mcg DrugName 4000 mcg

orderedNumeric(out$DOSE, prefix = "DrugName", suffix = "mcg", special.x = list(0, "Placebo"))

#>  [1] Placebo           Placebo           Placebo           Placebo           Placebo           DrugName 500 mcg  DrugName 500 mcg  DrugName 500 mcg 
#>  [9] DrugName 500 mcg  DrugName 500 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 100 mcg  DrugName 4000 mcg
#> [17] DrugName 4000 mcg DrugName 4000 mcg DrugName 4000 mcg DrugName 4000 mcg
#> Levels: Placebo DrugName 100 mcg DrugName 500 mcg DrugName 4000 mcg

orderedNumeric(out$DOSE, prefix = "DrugName ", suffix = "mcg", special.x = list(0, "Vehicle"),
paste.function = paste0)

#>  [1] Vehicle          Vehicle          Vehicle          Vehicle          Vehicle          DrugName 500mcg  DrugName 500mcg  DrugName 500mcg 
#>  [9] DrugName 500mcg  DrugName 500mcg  DrugName 100mcg  DrugName 100mcg  DrugName 100mcg  DrugName 100mcg  DrugName 100mcg  DrugName 4000mcg
#> [17] DrugName 4000mcg DrugName 4000mcg DrugName 4000mcg DrugName 4000mcg
#> Levels: Vehicle DrugName 100mcg DrugName 500mcg DrugName 4000mcg