Plotting function for nls.predict and nlme.predict

Plots residual, predicted, or partial residial plots based on x generated by qP functions nlme.predict and nls.predict.

# S3 method for fit
plot(
  x,
  ...,
  newFunc,
  yLabel,
  xLabel,
  label.cex = 1.25,
  xLimits,
  yLimits,
  title,
  plot.title = TRUE,
  layout,
  aspect,
  mnplot = TRUE,
  seplot = TRUE,
  logX = FALSE,
  logY = FALSE,
  pointcol = gray[8],
  linecol = qp.blue,
  axis.lim.widener = 0.035,
  do.plot = TRUE,
  abline = NULL
)

Arguments

x
the output from nls.predict or nlme.predict, i.e. class 'fit'

...
any other arguments to be passed on to the lattice call

newFunc
the function to plot. If not suppliued will be taken from x

yLabel
same as ylab

xLabel
same as xlab

label.cex
fontmszie of axes labels

xLimits
same as xlim

yLimits
same as ylim

title
place a title above

plot.title
logical indicating if a plot title should be written

layout
will be passed on a lattice argument layout

aspect
lattice banking aspect

mnplot
logical indicating an average plot to be created as opposed to 'raw data

seplot
logical indicating is error bars need to be drawn

logX
logarithmic X axis?

logY
logarithmic Y axis?

pointcol
color of dots

linecol
color of

axis.lim.widener
scalar to stretch the x and y axis. Use in case data or predictions are not visible anymore

do.plot
Defaults to T (create the plot) if F it will output the prediction data

abline
will be passed on to the lattice call as is

Value

A plot

Examples

pkpdData = example.pkpdData()

DNase1 <- subset(DNase, Run == 1)

## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
summary(fm1DNase1)

#> 
#> Formula: density ~ SSlogis(log(conc), Asym, xmid, scal)
#> 
#> Parameters:
#>      Estimate Std. Error t value Pr(>|t|)    
#> Asym  2.34518    0.07815   30.01 2.17e-13 ***
#> xmid  1.48309    0.08135   18.23 1.22e-10 ***
#> scal  1.04146    0.03227   32.27 8.51e-14 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.01919 on 13 degrees of freedom
#> 
#> Number of iterations to convergence: 0 
#> Achieved convergence tolerance: 8.283e-06
#> 

fm1DNase1.predict = nls.predict(density ~ conc,object = fm1DNase1)

#> Error in eval(if ("data" %in% names(object)) object$data else mCall$data): object 'DNase1' not found

plot(fm1DNase1.predict)

#> Error in plot(fm1DNase1.predict): object 'fm1DNase1.predict' not found

fm1DNase1.predict = nls.predict(density ~ conc,object = fm1DNase1)

#> Error in eval(if ("data" %in% names(object)) object$data else mCall$data): object 'DNase1' not found

plot(fm1DNase1.predict, logX = TRUE)

#> Error in plot(fm1DNase1.predict, logX = TRUE): object 'fm1DNase1.predict' not found

EFF.1comp.1abs <- function(dose, tob, cl, v, ka, keo)
{
  # Effect-site concentration for 1-compartment model, 1st-order absorption

  kel = cl / v

  # Define coefficients
  A = 1/(kel-ka) / (keo-ka)
  B = 1/(ka-kel) / (keo-kel)
  C = 1/(ka-keo) / (kel-keo)

  # Return effect-site concentration
  dose*ka*keo/v * (A*exp(-ka*tob) + B*exp(-kel*tob) + C*exp(-keo*tob))
}
fit.PD004.nlme = nlme.run(model = value ~ base +
 EFF.1comp.1abs(dose, time, cl * exp(cl.eta), v, ka, keo)
 , data = subset(pkpdData,type == "PD" & dose > 0 & value > 0.5),
                          fixed = base + cl + v + ka + keo ~ 1,
                          random = cl.eta ~ 1,
                          groups = ~ id,
                          start = c(base = 1, cl = 1, v = 10
                          , ka = 1, keo = 0.01),
                          problem = "True Model",
                          reference = 4)

#> Error in subset(pkpdData, type == "PD" & dose > 0 & value > 0.5): object 'pkpdData' not found

summary(fit.PD004.nlme$object)

#> Error in summary(fit.PD004.nlme$object): object 'fit.PD004.nlme' not found

nlme.extract(fit.PD004.nlme$object)$table

#> Error in nlme.extract(fit.PD004.nlme$object): object 'fit.PD004.nlme' not found

# simple fit vs time
fit.PD004.pred.nlme = nlme.predict(func = value ~ time
   , fit.PD004.nlme$object)

#> Error in getData(object): object 'fit.PD004.nlme' not found

plot(fit.PD004.pred.nlme)

#> Error in plot(fit.PD004.pred.nlme): object 'fit.PD004.pred.nlme' not found

fit.PD004.pred.nlme = nlme.predict(func = value ~ time
   , fit.PD004.nlme$object, method = "partial.residuals")

#> Error in getData(object): object 'fit.PD004.nlme' not found

plot(fit.PD004.pred.nlme) ## this is the same: PARTIAL RESIDUALS

#> Error in plot(fit.PD004.pred.nlme): object 'fit.PD004.pred.nlme' not found

fit.PD004.pred.nlme = nlme.predict(func = value ~ time
   , fit.PD004.nlme$object, method = "prediction")

#> Error in getData(object): object 'fit.PD004.nlme' not found

plot(fit.PD004.pred.nlme) ## 'simple' prediction for all xCovariate

#> Error in plot(fit.PD004.pred.nlme): object 'fit.PD004.pred.nlme' not found

## note that prediction and partial residual type of model fit
## plots are very different

# fit vs time by dose
fit.PD004.pred.nlme = nlme.predict(func = value ~ time | dose
   , fit.PD004.nlme$object)

#> Error in getData(object): object 'fit.PD004.nlme' not found

plot(fit.PD004.pred.nlme)

#> Error in plot(fit.PD004.pred.nlme): object 'fit.PD004.pred.nlme' not found

fit.PD004.pred.nlme = nlme.predict(func = value ~ time | dose
   , fit.PD004.nlme$object, method = "residuals")

#> Error in getData(object): object 'fit.PD004.nlme' not found

plot(fit.PD004.pred.nlme, yLimits = c(-1,1))

#> Error in plot(fit.PD004.pred.nlme, yLimits = c(-1, 1)): object 'fit.PD004.pred.nlme' not found

plot(fit.PD004.pred.nlme, yLimits = c(-1,1)
   , abline = list(h = 0, lty = 2))

#> Error in plot(fit.PD004.pred.nlme, yLimits = c(-1, 1), abline = list(h = 0,     lty = 2)): object 'fit.PD004.pred.nlme' not found