Format parameter estimate table

Process parameter table for inclusion into LaTeX, or MS Excel or Word

process.parTable(
  nm,
  index = 1,
  ci = 0.95,
  digits = 3,
  formatted = TRUE,
  latex = FALSE,
  align.dot = FALSE,
  transformations,
  Descriptor,
  plain = FALSE,
  missing.format = ".....",
  remove.fixed.sigma = FALSE
)

Arguments

nm
output from nm.params.table

index
which estimation (only relevant in case of multiple $EST)

ci
confidence interval for 95% CI

digits
number of significant digits in output

formatted
whether to enforce length of significant digits, see formatted.signif

latex
passed to formatted.signif

align.dot
passed to formatted.signif

transformations
a list with elements log or logit, that, in turn, are numeric vectors representating the THETA numbers that are logged or logit-transformed, respectively.

Descriptor
vector of descriptor (alias) names for numbered parameter estimates. For example c("CL","V1","SD(CL)","cor(CL,V1)","SD(V1)","prop. residual var.")

plain
logical (defaults to F) indicating if LaTeX formatting of the table should be skipped, allowing the result to be easily available for Microsoft products.

missing.format
what to replace non-estimated / fixed parameter estimates, or

remove.fixed.sigma
logical option to remove sigma estimates from the table

Value

Either a LaTeX preformatted parameter estimate table (default) or a fully processed parameter estimate table without LaTeX preformatting (for inclusion into Microsoft products).

See also

Examples

process.parTable(
   nm = nm.params.table(run = "example1", path = getOption("qpExampleDir"), runIndex = 1)
 , transformations = list(log=c(1:12))
 , missing.format = "...."
)

#> Joining, by = "Parameter"

#>          Parameter Estimate CV.perc      SE estimated              CI95 transformed
#> 1    $\\theta_{1}$     5.40    ....    .... estimated     (4.96 - 5.88)         log
#> 2    $\\theta_{2}$     5.01    ....    .... estimated     (4.56 - 5.50)         log
#> 3    $\\theta_{3}$     2.27    ....    .... estimated     (1.98 - 2.60)         log
#> 4    $\\theta_{4}$     10.9    ....    .... estimated     (9.88 - 12.1)         log
#> 5  $\\omega_{1.1}$     1.18    ....    .... estimated     (1.12 - 1.24)         log
#> 6  $\\omega_{2.1}$    0.999    ....    .... estimated    (0.959 - 1.04)         log
#> 7  $\\omega_{2.2}$     1.14    ....    .... estimated     (1.07 - 1.21)         log
#> 8  $\\omega_{3.1}$     1.01    ....    .... estimated    (0.955 - 1.07)         log
#> 9  $\\omega_{3.2}$     1.02    ....    .... estimated    (0.947 - 1.09)         log
#> 10 $\\omega_{3.3}$     1.21    ....    .... estimated     (1.05 - 1.39)         log
#> 11 $\\omega_{4.1}$    0.987    ....    .... estimated    (0.943 - 1.03)         log
#> 12 $\\omega_{4.2}$     1.01    ....    .... estimated    (0.967 - 1.06)         log
#> 13 $\\omega_{4.3}$   0.0333     128  0.0426 estimated (-0.0502 - 0.117)          no
#> 14 $\\omega_{4.4}$    0.150    25.5  0.0382 estimated  (0.0750 - 0.225)          no
#> 15 $\\sigma_{1.1}$   0.0572    13.2 0.00757 estimated (0.0423 - 0.0720)          no

process.parTable(
   nm = nm.params.table(run = "example1", path = getOption("qpExampleDir"))
 , transformations = list(logit=c(7,9))
 , plain = TRUE
 , missing.format = "...."
)

#> Joining, by = "Parameter"

#>    Parameter  Estimate CV.perc      SE estimated               CI95 transformed
#> 1     THETA1      1.69     2.6  0.0432 estimated      (1.60 - 1.77)          no
#> 2     THETA2      1.61       3  0.0476 estimated      (1.52 - 1.70)          no
#> 3     THETA3     0.820     8.5  0.0700 estimated    (0.682 - 0.957)          no
#> 4     THETA4      2.39     2.2  0.0516 estimated      (2.29 - 2.49)          no
#> 5   OMEGA1.1     0.165    16.5  0.0272 estimated    (0.112 - 0.218)          no
#> 6   OMEGA2.1 -0.000742  2803.1  0.0208 estimated (-0.0415 - 0.0400)          no
#> 7   OMEGA2.2     0.533    ....    .... estimated    (0.518 - 0.548)       logit
#> 8   OMEGA3.1    0.0124   240.9  0.0299 estimated (-0.0462 - 0.0710)          no
#> 9   OMEGA3.2     0.504    ....    .... estimated    (0.486 - 0.521)       logit
#> 10  OMEGA3.3     0.188    38.3  0.0719 estimated   (0.0466 - 0.328)          no
#> 11  OMEGA4.1   -0.0127   185.1  0.0236 estimated (-0.0589 - 0.0335)          no
#> 12  OMEGA4.2    0.0139   173.1  0.0241 estimated (-0.0333 - 0.0611)          no
#> 13  OMEGA4.3    0.0333     128  0.0426 estimated  (-0.0502 - 0.117)          no
#> 14  OMEGA4.4     0.150    25.5  0.0382 estimated   (0.0750 - 0.225)          no
#> 15  SIGMA1.1    0.0572    13.2 0.00757 estimated  (0.0423 - 0.0720)          no

process.parTable(
   nm = nm.params.table(run = "example1", path = getOption("qpExampleDir"))
 , transformations = list(logit=c(7,9))
 , plain = TRUE
 , formatted = FALSE
 , missing.format = "...."
)

#> Joining, by = "Parameter"

#>    Parameter  Estimate CV.perc      SE estimated               CI95 transformed
#> 1     THETA1      1.69     2.6  0.0432 estimated       (1.6 - 1.77)          no
#> 2     THETA2      1.61       3  0.0476 estimated       (1.52 - 1.7)          no
#> 3     THETA3     0.820     8.5    0.07 estimated    (0.682 - 0.957)          no
#> 4     THETA4      2.39     2.2  0.0516 estimated      (2.29 - 2.49)          no
#> 5   OMEGA1.1     0.165    16.5  0.0272 estimated    (0.112 - 0.218)          no
#> 6   OMEGA2.1 -0.000742  2803.1  0.0208 estimated   (-0.0415 - 0.04)          no
#> 7   OMEGA2.2     0.533    ....    .... estimated    (0.518 - 0.548)       logit
#> 8   OMEGA3.1    0.0124   240.9  0.0299 estimated  (-0.0462 - 0.071)          no
#> 9   OMEGA3.2     0.504    ....    .... estimated    (0.486 - 0.521)       logit
#> 10  OMEGA3.3     0.188    38.3  0.0719 estimated   (0.0466 - 0.328)          no
#> 11  OMEGA4.1   -0.0127   185.1  0.0236 estimated (-0.0589 - 0.0335)          no
#> 12  OMEGA4.2    0.0139   173.1  0.0241 estimated (-0.0333 - 0.0611)          no
#> 13  OMEGA4.3    0.0333     128  0.0426 estimated  (-0.0502 - 0.117)          no
#> 14  OMEGA4.4     0.150    25.5  0.0382 estimated    (0.075 - 0.225)          no
#> 15  SIGMA1.1    0.0572    13.2 0.00757 estimated   (0.0423 - 0.072)          no