Format runrecord for reports

Process the run record into a data frame ready for integration into LaTeX report. Note that the function assumes runrecord has been run with PsN option -maxlvl = 0 and the function depends on input from read.runrec

process.runrec(
  runrec,
  improvement = list(value = -6.63, color = "blue"),
  carryAlong = NULL,
  plain = FALSE
)

Arguments

runrec
outut from read.runrec

improvement
a list with elements value (numeric) and color (character) to emphasize significant improvements. Value defaults to -6.63, for a p<0.01 improvement for a single parameter added.

carryAlong
any additional variable to be included (would need to be present in runrec)

plain
optional toggle to create a non-LaTeX formatted table

Value

Either a LaTeX preformatted runrecord table (default) or a fully processed runrecord table without LaTeX preformatting (for inclusion into Microsoft products).

See also

Examples

rr = read.runrec(
 filename = "AAruninfo.txt",
 path = system.file(package = 'qpToolkit','NONMEM')
)
process.runrec(rr)

#>   Run Ref      OFV                     $\\Delta$OFV CondNum Minimization                                  Description   action
#> 1   1   0  -1121.0                             <NA>      27   Successful                          2 compartment model        ~
#> 2   2   1 -10772.1 \\textbf{\\color{blue}{-9651.1}}     158   Successful Covariate Model CL and V with Gender and Age accepted

process.runrec(rr, plain = TRUE)

#>   Run Ref      OFV    dOFV CondNum Minimization                                  Description   action
#> 1   1   0  -1121.0    <NA>      27   Successful                          2 compartment model        ~
#> 2   2   1 -10772.1 -9651.1     158   Successful Covariate Model CL and V with Gender and Age accepted