Parse the information from the .ext file

Read and parse the information embedded in the .ext file from a completed NONMEM run. In case of multiple $EST calls, these will be parsed one by one and included with their respective $EST names into the resulting list

read.ext(run, path = getOption("nmDir"), file.ext = ".ext", quiet = TRUE)

Arguments

run
run rootname (e.g. run1)

path
directory where run resides

file.ext
file extension of the .ext file

quiet
if TRUE no message will be returned. #' @return A list with data.frames (as many as there where $EST calls in the NONMEM control stream) containing the .ext output

See also

Examples

read.ext(run = "example1", path = getOption("qpExampleDir"))

#> $`First Order Conditional Estimation with Interaction (No Prior)`
#>    ITERATION    THETA1    THETA2    THETA3    THETA4 SIGMA(1,1) OMEGA(1,1)   OMEGA(2,1) OMEGA(2,2) OMEGA(3,1)   OMEGA(3,2) OMEGA(3,3) OMEGA(4,1)
#> 1    0.0e+00  1.633960 1.5576000 0.7458190 2.3478500 0.05798520  0.1770000 -0.000428264  0.1561880  0.0126053 -0.000713190  0.1975260 -0.0179167
#> 2    5.0e+00  1.673980 1.6140900 0.8302580 2.3914000 0.05479820  0.1569080 -0.000403712  0.1359930  0.0122718 -0.000663722  0.1859870 -0.0133695
#> 3    1.0e+01  1.676700 1.6118700 0.8340870 2.3940800 0.05571030  0.1618530 -0.000432699  0.1295100  0.0159343 -0.000517306  0.1960030 -0.0087187
#> 4    1.5e+01  1.686310 1.6111900 0.8139870 2.3904700 0.05665560  0.1651220 -0.000561129  0.1333960  0.0126990  0.000326150  0.1910740 -0.0121500
#> 5    2.0e+01  1.686540 1.6108000 0.8183850 2.3906900 0.05726200  0.1646560 -0.001986370  0.1312260  0.0119768  0.015282300  0.1866400 -0.0131544
#> 6    2.5e+01  1.686750 1.6112100 0.8191190 2.3911300 0.05716500  0.1650260 -0.000778044  0.1314490  0.0123702  0.015890100  0.1873990 -0.0127808
#> 7    3.0e+01  1.686920 1.6112900 0.8196180 2.3916300 0.05716410  0.1650600 -0.000744653  0.1314240  0.0124075  0.015960200  0.1875490 -0.0127377
#> 8    3.3e+01  1.686930 1.6112900 0.8196050 2.3916100 0.05716320  0.1650610 -0.000741979  0.1314290  0.0124112  0.015956500  0.1875480 -0.0127355
#> 9   -1.0e+09  1.686930 1.6112900 0.8196050 2.3916100 0.05716320  0.1650610 -0.000741979  0.1314290  0.0124112  0.015956500  0.1875480 -0.0127355
#> 10  -1.0e+09  0.043180 0.0476109 0.0699832 0.0515872 0.00757063  0.0271869  0.020798200  0.0313100  0.0299001  0.035725600  0.0719070  0.0235760
#> 11  -1.0e+09  0.138801 0.3426480 0.3997250 0.4364790 0.52212300  0.6516330  0.784397000  0.7981560  0.8555510  0.879394000  1.0117600  1.2145000
#> 12  -1.0e+09 26.903000 0.1388010 3.7341600 0.0000000 0.00000000  0.0000000  0.000000000  0.0000000  0.0000000  0.000000000  0.0000000  0.0000000
#> 13  -1.0e+09  0.000000 0.0000000 0.0000000 0.0000000 0.23908800  0.4062770 -0.005037600  0.3625310  0.0705401  0.101633000  0.4330680 -0.0809622
#> 14  -1.0e+09  0.000000 0.0000000 0.0000000 0.0000000 0.01583230  0.0334585  0.141350000  0.0431825  0.1628810  0.226604000  0.0830204  0.1547840
#> 15  -1.0e+09  0.000000 0.0000000 0.0000000 0.0000000 0.00000000  0.0000000  0.000000000  0.0000000  0.0000000  0.000000000  0.0000000  0.0000000
#>    OMEGA(4,2) OMEGA(4,3) OMEGA(4,4)       OBJ
#> 1  0.01452820  0.0267534  0.1655500 -1115.895
#> 2  0.01238560  0.0415674  0.1501960 -1120.035
#> 3  0.00531919  0.0396320  0.1432600 -1120.268
#> 4  0.00921431  0.0358207  0.1510600 -1120.818
#> 5  0.01369540  0.0327462  0.1495190 -1121.025
#> 6  0.01386650  0.0332032  0.1498620 -1121.028
#> 7  0.01390470  0.0332660  0.1499010 -1121.028
#> 8  0.01390610  0.0332702  0.1499060 -1121.028
#> 9  0.01390610  0.0332702  0.1499060 -1121.028
#> 10 0.02407020  0.0425812  0.0382220     0.000
#> 11 1.43382000  1.7968600  3.7341600     0.000
#> 12 0.00000000  0.0000000  0.0000000     0.000
#> 13 0.09907190  0.1984220  0.3871770     0.000
#> 14 0.16592100  0.2081350  0.0493598     0.000
#> 15 0.00000000  0.0000000  0.0000000     0.000
#> 

## now parse an output where multiple $EST calls have been performed
read.ext(run = "example1", path = getOption("qpExampleDir"), file.ext = ".mext")

#> $`Iterative Two Stage (No Prior)`
#>    ITERATION     THETA1    THETA2    THETA3    THETA4 SIGMA(1,1) OMEGA(1,1) OMEGA(2,1) OMEGA(2,2)  OMEGA(3,1) OMEGA(3,2) OMEGA(3,3)  OMEGA(4,1)
#> 1   0.00e+00  2.0000000 2.0000000 2.0000000 2.0000000 0.60000000  0.1500000 0.01000000  0.1500000 0.010000000  0.0100000  0.1500000  0.01000000
#> 2   5.00e+00  1.7222300 1.6204400 0.9846850 2.4888600 0.06630990  0.1779420 0.01882720  0.1113590 0.008160170  0.0365088  0.1757050 -0.00580553
#> 3   1.00e+01  1.6867900 1.5898000 0.8452100 2.3904000 0.05794970  0.1662390 0.00710622  0.1191350 0.000922876  0.0366138  0.1886450 -0.01739400
#> 4   1.50e+01  1.6835200 1.5898000 0.8219940 2.3770800 0.05668360  0.1643060 0.00453556  0.1264720 0.001156440  0.0281036  0.1914710 -0.01844750
#> 5   2.00e+01  1.6830000 1.5894200 0.8150150 2.3740400 0.05593690  0.1641600 0.00397758  0.1300250 0.002389590  0.0225122  0.1964270 -0.01792900
#> 6   2.50e+01  1.6829800 1.5889000 0.8129080 2.3734700 0.05542440  0.1643560 0.00395403  0.1316580 0.003556400  0.0197104  0.2016880 -0.01723500
#> 7   3.00e+01  1.6830700 1.5885500 0.8124320 2.3735900 0.05506940  0.1646180 0.00410183  0.1325150 0.004496240  0.0183623  0.2059410 -0.01661550
#> 8   3.50e+01  1.6831800 1.5883500 0.8124240 2.3738200 0.05484080  0.1648400 0.00426502  0.1330060 0.005173160  0.0176964  0.2089230 -0.01614910
#> 9   4.00e+01  1.6832500 1.5882400 0.8125030 2.3740200 0.05469110  0.1650030 0.00439454  0.1333120 0.005627310  0.0173526  0.2108730 -0.01582890
#> 10  4.50e+01  1.6833000 1.5881900 0.8125790 2.3741500 0.05460050  0.1651090 0.00448224  0.1334950 0.005914760  0.0171647  0.2120820 -0.01562420
#> 11  5.00e+01  1.6833300 1.5881600 0.8126310 2.3742400 0.05454570  0.1651770 0.00454064  0.1336100 0.006093940  0.0170593  0.2128230 -0.01549540
#> 12  5.50e+01  1.6833500 1.5881400 0.8126640 2.3742900 0.05451680  0.1652170 0.00457388  0.1336730 0.006199410  0.0169990  0.2132570 -0.01541990
#> 13  6.00e+01  1.6833600 1.5881300 0.8126820 2.3743200 0.05449140  0.1652420 0.00459500  0.1337130 0.006262790  0.0169647  0.2135170 -0.01537440
#> 14  6.50e+01  1.6833700 1.5881200 0.8126950 2.3743400 0.05448070  0.1652550 0.00460598  0.1337360 0.006300060  0.0169435  0.2136680 -0.01534740
#> 15  7.00e+01  1.6833700 1.5881200 0.8127000 2.3743500 0.05447340  0.1652650 0.00461446  0.1337540 0.006323930  0.0169316  0.2137650 -0.01533050
#> 16  7.50e+01  1.6833700 1.5881200 0.8127040 2.3743500 0.05447040  0.1652710 0.00461983  0.1337650 0.006338620  0.0169245  0.2138260 -0.01531910
#> 17  8.00e+01  1.6833800 1.5881200 0.8127070 2.3743600 0.05446710  0.1652740 0.00462278  0.1337700 0.006347380  0.0169199  0.2138600 -0.01531320
#> 18  8.50e+01  1.6833800 1.5881200 0.8127080 2.3743600 0.05446650  0.1652750 0.00462368  0.1337710 0.006351210  0.0169174  0.2138750 -0.01531070
#> 19  9.00e+01  1.6833800 1.5881200 0.8127100 2.3743600 0.05446550  0.1652750 0.00462271  0.1337680 0.006350990  0.0169159  0.2138740 -0.01531110
#> 20  9.50e+01  1.6833800 1.5881200 0.8127100 2.3743600 0.05446730  0.1652750 0.00462271  0.1337680 0.006351070  0.0169157  0.2138740 -0.01531100
#> 21  1.00e+02  1.6833800 1.5881200 0.8127100 2.3743600 0.05446810  0.1652750 0.00462265  0.1337680 0.006351140  0.0169155  0.2138740 -0.01531090
#> 22  1.05e+02  1.6833800 1.5881200 0.8127110 2.3743600 0.05446470  0.1652740 0.00462212  0.1337660 0.006350150  0.0169159  0.2138710 -0.01531180
#> 23  1.05e+02  1.6833800 1.5881200 0.8127110 2.3743600 0.05446470  0.1652740 0.00462212  0.1337660 0.006350150  0.0169159  0.2138710 -0.01531180
#> 24 -1.00e+09  1.6833800 1.5881200 0.8127110 2.3743600 0.05446470  0.1652740 0.00462212  0.1337660 0.006350150  0.0169159  0.2138710 -0.01531180
#> 25 -1.00e+09  0.0461817 0.0490153 0.0652742 0.0539291 0.00761176  0.0289772 0.02313910  0.0349217 0.032498800  0.0382768  0.0648382  0.02761300
#> 26 -1.00e+09  0.1287040 0.2521200 0.3201370 0.3375210 0.44938400  0.5039670 0.61759500  0.7734270 0.950563000  1.0629000  1.3721500  1.44483000
#> 27 -1.00e+09 26.8244000 0.1287040 3.4524200 0.0000000 0.00000000  0.0000000 0.00000000  0.0000000 0.000000000  0.0000000  0.0000000  0.00000000
#> 28 -1.00e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.23337700  0.4065390 0.03108600  0.3657410 0.033775800  0.1000110  0.4624610 -0.09335600
#> 29 -1.00e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.01630790  0.0356388 0.15321000  0.0477411 0.170522000  0.2219530  0.0701012  0.17287300
#> 30 -1.00e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.00000000  0.0000000 0.00000000  0.0000000 0.000000000  0.0000000  0.0000000  0.00000000
#>    OMEGA(4,2) OMEGA(4,3) OMEGA(4,4)        OBJ
#> 1   0.0100000  0.0100000   0.150000  -234.3627
#> 2   0.0201868  0.0161868   0.134416 -1112.9740
#> 3   0.0196290  0.0327745   0.146616 -1119.9428
#> 4   0.0155689  0.0381413   0.151976 -1120.2901
#> 5   0.0135732  0.0422893   0.155309 -1120.3523
#> 6   0.0128063  0.0457926   0.157771 -1120.3406
#> 7   0.0125638  0.0484409   0.159556 -1120.3177
#> 8   0.0125130  0.0502571   0.160763 -1120.2987
#> 9   0.0125220  0.0514364   0.161552 -1120.2845
#> 10  0.0125408  0.0521652   0.162039 -1120.2758
#> 11  0.0125578  0.0526120   0.162339 -1120.2700
#> 12  0.0125677  0.0528749   0.162516 -1120.2668
#> 13  0.0125746  0.0530305   0.162621 -1120.2647
#> 14  0.0125789  0.0531227   0.162683 -1120.2635
#> 15  0.0125825  0.0531815   0.162723 -1120.2626
#> 16  0.0125848  0.0532185   0.162749 -1120.2624
#> 17  0.0125856  0.0532392   0.162763 -1120.2619
#> 18  0.0125858  0.0532486   0.162768 -1120.2619
#> 19  0.0125851  0.0532488   0.162768 -1120.2618
#> 20  0.0125850  0.0532490   0.162768 -1120.2618
#> 21  0.0125849  0.0532490   0.162768 -1120.2619
#> 22  0.0125848  0.0532469   0.162766 -1120.2618
#> 23  0.0125848  0.0532469   0.162766 -1120.2618
#> 24  0.0125848  0.0532469   0.162766 -1120.2618
#> 25  0.0273385  0.0450168   0.040638     0.0000
#> 26  1.5695600  1.7647200   3.452420     0.0000
#> 27  0.0000000  0.0000000   0.000000     0.0000
#> 28  0.0852889  0.2853890   0.403443     0.0000
#> 29  0.1794120  0.1857240   0.050364     0.0000
#> 30  0.0000000  0.0000000   0.000000     0.0000
#> 
#> $`Stochastic Approximation Expectation-Maximization (No Prior)`
#>    ITERATION     THETA1    THETA2    THETA3    THETA4 SIGMA(1,1) OMEGA(1,1)   OMEGA(2,1) OMEGA(2,2)  OMEGA(3,1)   OMEGA(3,2) OMEGA(3,3)  OMEGA(4,1)
#> 1   -3.0e+03  1.6833800 1.5881200 0.8127110 2.3743600 0.05446470  0.2479110  6.93318e-03   0.200649  0.00952522  0.025373900  0.3208060 -0.02296770
#> 2   -2.9e+03  1.6387100 1.5594500 0.7273210 2.3466100 0.05916100  0.1671730 -1.36751e-02   0.155428  0.02032360 -0.045769600  0.1625540 -0.00802073
#> 3   -2.8e+03  1.6236200 1.5591200 0.6730810 2.3293400 0.06180060  0.1641730 -1.92846e-03   0.141237  0.01136830 -0.033524300  0.1079540 -0.03172900
#> 4   -2.7e+03  1.6566400 1.6028100 0.7960210 2.3701600 0.06266950  0.1681500  1.12230e-03   0.136433  0.00110022  0.030926100  0.1217670 -0.01915940
#> 5   -2.6e+03  1.6540300 1.6199800 0.7351480 2.3575400 0.05939270  0.1924640  1.43644e-02   0.128209  0.03877490  0.030195300  0.1595340 -0.00520062
#> 6   -2.5e+03  1.6522700 1.5792600 0.7774180 2.3589600 0.06290830  0.1614820  2.73592e-03   0.117391 -0.00358101  0.025225200  0.1676190 -0.02270010
#> 7   -2.4e+03  1.6346500 1.5784400 0.7196130 2.3215500 0.06442210  0.1567660 -1.51455e-02   0.126365 -0.00592758 -0.000679692  0.1117680 -0.02344630
#> 8   -2.3e+03  1.6316000 1.5710800 0.6745370 2.3157300 0.05797910  0.1527460 -2.78948e-02   0.171447  0.01755620 -0.042057300  0.1414210 -0.02421130
#> 9   -2.2e+03  1.6367000 1.6305800 0.7052150 2.3317700 0.06687770  0.1730100 -1.86468e-04   0.113197 -0.00186648  0.006326660  0.0958769 -0.01908200
#> 10  -2.1e+03  1.6269200 1.5232100 0.6814660 2.3316700 0.05604720  0.1592740 -1.49024e-02   0.141980  0.00725141 -0.039506800  0.1369100 -0.01705770
#> 11  -2.0e+03  1.6277900 1.5500200 0.7087390 2.3472100 0.05291590  0.1549110 -5.46747e-03   0.160058  0.01292830 -0.010360800  0.1729740 -0.01198950
#> 12  -1.9e+03  1.6303100 1.5236600 0.7424690 2.3505300 0.05405190  0.1735590  1.35594e-02   0.151858  0.01507130 -0.038068100  0.2441320 -0.01098250
#> 13   0.0e+00  1.6211900 1.5317300 0.6996950 2.3331300 0.05405190  0.1707950  6.73434e-03   0.154136  0.02555920 -0.028328200  0.2245150 -0.01429500
#> 14   1.0e+02  1.6279800 1.5363900 0.7349190 2.3418600 0.05270070  0.1702030 -9.11465e-04   0.164074  0.02195790 -0.020277800  0.2114370 -0.00956001
#> 15   2.0e+02  1.6329700 1.5437900 0.7417480 2.3475600 0.05366130  0.1710560 -9.73948e-06   0.157078  0.02157150 -0.017722500  0.2113180 -0.00995849
#> 16   3.0e+02  1.6327400 1.5470200 0.7378460 2.3435200 0.05333160  0.1689530 -2.56594e-04   0.155001  0.01756410 -0.014272400  0.2104140 -0.01317660
#> 17   4.0e+02  1.6328200 1.5504200 0.7360380 2.3429900 0.05370980  0.1687880  6.16291e-04   0.155044  0.01664730 -0.012715300  0.2079370 -0.01479510
#> 18   5.0e+02  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -2.43010e-04   0.154364  0.01801720 -0.012497300  0.2116010 -0.01374220
#> 19  -1.0e+09  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -2.43010e-04   0.154364  0.01801720 -0.012497300  0.2116010 -0.01374220
#> 20  -1.0e+09  0.0473452 0.0549876 0.0678456 0.0533304 0.00709799  0.0283242  2.47905e-02   0.040713  0.03205130  0.046675000  0.0624189  0.02652110
#> 21  -1.0e+09  0.1708940 0.3052370 0.3743220 0.3848760 0.50568800  0.5466770  6.81905e-01   0.808982  0.92736300  1.165800000  1.2467400  1.33896000
#> 22  -1.0e+09 18.1641000 0.1708940 3.1041400 0.0000000 0.00000000  0.0000000  0.00000e+00   0.000000  0.00000000  0.000000000  0.0000000  0.00000000
#> 23  -1.0e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.23187000  0.4105920 -1.50640e-03   0.392891  0.09539330 -0.069148700  0.4600020 -0.08296530
#> 24  -1.0e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.01530600  0.0344918  1.53762e-01   0.051812  0.16353000  0.260037000  0.0678464  0.16336500
#> 25  -1.0e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.00000000  0.0000000  0.00000e+00   0.000000  0.00000000  0.000000000  0.0000000  0.00000000
#>      OMEGA(4,2)  OMEGA(4,3) OMEGA(4,4)   SAEMOBJ
#> 1   0.018877300  0.07987040  0.2441490 -2606.162
#> 2   0.008490050  0.01898890  0.1709730 -2430.559
#> 3   0.010777700 -0.02392630  0.1298120 -2449.484
#> 4   0.017809300 -0.00356824  0.1327270 -2432.628
#> 5   0.027842200  0.00657973  0.1406260 -2420.958
#> 6   0.033216400  0.03922240  0.1611090 -2427.960
#> 7  -0.004424840 -0.00544716  0.1445560 -2453.089
#> 8  -0.021294100  0.01019990  0.1393360 -2462.270
#> 9   0.027500400 -0.00391397  0.1437910 -2440.238
#> 10  0.025126600  0.00106890  0.1492130 -2463.403
#> 11  0.016914700  0.04166320  0.2027240 -2447.647
#> 12 -0.000613824  0.04053950  0.1557120 -2412.779
#> 13  0.003246850  0.03480000  0.1663990 -2447.128
#> 14  0.013157700  0.03819020  0.1601710 -2492.637
#> 15  0.012352400  0.03974030  0.1610990 -2491.971
#> 16  0.012753400  0.04010640  0.1619850 -2493.509
#> 17  0.012353600  0.03809500  0.1609590 -2492.725
#> 18  0.012028400  0.04053520  0.1627420 -2492.043
#> 19  0.012028400  0.04053520  0.1627420 -2492.043
#> 20  0.030859700  0.04238850  0.0408370     0.000
#> 21  1.653810000  1.78460000  3.1041400     0.000
#> 22  0.000000000  0.00000000  0.0000000     0.000
#> 23  0.075890500  0.21843600  0.4034120     0.000
#> 24  0.189427000  0.19138600  0.0506145     0.000
#> 25  0.000000000  0.00000000  0.0000000     0.000
#> 
#> $`Objective Function Evaluation by Importance Sampling (No Prior)`
#>    ITERATION     THETA1    THETA2    THETA3    THETA4 SIGMA(1,1) OMEGA(1,1)  OMEGA(2,1) OMEGA(2,2) OMEGA(3,1) OMEGA(3,2) OMEGA(3,3) OMEGA(4,1)
#> 1      0e+00  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 2      1e+00  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 3      2e+00  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 4      3e+00  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 5      4e+00  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 6      5e+00  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 7     -1e+09  1.6332900 1.5487000 0.7415410 2.3457000 0.05376390  0.1685860 -0.00024301  0.1543640  0.0180172 -0.0124973  0.2116010 -0.0137422
#> 8     -1e+09  0.0434084 0.0523151 0.0686892 0.0522198 0.00624247  0.0279757  0.02356600  0.0372362  0.0298380  0.0383233  0.0606987  0.0236205
#> 9     -1e+09  0.2040500 0.3598770 0.4671660 0.5071510 0.65072400  0.7188930  0.77450500  0.8623790  0.8953540  0.9877420  1.0998800  1.2612100
#> 10    -1e+09 13.7435000 0.2040500 2.8043600 0.0000000 0.00000000  0.0000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> 11    -1e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.23187000  0.4105920 -0.00150640  0.3928910  0.0953933 -0.0691487  0.4600020 -0.0829653
#> 12    -1e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.01346110  0.0340675  0.14611800  0.0473874  0.1513750  0.2104240  0.0659766  0.1460780
#> 13    -1e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.00000000  0.0000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#>    OMEGA(4,2) OMEGA(4,3) OMEGA(4,4)       OBJ
#> 1   0.0120284  0.0405352  0.1627420 -1145.140
#> 2   0.0120284  0.0405352  0.1627420 -1145.637
#> 3   0.0120284  0.0405352  0.1627420 -1145.012
#> 4   0.0120284  0.0405352  0.1627420 -1144.855
#> 5   0.0120284  0.0405352  0.1627420 -1144.071
#> 6   0.0120284  0.0405352  0.1627420 -1145.873
#> 7   0.0120284  0.0405352  0.1627420 -1145.873
#> 8   0.0264289  0.0364792  0.0378499     0.000
#> 9   1.4046000  2.0021000  2.8043600     0.000
#> 10  0.0000000  0.0000000  0.0000000     0.000
#> 11  0.0758905  0.2184360  0.4034120     0.000
#> 12  0.1640860  0.1636000  0.0469122     0.000
#> 13  0.0000000  0.0000000  0.0000000     0.000
#> 
#> $`First Order Conditional Estimation with Interaction (No Prior)`
#>    ITERATION     THETA1    THETA2    THETA3    THETA4 SIGMA(1,1) OMEGA(1,1)   OMEGA(2,1) OMEGA(2,2) OMEGA(3,1)   OMEGA(3,2) OMEGA(3,3) OMEGA(4,1)
#> 1    0.0e+00  1.6301100 1.5484000 0.7391950 2.3420900 0.05848970  0.1733470 -0.003258460  0.1559170 0.00585389 -0.003601550   0.201272 -0.0224928
#> 2    5.0e+00  1.6757200 1.6109000 0.8346650 2.3893300 0.05447550  0.1546320 -0.003053190  0.1352870 0.00601178 -0.003206850   0.185337 -0.0140070
#> 3    1.0e+01  1.6756900 1.6080400 0.8330760 2.3887100 0.05654850  0.1602260 -0.003214570  0.1264060 0.01109770 -0.000427698   0.174209 -0.0116656
#> 4    1.5e+01  1.6854300 1.6133200 0.8220930 2.3912400 0.05691200  0.1643700  0.000103787  0.1306050 0.01277070  0.018363300   0.187145 -0.0119951
#> 5    2.0e+01  1.6867500 1.6112000 0.8191030 2.3911200 0.05716510  0.1650240 -0.000783667  0.1314490 0.01236840  0.015886000   0.187397 -0.0127829
#> 6    2.3e+01  1.6869300 1.6112900 0.8196040 2.3916100 0.05716290  0.1650620 -0.000741584  0.1314290 0.01241200  0.015956400   0.187549 -0.0127355
#> 7   -1.0e+09  1.6869300 1.6112900 0.8196040 2.3916100 0.05716290  0.1650620 -0.000741584  0.1314290 0.01241200  0.015956400   0.187549 -0.0127355
#> 8   -1.0e+09  0.0432046 0.0478577 0.0704887 0.0518067 0.00935508  0.0273185  0.022288500  0.0320360 0.02980410  0.035340300   0.088357  0.0239593
#> 9   -1.0e+09  0.0994675 0.2875420 0.3602960 0.4093320 0.43926200  0.5374090  0.729687000  0.8091900 0.86088700  0.884479000   1.002930  1.1245900
#> 10  -1.0e+09 42.5318000 0.0994675 4.2305300 0.0000000 0.00000000  0.0000000  0.000000000  0.0000000 0.00000000  0.000000000   0.000000  0.0000000
#> 11  -1.0e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.23908800  0.4062780 -0.005034900  0.3625320 0.07054430  0.101633000   0.433069 -0.0809617
#> 12  -1.0e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.01956410  0.0336205  0.151557000  0.0441837 0.16130500  0.226252000   0.102013  0.1580650
#> 13  -1.0e+09  0.0000000 0.0000000 0.0000000 0.0000000 0.00000000  0.0000000  0.000000000  0.0000000 0.00000000  0.000000000   0.000000  0.0000000
#>    OMEGA(4,2) OMEGA(4,3) OMEGA(4,4)       OBJ
#> 1  0.01652150  0.0284230  0.1658080 -1115.016
#> 2  0.01204680  0.0429975  0.1498880 -1119.691
#> 3  0.00368172  0.0317690  0.1394510 -1120.068
#> 4  0.01349970  0.0338444  0.1480420 -1121.001
#> 5  0.01386540  0.0332020  0.1498600 -1121.028
#> 6  0.01390540  0.0332715  0.1499080 -1121.028
#> 7  0.01390540  0.0332715  0.1499080 -1121.028
#> 8  0.02376170  0.0499630  0.0410975     0.000
#> 9  1.40287000  1.8215200  4.2305300     0.000
#> 10 0.00000000  0.0000000  0.0000000     0.000
#> 11 0.09906640  0.1984280  0.3871790     0.000
#> 12 0.16371400  0.2394620  0.0530731     0.000
#> 13 0.00000000  0.0000000  0.0000000     0.000
#> 

## a large one
read.ext(run = "example1", path = getOption("qpExampleDir"))

#> $`First Order Conditional Estimation with Interaction (No Prior)`
#>    ITERATION    THETA1    THETA2    THETA3    THETA4 SIGMA(1,1) OMEGA(1,1)   OMEGA(2,1) OMEGA(2,2) OMEGA(3,1)   OMEGA(3,2) OMEGA(3,3) OMEGA(4,1)
#> 1    0.0e+00  1.633960 1.5576000 0.7458190 2.3478500 0.05798520  0.1770000 -0.000428264  0.1561880  0.0126053 -0.000713190  0.1975260 -0.0179167
#> 2    5.0e+00  1.673980 1.6140900 0.8302580 2.3914000 0.05479820  0.1569080 -0.000403712  0.1359930  0.0122718 -0.000663722  0.1859870 -0.0133695
#> 3    1.0e+01  1.676700 1.6118700 0.8340870 2.3940800 0.05571030  0.1618530 -0.000432699  0.1295100  0.0159343 -0.000517306  0.1960030 -0.0087187
#> 4    1.5e+01  1.686310 1.6111900 0.8139870 2.3904700 0.05665560  0.1651220 -0.000561129  0.1333960  0.0126990  0.000326150  0.1910740 -0.0121500
#> 5    2.0e+01  1.686540 1.6108000 0.8183850 2.3906900 0.05726200  0.1646560 -0.001986370  0.1312260  0.0119768  0.015282300  0.1866400 -0.0131544
#> 6    2.5e+01  1.686750 1.6112100 0.8191190 2.3911300 0.05716500  0.1650260 -0.000778044  0.1314490  0.0123702  0.015890100  0.1873990 -0.0127808
#> 7    3.0e+01  1.686920 1.6112900 0.8196180 2.3916300 0.05716410  0.1650600 -0.000744653  0.1314240  0.0124075  0.015960200  0.1875490 -0.0127377
#> 8    3.3e+01  1.686930 1.6112900 0.8196050 2.3916100 0.05716320  0.1650610 -0.000741979  0.1314290  0.0124112  0.015956500  0.1875480 -0.0127355
#> 9   -1.0e+09  1.686930 1.6112900 0.8196050 2.3916100 0.05716320  0.1650610 -0.000741979  0.1314290  0.0124112  0.015956500  0.1875480 -0.0127355
#> 10  -1.0e+09  0.043180 0.0476109 0.0699832 0.0515872 0.00757063  0.0271869  0.020798200  0.0313100  0.0299001  0.035725600  0.0719070  0.0235760
#> 11  -1.0e+09  0.138801 0.3426480 0.3997250 0.4364790 0.52212300  0.6516330  0.784397000  0.7981560  0.8555510  0.879394000  1.0117600  1.2145000
#> 12  -1.0e+09 26.903000 0.1388010 3.7341600 0.0000000 0.00000000  0.0000000  0.000000000  0.0000000  0.0000000  0.000000000  0.0000000  0.0000000
#> 13  -1.0e+09  0.000000 0.0000000 0.0000000 0.0000000 0.23908800  0.4062770 -0.005037600  0.3625310  0.0705401  0.101633000  0.4330680 -0.0809622
#> 14  -1.0e+09  0.000000 0.0000000 0.0000000 0.0000000 0.01583230  0.0334585  0.141350000  0.0431825  0.1628810  0.226604000  0.0830204  0.1547840
#> 15  -1.0e+09  0.000000 0.0000000 0.0000000 0.0000000 0.00000000  0.0000000  0.000000000  0.0000000  0.0000000  0.000000000  0.0000000  0.0000000
#>    OMEGA(4,2) OMEGA(4,3) OMEGA(4,4)       OBJ
#> 1  0.01452820  0.0267534  0.1655500 -1115.895
#> 2  0.01238560  0.0415674  0.1501960 -1120.035
#> 3  0.00531919  0.0396320  0.1432600 -1120.268
#> 4  0.00921431  0.0358207  0.1510600 -1120.818
#> 5  0.01369540  0.0327462  0.1495190 -1121.025
#> 6  0.01386650  0.0332032  0.1498620 -1121.028
#> 7  0.01390470  0.0332660  0.1499010 -1121.028
#> 8  0.01390610  0.0332702  0.1499060 -1121.028
#> 9  0.01390610  0.0332702  0.1499060 -1121.028
#> 10 0.02407020  0.0425812  0.0382220     0.000
#> 11 1.43382000  1.7968600  3.7341600     0.000
#> 12 0.00000000  0.0000000  0.0000000     0.000
#> 13 0.09907190  0.1984220  0.3871770     0.000
#> 14 0.16592100  0.2081350  0.0493598     0.000
#> 15 0.00000000  0.0000000  0.0000000     0.000
#>