Reorder names in a data.frame

It can be tedious to reorder the columns in a data.frame. This function lets you move specific columns relative to some other named column.

# S3 method for data.frame
reorder(x, ..., who, after = NA)

Arguments

x
data.frame

...
ignored

who
a character vector of column names to move, or a logical vector of length names(x), or a vector of indices

after
the column after which to put who: may be character, integer, NA, or NULL

Value

data.frame: a variant of x

Examples

testData = expand.grid(start = 1,middle = LETTERS[1:3], end = c("K","L"))
reorder(testData, who = "start", after = "end")

#>   middle end start
#> 1      A   K     1
#> 2      B   K     1
#> 3      C   K     1
#> 4      A   L     1
#> 5      B   L     1
#> 6      C   L     1