Tabulate a summary

Tabulate a summary

tabSummarize(
  formula,
  data,
  digits = 3,
  extra.blank.line = TRUE,
  digits.categorical = 1
)

Arguments

formula
formula specifying what to summarize by what

data
data to summarize

digits
number of significant digits in output

extra.blank.line
a logical indicating if an empty line needs to be inserted between two summaries to improve layout and legibility

digits.categorical
passed to tabStats

Value

Table with summarized data, of class demoTable

Note

This function is primarily used for demographics tables

See also

Examples

options(width = 150)
pkpdData = example.pkpdData()
ok = duplicated(pkpdData$id) == FALSE
tabSummarize(formula = dose ~ race + wt + bmi + sex, data = pkpdData[ok, ], digits = 3)

#>         parameter                  0                100                250                500                All
#> 1                           (N = 32)           (N = 32)           (N = 32)           (N = 32)          (N = 128)
#> 2                                                                                                               
#> 3            Race                                                                                               
#> 4           Asian          6 (18.8%)          6 (18.8%)           3 (9.4%)           3 (9.4%)         18 (14.1%)
#> 5           Black          5 (15.6%)          6 (18.8%)          5 (15.6%)           2 (6.2%)         18 (14.1%)
#> 6       Caucasian         18 (56.2%)         18 (56.2%)         24 (75.0%)         23 (71.9%)         83 (64.8%)
#> 7           Other           3 (9.4%)           2 (6.2%)           0 (0.0%)          4 (12.5%)           9 (7.0%)
#> 8                                                                                                               
#> 9     Weight (kg)                                                                                               
#> 10      Mean (SD)         75.3 (9.0)          76.2 (13)          78.0 (10)          75.7 (10)          76.3 (10)
#> 11 Median (range) 76.0 (56.0 - 92.0)  74.5 (46.0 - 105) 79.5 (54.0 - 98.0)  78.0 (43.0 - 104)  77.0 (43.0 - 105)
#> 12                                                                                                              
#> 13    BMI (kg/m2)                                                                                               
#> 14      Mean (SD)         22.9 (4.1)         23.6 (5.8)         24.6 (4.8)         24.2 (4.3)         23.8 (4.8)
#> 15 Median (range) 23.0 (16.7 - 35.5) 23.0 (12.0 - 36.4) 24.6 (16.0 - 38.8) 24.2 (12.0 - 32.8) 23.6 (12.0 - 38.8)
#> 16                                                                                                              
#> 17         Gender                                                                                               
#> 18              F         15 (46.9%)         11 (34.4%)         12 (37.5%)         15 (46.9%)         53 (41.4%)
#> 19              M         17 (53.1%)         21 (65.6%)         20 (62.5%)         17 (53.1%)         75 (58.6%)

pkpdData$sex <- factor(pkpdData$sex, levels = c('M','F'))
tabSummarize(formula = dose ~ race + wt + bmi + sex, data = pkpdData[ok, ], digits = 3)

#>         parameter                  0                100                250                500                All
#> 1                           (N = 32)           (N = 32)           (N = 32)           (N = 32)          (N = 128)
#> 2                                                                                                               
#> 3            Race                                                                                               
#> 4           Asian          6 (18.8%)          6 (18.8%)           3 (9.4%)           3 (9.4%)         18 (14.1%)
#> 5           Black          5 (15.6%)          6 (18.8%)          5 (15.6%)           2 (6.2%)         18 (14.1%)
#> 6       Caucasian         18 (56.2%)         18 (56.2%)         24 (75.0%)         23 (71.9%)         83 (64.8%)
#> 7           Other           3 (9.4%)           2 (6.2%)           0 (0.0%)          4 (12.5%)           9 (7.0%)
#> 8                                                                                                               
#> 9     Weight (kg)                                                                                               
#> 10      Mean (SD)         75.3 (9.0)          76.2 (13)          78.0 (10)          75.7 (10)          76.3 (10)
#> 11 Median (range) 76.0 (56.0 - 92.0)  74.5 (46.0 - 105) 79.5 (54.0 - 98.0)  78.0 (43.0 - 104)  77.0 (43.0 - 105)
#> 12                                                                                                              
#> 13    BMI (kg/m2)                                                                                               
#> 14      Mean (SD)         22.9 (4.1)         23.6 (5.8)         24.6 (4.8)         24.2 (4.3)         23.8 (4.8)
#> 15 Median (range) 23.0 (16.7 - 35.5) 23.0 (12.0 - 36.4) 24.6 (16.0 - 38.8) 24.2 (12.0 - 32.8) 23.6 (12.0 - 38.8)
#> 16                                                                                                              
#> 17         Gender                                                                                               
#> 18              M         17 (53.1%)         21 (65.6%)         20 (62.5%)         17 (53.1%)         75 (58.6%)
#> 19              F         15 (46.9%)         11 (34.4%)         12 (37.5%)         15 (46.9%)         53 (41.4%)

# check the categorical summary for race
round(table(pkpdData$race[ok], pkpdData$dose[ok]) / apply(table(pkpdData$race[ok],
pkpdData$dose[ok]), 2, sum) * 100)

#>            
#>              0 100 250 500
#>   Asian     19  19   9   9
#>   Black     16  19  16   6
#>   Caucasian 56  56  75  72
#>   Other      9   6   0  12

# OK
myFormula =  dose ~ race + wt + bmi + sex
format(
 tabSummarize(formula = myFormula, data = pkpdData[ok, ], digits = 3)
  , formula = myFormula
)

#>       ~                        0                    100                  250                  500                  All                 
#>  [1,] "\\hfill "               "(N = 32)"           "(N = 32)"           "(N = 32)"           "(N = 32)"           "(N = 128)"         
#>  [2,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#>  [3,] "\\textbf{Race}"         ""                   ""                   ""                   ""                   ""                  
#>  [4,] "\\hfill Asian"          "6 (18.8\\%)"        "6 (18.8\\%)"        "3 (9.4\\%)"         "3 (9.4\\%)"         "18 (14.1\\%)"      
#>  [5,] "\\hfill Black"          "5 (15.6\\%)"        "6 (18.8\\%)"        "5 (15.6\\%)"        "2 (6.2\\%)"         "18 (14.1\\%)"      
#>  [6,] "\\hfill Caucasian"      "18 (56.2\\%)"       "18 (56.2\\%)"       "24 (75.0\\%)"       "23 (71.9\\%)"       "83 (64.8\\%)"      
#>  [7,] "\\hfill Other"          "3 (9.4\\%)"         "2 (6.2\\%)"         "0 (0.0\\%)"         "4 (12.5\\%)"        "9 (7.0\\%)"        
#>  [8,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#>  [9,] "\\textbf{Weight (kg)}"  ""                   ""                   ""                   ""                   ""                  
#> [10,] "\\hfill Mean (SD)"      "75.3 (9.0)"         "76.2 (13)"          "78.0 (10)"          "75.7 (10)"          "76.3 (10)"         
#> [11,] "\\hfill Median (range)" "76.0 (56.0 - 92.0)" "74.5 (46.0 - 105)"  "79.5 (54.0 - 98.0)" "78.0 (43.0 - 104)"  "77.0 (43.0 - 105)" 
#> [12,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#> [13,] "\\textbf{BMI (kg/m2)}"  ""                   ""                   ""                   ""                   ""                  
#> [14,] "\\hfill Mean (SD)"      "22.9 (4.1)"         "23.6 (5.8)"         "24.6 (4.8)"         "24.2 (4.3)"         "23.8 (4.8)"        
#> [15,] "\\hfill Median (range)" "23.0 (16.7 - 35.5)" "23.0 (12.0 - 36.4)" "24.6 (16.0 - 38.8)" "24.2 (12.0 - 32.8)" "23.6 (12.0 - 38.8)"
#> [16,] "\\hfill "               ""                   ""                   ""                   ""                   ""                  
#> [17,] "\\textbf{Gender}"       ""                   ""                   ""                   ""                   ""                  
#> [18,] "\\hfill M"              "17 (53.1\\%)"       "21 (65.6\\%)"       "20 (62.5\\%)"       "17 (53.1\\%)"       "75 (58.6\\%)"      
#> [19,] "\\hfill F"              "15 (46.9\\%)"       "11 (34.4\\%)"       "12 (37.5\\%)"       "15 (46.9\\%)"       "53 (41.4\\%)"