NONMEM output extraction from XML output

Extracts all output embedded in the XML file of a NONMEM run

nm.extract.xml(run = "run1", path = getOption("nmDir"),
  filename = paste(path, run, paste(run, xml.extension, sep = ""), sep =
  "/"), signif = 4, signif.se = signif, signif.cv = signif,
  signif.est = signif, xml.extension = ".xml", zip.extension = ".7z",
  zip.program = "c:/progra~1/7-zip/7z", remove.obsolete = FALSE,
  na.value = "n.d.", control_stream = FALSE, get.xpose.tables = TRUE,
  quiet = TRUE)

Arguments

run
run rootname (e.g. run1)

path
directory where run resides

filename
with full path of the XML file to be parsed. This is automatically created using qP workflow standards but can be specified as well

signif
number of significant digits

signif.se
number of significant digits in SE (when point estimates = T)

signif.cv
number of significant digits in CV% (when point.estimates = T)

signif.est
number of significant digits in parameter estimates (when point.estimates = T)

xml.extension
extension of the xml file. Defaults to ".xml"

zip.extension
extension of the zip program. Defaults to ".7z"

zip.program
full path to the executable of the zip program (7zip or winzip). Defauls to "c:/progra~1/7-zip/7z"

remove.obsolete
logical (F) defining if non-estimated parameters should be dropped for the parameter table.

na.value
what to do with an NA value in the parameter table. Defaults to "n.d."

control_stream
logical (F) defining if the control stream should be included as well.

get.xpose.tables
logical (T) defining if the xpose tables should be integrated in the result as a single data frame using the function get.xpose.tables

quiet
if TRUE (default) silences any message returned by the function and subfunctions.

Value

An extensive list with all elements extracted from the XML output

Note

Works for NM7.2 (up to 9 OMEGAs) and NM7.3 (bug free)

Examples

tmp = nm.extract.xml("example2", path = getOption("qpExampleDir"))

#> Error: XML content does not seem to be XML: '/home/rstudio/rstudio/NONMEM/example2/example2.xml'

names(tmp)

#> Error in eval(expr, envir, enclos): object 'tmp' not found

tmp$final_objective_function

#> Error in eval(expr, envir, enclos): object 'tmp' not found

tmp$table

#> Error in eval(expr, envir, enclos): object 'tmp' not found

tmp$monitor

#> Error in eval(expr, envir, enclos): object 'tmp' not found