Table Statistics

Determines if variable is continuous or categorical and returns the appropriate summary statistics.

tabStats(
  x,
  BY,
  digits = 3,
  conFunc1 = conDataFun1,
  conFunc2 = conDataFun2,
  catFunc = catDataFun,
  digits.categorical = 1,
  parName
)

Arguments

x
numeric vector to summarize

BY
stratification argument

digits
number of significant digits in continuous data summaries

conFunc1
defaults to conDataFun1

conFunc2
defaults to conDataFun2

catFunc
defaults to catDataFun

digits.categorical
passed to catFunc

parName
is automatically populated but the parameter name can be specified here by the user.

Value

Appropriate statistic for x

Note

This function is primarily used for demographics tables

See also

Examples

pkpdData = example.pkpdData()
ok = duplicated(pkpdData$id) == FALSE
tabStats(x = pkpdData$race[ok], BY = list(dose = pkpdData$dose[ok]))

#>           parameter          0        100        250        500        All
#> 1 pkpdData$race[ok]                                                       
#> 2             Asian  6 (18.8%)  6 (18.8%)   3 (9.4%)   3 (9.4%) 18 (14.1%)
#> 3             Black  5 (15.6%)  6 (18.8%)  5 (15.6%)   2 (6.2%) 18 (14.1%)
#> 4         Caucasian 18 (56.2%) 18 (56.2%) 24 (75.0%) 23 (71.9%) 83 (64.8%)
#> 5             Other   3 (9.4%)   2 (6.2%)   0 (0.0%)  4 (12.5%)   9 (7.0%)

tabStats(x = pkpdData$race[ok], BY = list(dose = pkpdData$dose[ok]),
parName = "Race", digits.categorical = 0)

#>   parameter        0      100      250      500      All
#> 1      Race                                             
#> 2     Asian  6 (19%)  6 (19%)   3 (9%)   3 (9%) 18 (14%)
#> 3     Black  5 (16%)  6 (19%)  5 (16%)   2 (6%) 18 (14%)
#> 4 Caucasian 18 (56%) 18 (56%) 24 (75%) 23 (72%) 83 (65%)
#> 5     Other   3 (9%)   2 (6%)   0 (0%)  4 (12%)   9 (7%)

tabStats(pkpdData$sex[ok], list(dose = pkpdData$dose[ok]))

#>          parameter          0        100        250        500        All
#> 1 pkpdData$sex[ok]                                                       
#> 2                F 15 (46.9%) 11 (34.4%) 12 (37.5%) 15 (46.9%) 53 (41.4%)
#> 3                M 17 (53.1%) 21 (65.6%) 20 (62.5%) 17 (53.1%) 75 (58.6%)

tabStats(x = pkpdData$wt, BY = list(dose = pkpdData$dose))

#>        parameter                  0               100                250               500               All
#> 1    pkpdData$wt                                                                                            
#> 2      Mean (SD)         75.3 (8.9)         76.2 (12)         78.0 (9.8)        75.7 (9.9)         76.3 (10)
#> 3 Median (range) 76.0 (56.0 - 92.0) 74.5 (46.0 - 105) 79.5 (54.0 - 98.0) 78.0 (43.0 - 104) 77.0 (43.0 - 105)

tabStats(x = pkpdData$bmi, BY = list(dose = pkpdData$dose))

#>        parameter                  0                100                250                500                All
#> 1   pkpdData$bmi                                                                                               
#> 2      Mean (SD)         22.9 (4.0)         23.6 (5.7)         24.6 (4.7)         24.2 (4.3)         23.8 (4.8)
#> 3 Median (range) 23.0 (16.7 - 35.5) 23.0 (12.0 - 36.4) 24.6 (16.0 - 38.8) 24.2 (12.0 - 32.8) 23.6 (12.0 - 38.8)